% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{add_roxygen_input}
\alias{add_roxygen_input}
\title{Input: message content template for add_roxygen()}
\format{
## `add_roxygen`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
add_roxygen_input
}
\description{
This message content serves as user input for the add_roxygen function.
}
\keyword{datasets}
