% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{add_upload_part}
\alias{add_upload_part}
\title{Add upload part (convenience function)}
\usage{
add_upload_part(upload_id, data)
}
\arguments{
\item{upload_id}{Upload ID}

\item{data}{Raw data}
}
\value{
Upload part
}
\description{
Add upload part (convenience function)
}
