% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{AssistantsClient}
\alias{AssistantsClient}
\title{Assistants Client (Beta)}
\description{
Client for OpenAI Assistants API v2.
Build AI assistants with tools and persistent state.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AssistantsClient-new}{\code{AssistantsClient$new()}}
\item \href{#method-AssistantsClient-create}{\code{AssistantsClient$create()}}
\item \href{#method-AssistantsClient-list}{\code{AssistantsClient$list()}}
\item \href{#method-AssistantsClient-retrieve}{\code{AssistantsClient$retrieve()}}
\item \href{#method-AssistantsClient-update}{\code{AssistantsClient$update()}}
\item \href{#method-AssistantsClient-delete}{\code{AssistantsClient$delete()}}
\item \href{#method-AssistantsClient-clone}{\code{AssistantsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$create(
  model,
  name = NULL,
  description = NULL,
  instructions = NULL,
  tools = NULL,
  tool_resources = NULL,
  metadata = NULL,
  temperature = NULL,
  top_p = NULL,
  response_format = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$list(limit = NULL, order = NULL, after = NULL, before = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$retrieve(assistant_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-update"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$update(assistant_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$delete(assistant_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssistantsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AssistantsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AssistantsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
