% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{BatchClient}
\alias{BatchClient}
\title{Batch Client}
\description{
Client for OpenAI Batch API.
Process large numbers of requests asynchronously.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BatchClient-new}{\code{BatchClient$new()}}
\item \href{#method-BatchClient-create}{\code{BatchClient$create()}}
\item \href{#method-BatchClient-list}{\code{BatchClient$list()}}
\item \href{#method-BatchClient-retrieve}{\code{BatchClient$retrieve()}}
\item \href{#method-BatchClient-cancel}{\code{BatchClient$cancel()}}
\item \href{#method-BatchClient-clone}{\code{BatchClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$create(
  input_file_id,
  endpoint,
  completion_window = "24h",
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$list(after = NULL, limit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$retrieve(batch_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-cancel}{}}}
\subsection{Method \code{cancel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$cancel(batch_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BatchClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BatchClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BatchClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
