% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakOpts.R
\name{breakOpts}
\alias{breakOpts}
\title{Define \code{breaks} options for \code{openair} plots}
\usage{
breakOpts(
  breaks = NULL,
  labels = NULL,
  method = c("number", "interval", "width", "pretty", "wd"),
  max.bins = NULL,
  include.lowest = NULL,
  right = NULL,
  dig.lab = NULL
)
}
\arguments{
\item{breaks, labels}{If a categorical colour scale is required, \code{breaks}
should be specified. This can be either of:
\itemize{
\item A single value, which will use the strategy specified by \code{method} to
bin the scale. By default, this uses the same logic as \code{\link[=cutData]{cutData()}}, which
splits the scale into quantiles.
\item A numeric vector, which will define the specific breakpoints. For
example, \code{c(0, 50, 100)} will bin the data into \verb{0 to 50}, \verb{50 to 100}, and
so on. If \code{breaks} does not cover the full range of the data, the outer
limits will be extended so that the full colour scale is covered while
retaining the desired number of breaks.
}

By default, \code{breaks} will generate nicely formatted labels for each
category. The \code{labels} argument overrides this - for example, a user could
define \verb{breaks = 3, labels = c("low", "medium", "high")}. Care should be
taken to provide the appropriate number of \code{labels} - it should always be
equal to the number of bins, or one less than the number of breakpoints.}

\item{method}{When \code{breaks} is a single number (e.g., \code{breaks = 5}), \code{method}
controls how this value is used to create breaks. Can be one of:
\itemize{
\item \code{"number"} - splits the range into \emph{n} quantiles (i.e., bins with roughly
the same number of measurements in them). This is equivalent to
\code{\link[=cutData]{cutData()}}.
\item \code{"interval"} - splits the range into \emph{n} bins of equal width.
\item \code{"width"} - splits the range into some number of bins with width \emph{n}.
\item \code{"pretty"} - splits the range into \emph{approximately} \emph{n} bins with
aesthetically pleasing breakpoints using \code{\link[=pretty]{pretty()}}.
\item \code{"wd"} - splits the range into wind direction bins by passing it
to \code{\link[=cutData]{cutData()}}. Note that this option requires \code{breaks} to be one of \code{4},
\code{8}, \code{16} or \code{32} and ignores all other arguments.
}}

\item{max.bins}{If \code{max.bins} is set, only that number of bins will ever be
created. For example, if \code{breaks = c(0, 10, 20, 30, 40, 50)} and \code{max.bins = 3}, \code{breaks} will be set to \code{c(0, 10, 20, 50)} to ensure only three bins
are made.}

\item{include.lowest}{logical, indicating if an \code{x[i]} equal to the lowest
(or highest, for \code{right = FALSE}) 'breaks' value should be included.}

\item{right}{logical, indicating if the intervals should be closed on the
right (and open on the left) or vice versa.}

\item{dig.lab}{Integer which is used when labels are not given. It determines
the number of digits used in formatting the break numbers. If \code{NULL}, the
labels will use the minimum number of decimal places possible to ensure the
breakpoints are unique, dropping all trailing zeroes.}
}
\description{
This function provides a convenient way to set default options for \code{breaks}
arguments in \code{openair} plots, which cut continuous colour scales into
discrete bins.
}
