% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{cancel_batch}
\alias{cancel_batch}
\title{Cancel a batch (convenience function)}
\usage{
cancel_batch(batch_id)
}
\arguments{
\item{batch_id}{Batch ID}
}
\value{
Cancelled batch
}
\description{
Cancel a batch (convenience function)
}
