% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{cancel_response}
\alias{cancel_response}
\title{Cancel a response (convenience function)}
\usage{
cancel_response(response_id)
}
\arguments{
\item{response_id}{Response ID}
}
\value{
Cancelled response
}
\description{
Cancel a response (convenience function)
}
