% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{cancel_run}
\alias{cancel_run}
\title{Cancel a run (convenience function)}
\usage{
cancel_run(thread_id, run_id)
}
\arguments{
\item{thread_id}{Thread ID}

\item{run_id}{Run ID}
}
\value{
Cancelled run
}
\description{
Cancel a run (convenience function)
}
