% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{cancel_upload}
\alias{cancel_upload}
\title{Cancel an upload (convenience function)}
\usage{
cancel_upload(upload_id)
}
\arguments{
\item{upload_id}{Upload ID}
}
\value{
Cancelled upload
}
\description{
Cancel an upload (convenience function)
}
