% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{ChatCompletionsClient}
\alias{ChatCompletionsClient}
\title{Chat Completions Interface}
\description{
Provides methods to create, manage and retrieve chat completions.
Access via \code{client$chat$completions}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChatCompletionsClient-new}{\code{ChatCompletionsClient$new()}}
\item \href{#method-ChatCompletionsClient-create}{\code{ChatCompletionsClient$create()}}
\item \href{#method-ChatCompletionsClient-retrieve}{\code{ChatCompletionsClient$retrieve()}}
\item \href{#method-ChatCompletionsClient-update}{\code{ChatCompletionsClient$update()}}
\item \href{#method-ChatCompletionsClient-list}{\code{ChatCompletionsClient$list()}}
\item \href{#method-ChatCompletionsClient-delete}{\code{ChatCompletionsClient$delete()}}
\item \href{#method-ChatCompletionsClient-clone}{\code{ChatCompletionsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$create(
  messages,
  model = "gpt-3.5-turbo",
  frequency_penalty = NULL,
  logit_bias = NULL,
  logprobs = NULL,
  top_logprobs = NULL,
  max_tokens = NULL,
  max_completion_tokens = NULL,
  n = NULL,
  presence_penalty = NULL,
  response_format = NULL,
  seed = NULL,
  stop = NULL,
  stream = NULL,
  stream_options = NULL,
  temperature = NULL,
  top_p = NULL,
  tools = NULL,
  tool_choice = NULL,
  parallel_tool_calls = NULL,
  user = NULL,
  store = NULL,
  metadata = NULL,
  callback = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$retrieve(completion_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-update"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$update(completion_id, metadata = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$list(
  model = NULL,
  after = NULL,
  limit = NULL,
  order = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$delete(completion_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
