% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{ChatCompletionsMessagesClient}
\alias{ChatCompletionsMessagesClient}
\title{Chat Completions Messages Client}
\description{
Lists messages from stored chat completions.
Access via \code{client$chat$completions$messages}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ChatCompletionsMessagesClient-new}{\code{ChatCompletionsMessagesClient$new()}}
\item \href{#method-ChatCompletionsMessagesClient-list}{\code{ChatCompletionsMessagesClient$list()}}
\item \href{#method-ChatCompletionsMessagesClient-clone}{\code{ChatCompletionsMessagesClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsMessagesClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsMessagesClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsMessagesClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsMessagesClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsMessagesClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsMessagesClient$list(
  completion_id,
  after = NULL,
  limit = NULL,
  order = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChatCompletionsMessagesClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChatCompletionsMessagesClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChatCompletionsMessagesClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
