% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{chatlog-class}
\alias{chatlog-class}
\title{chatlog class}
\description{
chatlog class
}
\section{Slots}{

\describe{
\item{\code{messages}}{The message data of the object}

\item{\code{chatlog_id}}{The chatlog's ID}
}}

