% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatlog.R
\name{chatlog}
\alias{chatlog}
\title{Extract chatlog from OpenAI API response object}
\usage{
chatlog(x)
}
\arguments{
\item{x}{either a list object representing a prompt response, or a character string representing the id of a current chatlog.}
}
\value{
a dataframe representing the chat log
}
\description{
This function extracts the current chatlog as a data.frame/tibble from the response of an API call to the
OpenAI chat completions endpoint.
}
