% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_output.R
\name{clean_output}
\alias{clean_output}
\title{Clean Output From Code Responses}
\usage{
clean_output(text)
}
\arguments{
\item{text}{A character string containing the code or text content with '```' markers.}
}
\value{
A character string containing the extracted code or text content.
}
\description{
This function extracts the content from a given text string that is enclosed between the '```' markers. It can be used to extract any kind of code or text content.
}
\examples{
code_text <- "```
example_code <- function(x) {
  return(x * 2)
}
```"
clean_output(code_text)
}
