% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_TSV.R
\name{cluster_TSV}
\alias{cluster_TSV}
\title{Calculate the Total Spatial Variance TSV for a given trajectory cluster dataset.}
\usage{
cluster_TSV(cluster_res, output = "TSV", cluster_sp = NULL, verbose = FALSE)
}
\arguments{
\item{cluster_res}{An output from trajCluster().}

\item{output}{one of either 'CSV' or 'TSV'. In the former case, a list of CSV values for all clusters is returned.}

\item{cluster_sp}{NULL or numeric to calculate CSV for a specific cluster.}

\item{verbose}{TRUE or FALSE to display progress text.}
}
\description{
Uses an adaptation of the intra-cluster variance quantification method employed by
the HYSPLIT clustering algorithm. See \code{\link{https://www.ready.noaa.gov/documents/Tutorial/html/traj_cluseqn.html}}
for the inspiration. It uses the haversine calculation to ensure distances between trajectory and cluster mean endpoints
are accurate at higher latitudes.
}
