% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{complete_upload}
\alias{complete_upload}
\title{Complete an upload (convenience function)}
\usage{
complete_upload(upload_id, part_ids, ...)
}
\arguments{
\item{upload_id}{Upload ID}

\item{part_ids}{Part IDs}

\item{...}{Additional parameters}
}
\value{
File object
}
\description{
Complete an upload (convenience function)
}
