% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{completions}
\alias{completions}
\title{Generate text using the OpenAI API}
\usage{
completions(
  prompt,
  model = "text-davinci-003",
  temperature = 0.5,
  max_tokens = 2048,
  n = 1,
  stop = "\\n",
  echo = FALSE,
  suffix = NULL,
  top_p = NULL,
  stream = FALSE,
  logprobs = NULL,
  presence_penalty = NULL,
  frequency_penalty = NULL,
  best_of = NULL,
  logit_bias = NULL,
  user = NULL
)
}
\arguments{
\item{prompt}{A character string containing the prompt to generate text from.}

\item{model}{A character string specifying the ID of the model to use. Default is "text-davinci-003". See list_models() for an overview of available models.}

\item{temperature}{An optional numeric scalar specifying the sampling temperature to use. The default value is 0.5.}

\item{max_tokens}{An optional integer scalar specifying the maximum number of tokens to generate in the text. The default value is 2048.}

\item{n}{An optional integer scalar specifying the number of text completions to generate. The default value is 1.}

\item{stop}{An optional character string or character vector specifying one or more stop sequences to use when generating the text.}

\item{echo}{An optional logical scalar specifying whether to include the prompt in the generated text. The default value is \code{FALSE}.}

\item{suffix}{An optional character string specifying a suffix to add to the generated text.}

\item{top_p}{An optional numeric scalar specifying the cumulative probability of the top tokens to use in the generated text. The default value is NULL.}

\item{stream}{An optional logical scalar specifying whether to generate text continuously. The default value is \code{FALSE}.}

\item{logprobs}{An optional integer scalar specifying the number of log probabilities to return for each token in the generated text. The default value is NULL.}

\item{presence_penalty}{An optional numeric scalar specifying the presence penalty to use when generating the text. The default value is NULL.}

\item{frequency_penalty}{An optional numeric scalar specifying the frequency penalty to use when generating the text. The default value is NULL.}

\item{best_of}{An optional integer scalar specifying the number of completions to generate and return the best one. The default value is NULL.}

\item{logit_bias}{An optional named numeric vector specifying the logit bias to use for each token in the generated text.}

\item{user}{A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse.}
}
\value{
A character vector containing the generated text(s).
}
\description{
This function generates natural language text given a prompt using the OpenAI API. It wraps the \code{completions} endpoint of the OpenAI API.
}
\examples{
\dontrun{
openai_api_key("your_api_key_here")
completions("Once upon a time,", n = 3)
}

}
