% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{CompletionsClient}
\alias{CompletionsClient}
\title{Completions Client (Legacy)}
\description{
Client for OpenAI Completions API (legacy).
For most use cases, use Chat Completions instead.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompletionsClient-new}{\code{CompletionsClient$new()}}
\item \href{#method-CompletionsClient-create}{\code{CompletionsClient$create()}}
\item \href{#method-CompletionsClient-clone}{\code{CompletionsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionsClient$create(
  prompt,
  model = "gpt-3.5-turbo-instruct",
  max_tokens = NULL,
  temperature = NULL,
  top_p = NULL,
  n = NULL,
  stream = NULL,
  logprobs = NULL,
  echo = NULL,
  stop = NULL,
  presence_penalty = NULL,
  frequency_penalty = NULL,
  best_of = NULL,
  logit_bias = NULL,
  user = NULL,
  callback = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
