% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalEval.R
\name{conditionalEval}
\alias{conditionalEval}
\title{Conditional quantile estimates with additional variables for model evaluation}
\usage{
conditionalEval(
  mydata,
  obs = "obs",
  mod = "mod",
  var.obs = "var.obs",
  var.mod = "var.mod",
  type = "default",
  bins = 31,
  statistic = "MB",
  cols = "YlOrRd",
  col.var = "Set1",
  var.names = NULL,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the field \code{obs} and \code{mod} representing
observed and modelled values.}

\item{obs}{The name of the observations in \code{mydata}.}

\item{mod}{The name of the predictions (modelled values) in \code{mydata}.}

\item{var.obs}{Other variable observations for which statistics should be
calculated. Can be more than length one e.g. \code{var.obs = c("nox.obs", "ws.obs")}.}

\item{var.mod}{Other variable predictions for which statistics should be
calculated. Can be more than length one e.g. \code{var.mod = c("nox.mod", "ws.mod")}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{bins}{Number of bins to be used in calculating the different quantile
levels.}

\item{statistic}{Statistic(s) to be plotted. Can be \dQuote{MB},
\dQuote{NMB}, \dQuote{r}, \dQuote{COE}, \dQuote{MGE}, \dQuote{NMGE},
\dQuote{RMSE} and \dQuote{FAC2}. \code{statistic} can also be a variable
name in the data frame or a date-based type (e.g. \dQuote{season}), in
which case the plot shows the proportions of that variable across the
prediction intervals. A special case is \dQuote{cluster}.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{col.var}{Colours for the additional variables. See \code{openColours}
for more details.}

\item{var.names}{Variable names to be shown in the legend for \code{var.obs}
and \code{var.mod}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Other graphical parameters passed onto \code{\link[=conditionalQuantile]{conditionalQuantile()}}
and \code{\link[=cutData]{cutData()}}.}
}
\description{
This function enhances \code{\link[=conditionalQuantile]{conditionalQuantile()}} by also considering how other
variables vary over the same intervals. Conditional quantiles are very useful
on their own for model evaluation, but provide no direct information on how
other variables change at the same time. For example, a conditional quantile
plot of ozone concentrations may show that low concentrations of ozone tend
to be under-predicted. However, the cause of the under-prediction can be
difficult to determine. However, by considering how well the model predicts
other variables over the same intervals, more insight can be gained into the
underlying reasons why model performance is poor.
}
\details{
The \code{conditionalEval} function provides information on how other
variables vary across the same intervals as shown on the conditional quantile
plot. There are two types of variable that can be considered by setting the
value of \code{statistic}. First, \code{statistic} can be another variable in
the data frame. In this case the plot will show the different proportions of
\code{statistic} across the range of predictions. For example \code{statistic = "season"} will show for each interval of \code{mod} the proportion of
predictions that were spring, summer, autumn or winter. This is useful
because if model performance is worse for example at high concentrations of
\code{mod} then knowing that these tend to occur during a particular season
etc. can be very helpful when trying to understand \emph{why} a model fails.
See \code{\link[=cutData]{cutData()}} for more details on the types of variable that can
be \code{statistic}. Another example would be \code{statistic = "ws"} (if
wind speed were available in the data frame), which would then split wind
speed into four quantiles and plot the proportions of each.

Second, \code{conditionalEval} can simultaneously plot the model performance
of other observed/predicted variable \strong{pairs} according to different
model evaluation statistics. These statistics derive from the
\code{\link[=modStats]{modStats()}} function and include \dQuote{MB}, \dQuote{NMB},
\dQuote{r}, \dQuote{COE}, \dQuote{MGE}, \dQuote{NMGE}, \dQuote{RMSE} and
\dQuote{FAC2}. More than one statistic can be supplied e.g. \code{statistic = c("NMB", "COE")}. Bootstrap samples are taken from the corresponding values
of other variables to be plotted and their statistics with 95\\% confidence
intervals calculated. In this case, the model \emph{performance} of other
variables is shown across the same intervals of \code{mod}, rather than just
the values of single variables. In this second case the model would need to
provide observed/predicted pairs of other variables.

For example, a model may provide predictions of NOx and wind speed (for which
there are also observations available). The \code{conditionalEval} function
will show how well these other variables are predicted for the same intervals
of the main variables assessed in the conditional quantile e.g. ozone. In
this case, values are supplied to \code{var.obs} (observed values for other
variables) and \code{var.mod} (modelled values for other variables). For
example, to consider how well the model predicts NOx and wind speed
\code{var.obs = c("nox.obs", "ws.obs")} and \code{var.mod = c("nox.mod", "ws.mod")} would be supplied (assuming \verb{nox.obs, nox.mod, ws.obs, ws.mod} are present in the data frame). The analysis could show for example,
when ozone concentrations are under-predicted, the model may also be shown to
over-predict concentrations of NOx at the same time, or under-predict wind
speeds. Such information can thus help identify the underlying causes of poor
model performance.

A special case is \code{statistic = "cluster"}. In this case a data frame is
provided that contains the cluster calculated by \code{\link[=trajCluster]{trajCluster()}}
and \code{\link[=importTraj]{importTraj()}}. Note that \code{statistic = "cluster"} cannot be
used together with the ordinary model evaluation statistics such as MB.
The output will be a bar chart showing the proportion of each interval of
\code{mod} by cluster number.
}
\references{
Wilks, D. S., 2005. Statistical Methods in the Atmospheric
Sciences, Volume 91, Second Edition (International Geophysics), 2nd
Edition. Academic Press.
}
\seealso{
The \code{verification} package for comprehensive functions for
forecast verification.

Other model evaluation functions: 
\code{\link{TaylorDiagram}()},
\code{\link{conditionalQuantile}()},
\code{\link{modStats}()}
}
\author{
David Carslaw
}
\concept{model evaluation functions}
