% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_roxygen.R
\name{contains_roxygen}
\alias{contains_roxygen}
\title{Check if a text file or character string contains Roxygen2 documentation}
\usage{
contains_roxygen(input)
}
\arguments{
\item{input}{A character string specifying the path to the text file, or a
character
  string containing the text to be checked for Roxygen2 documentation.}
}
\value{
A logical value (TRUE or FALSE) indicating whether the input
  contains Roxygen2 documentation.
}
\description{
This function parses a text file or a character string and returns TRUE if it
contains
Roxygen2 documentation. It uses a regex pattern to identify
possible Roxygen2 documentation lines.
}
