% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{Correlation matrices with conditioning}
\usage{
corPlot(
  mydata,
  pollutants = NULL,
  type = "default",
  cluster = TRUE,
  method = "pearson",
  use = "pairwise.complete.obs",
  annotate = c("cor", "signif", "stars", "none"),
  dendrogram = FALSE,
  triangle = c("both", "upper", "lower"),
  diagonal = TRUE,
  breaks = NULL,
  labels = NULL,
  cols = "default",
  r.thresh = 0.8,
  text.col = c("black", "black"),
  key.title = NULL,
  key.position = "right",
  strip.position = "top",
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame which should consist of some numeric columns.}

\item{pollutants}{the names of data-series in \code{mydata} to be plotted by
\code{corPlot}. The default option \code{NULL} and the alternative \code{"all"} use all
available valid (numeric) data.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{cluster}{Should the data be ordered according to cluster analysis. If
\code{TRUE} hierarchical clustering is applied to the correlation matrices using
\code{\link[=hclust]{hclust()}} to group similar variables together. With many variables
clustering can greatly assist interpretation.}

\item{method}{The correlation method to use. Can be \code{"pearson"}, \code{"spearman"}
or \code{"kendall"}.}

\item{use}{How to handle missing values in the \code{cor} function. The default is
\code{"pairwise.complete.obs"}. Care should be taken with the choice of how to
handle missing data when considering pair-wise correlations.}

\item{annotate}{What to annotate each correlation tile with. One of:
\itemize{
\item \code{"cor"}, the correlation coefficient to 2 decimal places.
\item \code{"signif"}, an X marker if the correlation is significant.
\item \code{"stars"}, standard significance stars.
\item \code{"none"}, no annotation.
}}

\item{dendrogram}{Should a dendrogram be plotted? When \code{TRUE} a dendrogram is
shown on the plot. Note that this will only work for \code{type = "default"}.
Defaults to \code{FALSE}.}

\item{triangle}{Which 'triangles' of the correlation plot should be shown?
Can be \code{"both"}, \code{"lower"} or \code{"upper"}. Defaults to \code{"both"}.}

\item{diagonal}{Should the 'diagonal' of the correlation plot be shown? The
diagonal of a correlation matrix is axiomatically always \code{1} as it
represents correlating a variable with itself. Defaults to \code{TRUE}.}

\item{breaks, labels}{If a categorical colour scale is required then \code{breaks}
should be specified. These should be provided as a numeric vector, e.g.,
\code{breaks = c(0, 50, 100, 1000)}. Users should set the maximum value of
\code{breaks} to exceed the maximum data value to ensure it is within the
maximum final range, e.g., 100--1000 in this case. Labels will
automatically be generated, but can be customised by passing a character
vector to \code{labels}, e.g., \code{labels = c("good", "bad", "very bad")}. In this
example, \code{0 - 50} will be \code{"good"} and so on. Note there is one less label
than break.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{r.thresh}{Values of greater than \code{r.thresh} will be shown in bold type.
This helps to highlight high correlations.}

\item{text.col}{The colour of the text used to show the correlation values.
The first value controls the colour of negative correlations and the second
positive.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label. All of these
are passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
Function to to draw and visualise correlation matrices. The primary purpose
is as a tool for exploratory data analysis. Hierarchical clustering is used
to group similar variables.
}
\details{
The \code{\link[=corPlot]{corPlot()}} function plots correlation matrices. The implementation
relies heavily on that shown in Sarkar (2007), with a few extensions.

Correlation matrices are a very effective way of understating relationships
between many variables. The \code{\link[=corPlot]{corPlot()}} shows the correlation coded in three
ways: by shape (ellipses), colour and the numeric value. The ellipses can be
thought of as visual representations of scatter plot. With a perfect positive
correlation a line at 45 degrees positive slope is drawn. For zero
correlation the shape becomes a circle. See examples below.

With many different variables it can be difficult to see relationships
between variables, i.e., which variables tend to behave most like one
another. For this reason hierarchical clustering is applied to the
correlation matrices to group variables that are most similar to one another
(if \code{cluster = TRUE}).

If clustering is chosen it is also possible to add a dendrogram using the
option \code{dendrogram = TRUE}. Note that dendrogramscan only be plotted for
\code{type = "default"} i.e. when there is only a single panel. The dendrogram can
also be recovered from the plot object itself and plotted more clearly; see
examples below.

It is also possible to use the \code{openair} type option to condition the data in
many flexible ways, although this may become difficult to visualise with too
many panels.
}
\examples{
# basic plot
corPlot(mydata)

\dontrun{
# plot by season
corPlot(mydata, type = "season")

# recover dendrogram when cluster = TRUE and plot it
res <- corPlot(mydata, plot = FALSE)
plot(res$clust)

# a more interesting are hydrocarbon measurements
hc <- importAURN(site = "my1", year = 2005, hc = TRUE)

# now it is possible to see the hydrocarbons that behave most
# similarly to one another
corPlot(hc)
}

}
\author{
David Carslaw

Jack Davison

Adapted from the approach taken by Sarkar (2007)
}
