% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{create_assistant}
\alias{create_assistant}
\title{Create an assistant (convenience function)}
\usage{
create_assistant(model, ...)
}
\arguments{
\item{model}{Model ID}

\item{...}{Additional parameters}
}
\value{
Assistant object
}
\description{
Create an assistant (convenience function)
}
