% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{create_batch}
\alias{create_batch}
\title{Create a batch (convenience function)}
\usage{
create_batch(input_file_id, endpoint, ...)
}
\arguments{
\item{input_file_id}{File ID}

\item{endpoint}{Endpoint}

\item{...}{Additional parameters}
}
\value{
Batch object
}
\description{
Create a batch (convenience function)
}
