% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{create_completion}
\alias{create_completion}
\title{Create a completion (convenience function)}
\usage{
create_completion(prompt, model = "gpt-3.5-turbo-instruct", ...)
}
\arguments{
\item{prompt}{Text prompt}

\item{model}{Model to use}

\item{...}{Additional parameters}
}
\value{
Completion response
}
\description{
Create a completion (convenience function)
}
