% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{create_message}
\alias{create_message}
\title{Create a message (convenience function)}
\usage{
create_message(thread_id, role, content, ...)
}
\arguments{
\item{thread_id}{Thread ID}

\item{role}{Message role}

\item{content}{Message content}

\item{...}{Additional parameters}
}
\value{
Message object
}
\description{
Create a message (convenience function)
}
