% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderations.R
\name{create_moderation}
\alias{create_moderation}
\title{Create a moderation (convenience function)}
\usage{
create_moderation(input, model = "omni-moderation-latest")
}
\arguments{
\item{input}{Text or texts to moderate}

\item{model}{Moderation model}
}
\value{
Moderation response
}
\description{
Create a moderation (convenience function)
}
