% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{create_response}
\alias{create_response}
\title{Create a response (convenience function)}
\usage{
create_response(model, input, ...)
}
\arguments{
\item{model}{Model ID}

\item{input}{Input}

\item{...}{Additional parameters}
}
\value{
Response object
}
\description{
Create a response (convenience function)
}
