% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{create_run}
\alias{create_run}
\title{Create a run (convenience function)}
\usage{
create_run(thread_id, assistant_id, ...)
}
\arguments{
\item{thread_id}{Thread ID}

\item{assistant_id}{Assistant ID}

\item{...}{Additional parameters}
}
\value{
Run object
}
\description{
Create a run (convenience function)
}
