% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{create_thread}
\alias{create_thread}
\title{Create a thread (convenience function)}
\usage{
create_thread(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
Thread object
}
\description{
Create a thread (convenience function)
}
