% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{create_upload}
\alias{create_upload}
\title{Create an upload (convenience function)}
\usage{
create_upload(purpose, filename, bytes, ...)
}
\arguments{
\item{purpose}{Upload purpose}

\item{filename}{Filename}

\item{bytes}{File size}

\item{...}{Additional parameters}
}
\value{
Upload object
}
\description{
Create an upload (convenience function)
}
