% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{create_vector_store}
\alias{create_vector_store}
\title{Create a vector store (convenience function)}
\usage{
create_vector_store(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
Vector store object
}
\description{
Create a vector store (convenience function)
}
