% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/created.R
\name{created}
\alias{created}
\title{Extract Created DateTime from OpenAI API response}
\usage{
created(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a Date object representing the date-time string of when the response was created
}
\description{
This function extracts the date-time string of when the response was created from the parsed HTTP response of an API call to the
OpenAI chat completions endpoint.
}
