% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutData.R
\name{cutData}
\alias{cutData}
\alias{cutData.data.frame}
\alias{cutData.default}
\alias{cutData.numeric}
\alias{cutData.double}
\alias{cutData.integer}
\alias{cutData.character}
\alias{cutData.factor}
\alias{cutData.POSIXct}
\alias{cutData.Date}
\title{Function to split data in different ways for conditioning}
\usage{
cutData(
  x,
  type = "default",
  drop = c("default", "empty", "outside", "none"),
  ...
)

\method{cutData}{data.frame}(
  x,
  type = "default",
  names = NULL,
  suffix = NULL,
  hemisphere = "northern",
  n.levels = 4,
  start.day = 1,
  is.axis = FALSE,
  local.tz = NULL,
  latitude = 51,
  longitude = -0.5,
  drop = c("default", "empty", "outside", "none"),
  ...
)

\method{cutData}{default}(x, ...)

\method{cutData}{numeric}(x, type = c("default", "wd"), n.levels = 4, drop = "default", ...)

\method{cutData}{double}(x, type = c("default", "wd"), n.levels = 4, drop = "default", ...)

\method{cutData}{integer}(x, type = c("default", "wd"), n.levels = 4, drop = "default", ...)

\method{cutData}{character}(x, type = "default", ...)

\method{cutData}{factor}(x, type = "default", ...)

\method{cutData}{POSIXct}(
  x,
  type = "default",
  hemisphere = "northern",
  start.day = 1,
  is.axis = FALSE,
  local.tz = NULL,
  latitude = 51,
  longitude = -0.5,
  drop = c("default", "empty", "outside", "none"),
  ...
)

\method{cutData}{Date}(
  x,
  type = "default",
  hemisphere = "northern",
  start.day = 1,
  is.axis = FALSE,
  local.tz = NULL,
  latitude = 51,
  longitude = -0.5,
  drop = c("default", "empty", "outside", "none"),
  ...
)
}
\arguments{
\item{x}{A data frame or vector. Data frames should contain a POSIXCt or Date
column named \code{date} for time-dependent \code{type} options.}

\item{type}{A string giving the way in which the data should be split (e.g.,
\code{"default"}, \code{"year"}, \code{"hour"}, \code{"month"}, \code{"season"}, etc.) or, for data
frames, a column name in \code{x} to be conditioned on (e.g., \code{"no2"}, \code{"site"},
etc.). See \code{\link[=cutData]{cutData()}} for all options.}

\item{drop}{How to handle empty factor levels. One of:
\itemize{
\item \code{"default"}: Sensible defaults selected on a case-by-case basis for
different \code{type} options.
\item \code{"empty"}: Drop all empty factor levels.
\item \code{"none"}: Retain all empty factor levels, where possible. For example,
for \code{type = "hour"}, all factor levels from \code{0} and \code{23} will be
represented.
\item \code{"outside"}: Retain empty factor levels within the range of the data.
For example, for \code{type = "hour"} when the data only contains data for 1 AM
and 5 AM, the factor levels, \code{1}, \code{2}, \code{3}, \code{4} and \code{5} will be retained.
}

Some of these options only apply to certain \code{type} options. For example,
for \code{type = "year"}, \code{"outside"} is equivalent to \code{"none"} as there is no
fixed range of years to use in the \code{"none"} case.}

\item{...}{All additional parameters are passed on to next function(s).}

\item{names}{By default, the columns created by \code{\link[=cutData]{cutData()}} are named after
their \code{type} option. Specifying \code{names} defines other names for the
columns, which map onto the \code{type} options in the same order they are
given. The length of \code{names} should therefore be equal to the length of
\code{type}.}

\item{suffix}{If \code{name} is not specified, \code{suffix} will be appended to any
added columns that would otherwise overwrite existing columns. For example,
\code{cutData(mydata, "nox", suffix = "_cuts")} would append a \code{nox_cuts} column
rather than overwriting \code{nox}.}

\item{hemisphere}{Can be \code{"northern"} or \code{"southern"}, used when \code{type = "season"} or \code{type = "seasonyear"}.}

\item{n.levels}{Number of quantiles to split numeric data into.}

\item{start.day}{What day of the week should the \code{type = "weekday"} start on?
The user can change the start day by supplying an integer between 0 and 6.
Sunday = 0, Monday = 1, ... For example to start the weekday plots on a
Saturday, choose \code{start.day = 6}.}

\item{is.axis}{A logical (\code{TRUE}/\code{FALSE}), used to request shortened cut
labels for axes.}

\item{local.tz}{Used for identifying whether a date has daylight savings time
(DST) applied or not. Examples include \code{local.tz = "Europe/London"},
\code{local.tz = "America/New_York"}, i.e., time zones that assume DST.
\url{https://en.wikipedia.org/wiki/List_of_zoneinfo_time_zones} shows time
zones that should be valid for most systems. It is important that the
original data are in GMT (UTC) or a fixed offset from GMT.}

\item{latitude, longitude}{The decimal latitude and longitudes used when \code{type = "daylight"}. Note that locations west of Greenwich have negative
longitudes.}
}
\value{
If \code{x} is a data.frame, returns the data frame with columns appended
as defined by \code{type} and \code{name}. If \code{x} is a vector, returns a factor
vector.
}
\description{
\code{\link[=cutData]{cutData()}} splits data up in various ways for use in plots and other
functions. It is widely used on \code{data.frame}s by many \code{openair} functions
usually through the option \code{type}. Can be used on a \code{data.frame} to create
new column(s) will with the data splits, or on various vector types to return
split vectors.
}
\section{\code{type} options}{


This section give a brief description of each of the define levels of
\code{type}.
\subsection{The \code{"default"} Type}{
\itemize{
\item \code{"default"} does not split the data but will describe the levels as a date
range in the format "day month year" if a \code{date} column is available, or
return a generic \code{"all data"} if not. For numeric vectors, \code{"default"} uses
default numeric data handling described below.
}
}

\subsection{Numeric Types (NB: these require any numeric column)}{
\itemize{
\item Any numeric column (e.g., \code{"no2"}) splits the data into \code{n.levels} quantiles.
\item \code{"wd"} splits the data by 8 wind sectors and requires a column \code{wd}: "NE",
"E", "SE", "S", "SW", "W", "NW", "N".
}
}

\subsection{Time-Dependent Types (NB: these require a column \code{date})}{
\itemize{
\item \code{"year"} splits the data by each year.
\item \code{"month"} splits the data by month of the year.
\item \code{"hour"} splits the data by hour of the day.
\item \code{"monthyear"} splits the data by year and month. It differs from month in
that a level is defined for each month of the data set. This is useful
sometimes to show an ordered sequence of months if the data set starts half
way through a year; rather than starting in January.
\item \code{"weekend"} splits the data by weekday and weekend.
\item \code{"weekday"} splits the data by day of the week - ordered to start Monday.
\item \code{"season"} splits data up by season. In the northern hemisphere winter =
December, January, February; spring = March, April, May etc. These
definitions will change of \code{hemisphere = "southern"}.
\item \code{"seasonyear"} (or \code{"yearseason"}) will split the data into year-season
intervals, keeping the months of a season together. For example, December
2010 is considered as part of winter 2011 (with January and February 2011).
This makes it easier to consider contiguous seasons. In contrast, \code{type =   "season"} will just split the data into four seasons regardless of the
year.
\item \code{"daylight"} splits the data relative to estimated sunrise and sunset to
give either daylight or nighttime. The cut is made by \code{cutDaylight} but
more conveniently accessed via \code{cutData}, e.g. \code{cutData(mydata, type =   "daylight", latitude = my.latitude, longitude = my.longitude)}. The
daylight estimation, which is valid for dates between 1901 and 2099, is
made using the measurement location, date, time and astronomical algorithms
to estimate the relative positions of the Sun and the measurement location
on the Earth's surface, and is based on NOAA methods. Measurement location
should be set using \code{latitude} (+ to North; - to South) and \code{longitude} (+
to East; - to West).
\item \code{"dst"} will split the data by hours that are in daylight saving time (DST)
and hours that are not for appropriate time zones. The option also requires
that the local time zone is given e.g. \code{local.tz = "Europe/London"},
\code{local.tz = "America/New_York"}. Each of the two periods will be in
\emph{local time}. The main purpose of this option is to test whether there
is a shift in the diurnal profile when DST and non-DST hours are compared.
This option is particularly useful with the \code{\link[=timeVariation]{timeVariation()}} function. For
example, close to the source of road vehicle emissions, "rush-hour" will
tend to occur at the same \emph{local time} throughout the year, e.g., 8 am and
5 pm. Therefore, comparing non-DST hours with DST hours will tend to show
similar diurnal patterns (at least in the timing of the peaks, if not
magnitude) when expressed in local time. By contrast a variable such as
wind speed or temperature should show a clear shift when expressed in local
time. In essence, this option when used with \code{timeVariation()} may help
determine whether the variation in a pollutant is driven by man-made
emissions or natural processes.
}
}
}

\section{Methods}{


This function is a \strong{generic}, which means that packages can provide
implementations (methods) for other classes. See the documentation of
individual methods for extra arguments and differences in behaviour.
}

\examples{
# split data by day of the week
mydata <- cutData(mydata, type = "weekday")
names(mydata)
head(mydata)

# use many types
cutData(mydata, type = c("weekday", "weekend", "no2"))

# use different names
cutData(mydata, type = c("wd", "ws"), names = c("winddir", "windspd"))
cutData(mydata, type = c("no2", "nox"), suffix = "_cuts")

# apply to vectors
head(mydata$no2)
head(cutData(mydata$no2))

head(cutData(mydata$date, "weekday"))
}
\author{
David Carslaw

Jack Davison

Karl Ropkins (\code{"daylight"} option)
}
