% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{delete_assistant}
\alias{delete_assistant}
\title{Delete an assistant (convenience function)}
\usage{
delete_assistant(assistant_id)
}
\arguments{
\item{assistant_id}{Assistant ID}
}
\value{
Deletion status
}
\description{
Delete an assistant (convenience function)
}
