% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file (convenience function)}
\usage{
delete_file(file_id)
}
\arguments{
\item{file_id}{File ID}
}
\value{
Deletion status
}
\description{
Delete a file (convenience function)
}
