% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{delete_response}
\alias{delete_response}
\title{Delete a response (convenience function)}
\usage{
delete_response(response_id)
}
\arguments{
\item{response_id}{Response ID}
}
\value{
Deletion status
}
\description{
Delete a response (convenience function)
}
