% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{delete_thread}
\alias{delete_thread}
\title{Delete a thread (convenience function)}
\usage{
delete_thread(thread_id)
}
\arguments{
\item{thread_id}{Thread ID}
}
\value{
Deletion status
}
\description{
Delete a thread (convenience function)
}
