% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{delete_vector_store}
\alias{delete_vector_store}
\title{Delete a vector store (convenience function)}
\usage{
delete_vector_store(vector_store_id)
}
\arguments{
\item{vector_store_id}{Vector store ID}
}
\value{
Deletion status
}
\description{
Delete a vector store (convenience function)
}
