% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise_opts.R
\name{disc_breaks}
\alias{disc_breaks}
\alias{disc_interval}
\alias{disc_number}
\alias{disc_width}
\alias{disc_width_n}
\alias{disc_pretty}
\title{Discretise continuous scales in \code{openair} plotting functions}
\usage{
disc_breaks(breaks, labels = NULL, closed = "right", dig.lab = 3)

disc_interval(n = NULL, length = NULL, labels = NULL)

disc_number(n = NULL, labels = NULL)

disc_width(
  width = NULL,
  center = NULL,
  boundary = NULL,
  labels = NULL,
  closed = "right"
)

disc_width_n(
  width = NULL,
  n_ints = 4L,
  start = NULL,
  closed = "right",
  dig.lab = 3
)

disc_pretty(n = NULL, closed = "right", dig.lab = 3)
}
\arguments{
\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{labels}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether right
or left edges of bins are included in the bin.}

\item{dig.lab}{integer which is used when labels are not given.  It
    determines the number of digits used in formatting the break numbers.}

\item{n}{number of intervals to create, OR}

\item{length}{length of each interval}

\item{width}{The bin width.}

\item{center, boundary}{Specify either the position of edge or the center of
a bin. Since all bins are aligned, specifying the position of a single bin
(which doesn't need to be in the range of the data) affects the location of
all bins. If not specified, uses the "tile layers algorithm", and sets
the boundary to half of the binwidth.

To center on integers, \code{width = 1} and \code{center = 0}.
\code{boundary = 0.5}.}

\item{n_ints}{In \code{\link[=disc_width_n]{disc_width_n()}}, the number of intervals to create using
\code{width}.}

\item{start}{In \code{\link[=disc_width_n]{disc_width_n()}}, where to start interval creation. If
\code{NULL}, will pick a 'pretty' start point based on the scale limits.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are used in the \code{discretise} argument of many \code{openair}
plotting functions to bin a continuous colour scale.
\itemize{
\item \code{\link[=disc_breaks]{disc_breaks()}} is a way to define breaks directly, by giving direct access to the \code{breaks} argument.
\item \code{\link[=disc_interval]{disc_interval()}}, \code{\link[=disc_number]{disc_number()}} and \code{\link[=disc_width]{disc_width()}} use \code{\link[ggplot2:cut_interval]{ggplot2::cut_interval()}}, \code{\link[ggplot2:cut_interval]{ggplot2::cut_number()}} or \code{\link[ggplot2:cut_interval]{ggplot2::cut_width()}}, respectively. These give convenient ways to discretise a scale.
\item \code{\link[=disc_width_n]{disc_width_n()}} acts like \code{\link[=disc_width]{disc_width()}} for a certain number of intervals, then combines the rest of the data into a single bin. This is similar logic to \code{\link[=windRose]{windRose()}}.
\item \code{\link[=disc_pretty]{disc_pretty()}} uses \code{\link[=pretty]{pretty()}} to work out pleasantly spaced breakpoints. Note that \code{n} is \emph{approximate} in this case; it may not be possible to create exactly \code{n} 'pretty' breakpoints. The upper and lower limits are also not likely to be 'pretty' as they will be set to the limits of the scale.
}
}
\examples{
plot_heatmap(mydata, "no2")

# 5 groups with equal range
plot_heatmap(mydata, "no2", discretise = disc_interval(5))

# 5 groups with the same number of observations
plot_heatmap(mydata, "no2", discretise = disc_number(5))

# however many groups with width 5
plot_heatmap(mydata, "no2", discretise = disc_width(5))

# defined (irregular) breaks
plot_heatmap(
  mydata,
  "no2",
  discretise = disc_breaks(c(20, 25, 30, 40, 50, 100))
)
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{facet_opts}()},
\code{\link{scale_opts}()},
\code{\link{windflow_opts}()}
}
\concept{ggplot2 plot utilities}
