% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise.R
\name{discretise_breaks}
\alias{discretise_breaks}
\alias{discretise_interval}
\alias{discretise_number}
\alias{discretise_width}
\title{Discretise continuous scales in \code{openair} plotting functions}
\usage{
discretise_breaks(
  breaks,
  labels = NULL,
  include.lowest = FALSE,
  right = TRUE,
  dig.lab = 3
)

discretise_interval(n = NULL, length = NULL, labels = NULL)

discretise_number(n = NULL, labels = NULL)

discretise_width(
  width = NULL,
  center = NULL,
  boundary = NULL,
  labels = NULL,
  closed = "right"
)
}
\arguments{
\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{labels}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}

\item{include.lowest}{logical, indicating if an \sQuote{x[i]} equal to
    the lowest (or highest, for \code{right = FALSE}) \sQuote{breaks}
    value should be included.}

\item{right}{logical, indicating if the intervals should be closed on
    the right (and open on the left) or vice versa.}

\item{dig.lab}{integer which is used when labels are not given.  It
    determines the number of digits used in formatting the break numbers.}

\item{n}{number of intervals to create, OR}

\item{length}{length of each interval}

\item{width}{The bin width.}

\item{center, boundary}{Specify either the position of edge or the center of
a bin. Since all bins are aligned, specifying the position of a single bin
(which doesn't need to be in the range of the data) affects the location of
all bins. If not specified, uses the "tile layers algorithm", and sets
the boundary to half of the binwidth.

To center on integers, \code{width = 1} and \code{center = 0}.
\code{boundary = 0.5}.}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether right
or left edges of bins are included in the bin.}
}
\description{
These functions are used in the \code{discretise} argument of many \code{openair}
plotting functions to bin a continuous colour scale. \code{\link[=discretise_breaks]{discretise_breaks()}} is
a way to define breaks directly. The other functions use one of
\code{\link[ggplot2:cut_interval]{ggplot2::cut_interval()}}, \code{\link[ggplot2:cut_interval]{ggplot2::cut_number()}} or \code{\link[ggplot2:cut_interval]{ggplot2::cut_width()}}
to discretise a continuous scale, which can be more useful short-hands.
}
\examples{
plot_heatmap(mydata, "no2")

# 5 groups with equal range
plot_heatmap(mydata, "no2", discretise = discretise_interval(5))

# 5 groups with the same number of observations
plot_heatmap(mydata, "no2", discretise = discretise_number(5))

# however many groups with width 5
plot_heatmap(mydata, "no2", discretise = discretise_width(5))

# defined (irregular) breaks
plot_heatmap(
  mydata,
  "no2",
  discretise = discretise_breaks(c(20, 25, 30, 40, 50, 100))
)
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{facet_opts}()},
\code{\link{scale_opts}()},
\code{\link{windflow_opts}()}
}
\concept{ggplot2 plot utilities}
