% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{discretise_opts}
\alias{discretise_opts}
\title{Define 'discretisation' options for openair \code{ggplot2} plots}
\usage{
discretise_opts(
  method = c("interval", "number", "width", "breaks"),
  n = NULL,
  length = NULL,
  width = NULL,
  labels = NULL,
  breaks = NULL
)
}
\arguments{
\item{method}{How the discretisation should occur. One of:
\itemize{
\item \code{"intervals"} (the default, which makes \emph{n} groups with equal range. Requires \code{n} or \code{length})
\item \code{"number"} (which makes \emph{n} groups with roughly equal numbers of observations. Requires \code{n})
\item \code{"width"} (which makes groups of width \emph{width}. Requires \code{width}.)
\item \code{"breaks"} (where \code{breaks} must be set manually)
}}

\item{n}{number of intervals to create, OR}

\item{length}{length of each interval}

\item{width}{The bin width.}

\item{labels}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}
}
\description{
This function conveniently allows for certain plots which plot a continuous
scale.
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{facet_opts}()}
}
\author{
Jack Davison
}
\concept{ggplot2 plot utilities}
