% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distAngle}
\alias{distAngle}
\title{Optimized Distance matrix based on similarity of trajectory angles}
\usage{
distAngle(Lonm, Latm)
}
\description{
Optimized Distance matrix based on similarity of trajectory angles
}
\keyword{internal}
