% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{explain_code_input}
\alias{explain_code_input}
\title{Input: message content template for explain_code()}
\format{
## `feedback_input`
A data frame with 2 rows and 2 columns representing the message object in API
calls.
}
\source{
Contributed by jldc
}
\usage{
explain_code_input
}
\description{
This message content serves as user input for the feedback function.
}
\keyword{datasets}
