% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_blocks_content.R
\name{extract_blocks_content}
\alias{extract_blocks_content}
\title{Extract the content parts of blocks}
\usage{
extract_blocks_content(block_list)
}
\arguments{
\item{block_list}{A list of blocks to extract the content from}
}
\value{
A list of content parts
}
\description{
This function takes a list of blocks and returns a list of their content
parts.
}
