% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extract_entities_prompt}
\alias{extract_entities_prompt}
\title{Input: message content template for extract_entities()}
\format{
## `extract_entities_prompt`
A data frame with 2 rows and 2 columns representing the messages object in
API calls.
}
\source{
Contributed by umatter
}
\usage{
extract_entities_prompt
}
\description{
This message content serves as user input for the extract_entities function.
}
\keyword{datasets}
