% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info.COMPLETIONS}
\alias{extract_info.COMPLETIONS}
\title{Generic Method for COMPLETIONS}
\usage{
\method{extract_info}{COMPLETIONS}(object, ...)
}
\arguments{
\item{object}{A COMPLETIONS object containing response of system}

\item{...}{Additional arguments (ignored)}
}
\value{
An character object
}
\description{
Generic Method for COMPLETIONS
}
