% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info.EMBEDDINGS}
\alias{extract_info.EMBEDDINGS}
\title{Generic Method for EMBEDDINGS}
\usage{
\method{extract_info}{EMBEDDINGS}(object, ...)
}
\arguments{
\item{object}{A EMBEDDINGS object containing embeddings data}

\item{...}{Additional arguments (ignored)}
}
\value{
An embeddings vector.
}
\description{
Generic Method for EMBEDDINGS
}
