% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Generic Function}
\usage{
extract_info(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments (ignored)}
}
\description{
Generic Function
}
