% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{facet_opts}
\alias{facet_opts}
\title{Define 'faceting' options for openair \code{ggplot2} plots}
\usage{
facet_opts(
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  space = "fixed",
  labeller = "label_value",
  drop = FALSE,
  strip.position = "top",
  switch = NULL,
  axes = "margins",
  axis.labels = "all"
)
}
\arguments{
\item{nrow, ncol}{The number of rows and/or columns for a 'wrap' layout (i.e.,
when there is only one \code{type}/'faceting' variable).}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:label_parsed]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:label_value]{label_value()}} is used by default,
check it for more details and pointers to other options.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{axes}{Determines which axes will be drawn. When \code{"margins"}
(default), axes will be drawn at the exterior margins. \code{"all_x"} and
\code{"all_y"} will draw the respective axes at the interior panels too, whereas
\code{"all"} will draw all axes at all panels.}

\item{axis.labels}{Determines whether to draw labels for interior axes when
the \code{axes} argument is not \code{"margins"}. When \code{"all"} (default), all
interior axes get labels. When \code{"margins"}, only the exterior axes get
labels and the interior axes get none. When \code{"all_x"} or \code{"all_y"}, only
draws the labels at the interior axes in the x- or y-direction
respectively.}
}
\description{
This function conveniently allows for options to be defined in relation to
how plots are \emph{faceted} (i.e., split into multiple panels). These arguments
are automatically passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}
or similar extension functions depending on the value the user has passed to
\code{type}.
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{discretise_breaks}()},
\code{\link{windflow_opts}()}
}
\author{
Jack Davison
}
\concept{ggplot2 plot utilities}
