% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\title{Obtain feedback on code using OpenAI's chat_completion function}
\usage{
feedback(task, code)
}
\arguments{
\item{task}{A character string that represents the task the user is trying
to accomplish.}

\item{code}{A character string that represents the code that the user has
written.}
}
\value{
A data frame with two columns:
\itemize{
 \item \code{lines}: A character vector containing the lines of code from
the user's input.
 \item \code{feedback}: A character vector containing feedback on each
line of code.
}
}
\description{
This function takes two arguments: task and code. It returns a processed
feedback data frame based on the user input of code and task. The feedback
is obtained by using OpenAI's chat_completion function and processing the
response.
}
\examples{
feedback("Solve the FizzBuzz problem", "for(i in 1:100)
{ if(i \%\% 3 == 0 && i \%\% 5 == 0) { print('FizzBuzz') } else if(i \%\% 3 == 0)
{ print('Fizz') } else if(i \%\% 5 == 0)
{ print('Buzz') } else { print(i) } }")

}
