% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_smooth.R
\name{GaussianSmooth}
\alias{GaussianSmooth}
\title{Calculate rolling Gaussian smooth of pollutant values}
\usage{
GaussianSmooth(
  mydata,
  pollutant = "o3",
  sigma = 24L,
  type = "default",
  data.thresh = 0,
  new.name = NULL,
  date.pad = FALSE,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing a \code{date} field. \code{mydata} must contain a
\code{date} field in \code{Date} or \code{POSIXct} format. The input time series must be
regular, e.g., hourly, daily.}

\item{pollutant}{The name of a pollutant, e.g., \code{pollutant = "o3"}.}

\item{sigma}{The value of \code{sigma} to use in the Gaussian.}

\item{type}{Used for splitting the data further. Passed to \code{\link[=cutData]{cutData()}}.}

\item{data.thresh}{The \% data capture threshold. No values are calculated if
data capture over the period of interest is less than this value.}

\item{new.name}{The name given to the new column. If not supplied it will
create a name based on the name of the pollutant and the averaging period
used.}

\item{date.pad}{Should missing dates be padded? Default is \code{FALSE}.}

\item{...}{Additional parameters passed to \code{\link[=cutData]{cutData()}}. For use with \code{type}.}
}
\value{
A tibble with two new columns for the Gaussian smooth value.
}
\description{
This is a utility function designed to calculate rolling Gaussian smooth (kernel smoothing). The function will try
and fill in missing time gaps to get a full time sequence but return a data
frame with the same number of rows supplied.
}
\details{
The function provides centre-aligned smoothing out to 3 sigma, which captures 99.7\% of the data.
}
\examples{
# Gaussian smoother with sigma = 24
mydata <- GaussianSmooth(mydata,
  pollutant = "o3", sigma = 24, data.thresh = 75)
}
\author{
David Carslaw
}
