% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chatlog.R
\name{get_chatlog}
\alias{get_chatlog}
\title{Get the current chatlog}
\usage{
get_chatlog(x = ".__CURRENTCHAT__")
}
\arguments{
\item{x}{either a chatlog object r, or a character string representing the id of a chatlog (the default is ".__CURRENTCHAT__", the current log of the chat()-function)}
}
\value{
a chatlog object
}
\description{
This function extracts the current chatlog of a given chat from the chat environment.
}
\author{
Ulrich Matter umatter@protonmail.com
}
