% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embeddings.R
\name{get_embeddings}
\alias{get_embeddings}
\title{get embeddings using openAI api}
\usage{
get_embeddings(
  input_text,
  api_key,
  api_base_url = "https://api.openai.com/v1/embeddings",
  model = "text-embedding-ada-002",
  max_tries = 1
)
}
\arguments{
\item{input_text}{The text to get embeddings for.}

\item{api_key}{Your OpenAI API key.}

\item{api_base_url}{The base URL for the API (default is "https://api.openai.com/v1/embeddings").}

\item{model}{A model to use for embeddings (default is "text-embedding-ada-002")}

\item{max_tries}{Automatically retry if the request fails.}
}
\value{
A list object containing embeddings.
}
\description{
This function retrieves embeddings for a given text using the OpenAI API. It provides a simple interface for embedding text into a vector space, which can be useful for a variety of natural language processing tasks. For more details, see "https://platform.openai.com/docs/guides/embeddings."
}
