% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{id}
\alias{id}
\title{Extract ID from OpenAI API response}
\usage{
id(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a character string representing the ID from the response
}
\description{
This function extracts the ID from the parsed HTTP response of an API call
to the OpenAI chat completions endpoint.
}
\author{
Ulrich Matter umatter@protonmail.com
}
