% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importADMS.R
\name{importADMS}
\alias{importADMS}
\alias{importADMSBgd}
\alias{importADMSMet}
\alias{importADMSMop}
\alias{importADMSPst}
\title{CERC Atmospheric Dispersion Modelling System (ADMS) data import function(s)
for openair}
\usage{
importADMS(
  file = file.choose(),
  file.type = "unknown",
  drop.case = TRUE,
  drop.input.dates = TRUE,
  keep.units = TRUE,
  simplify.names = TRUE,
  test.file.structure = TRUE,
  drop.delim = TRUE,
  add.prefixes = TRUE,
  names = NULL,
  all = FALSE,
  ...
)
}
\arguments{
\item{file}{The ADMS file to be imported. Default, \code{\link[=file.choose]{file.choose()}} opens
browser. Use of \code{\link[utils:read.csv]{utils::read.csv()}} also allows this to be a readable
text-mode connection or url (although these options are currently not fully
tested).}

\item{file.type}{Type of ADMS file to be imported. With default, "unknown",
the import uses the file extension to identify the file type and, where
recognised, uses this to identify the file structure and import method to
be applied. Where file extension is not recognised the choice may be forced
by setting \code{file.type} to one of the known \code{file.type} options:
"bgd", "met", "mop" or "pst".}

\item{drop.case}{Option to convert all data names to lower case. Default,
\code{TRUE}.  Alternative, \code{FALSE}, returns data with name cases as
defined in file.}

\item{drop.input.dates}{Option to remove ADMS "hour", "day", and "year" data
columns after generating openair "date" timeseries. Default, \code{TRUE}.
Alternative, \code{FALSE}, returns both "date" and the associated ADMS data
columns as part of openair data frame.}

\item{keep.units}{Option to retain ADMS data units. Default, \code{TRUE},
retains units (if recoverable) as character vector in data frame comment if
defined in \code{file}.  Alternative, \code{FALSE}, discards units. (NOTE:
currently, only \code{.bgd} and \code{.pst} files assign units. So, this
option is ignored when importing \code{.met} or \code{.mop} files.)}

\item{simplify.names}{Option to simplify data names in accordance with common
\code{openair} practices. Default, \code{TRUE}. Alternative, \code{FALSE},
returns data with names as interpreted by standard R. (NOTE: Some ADMS file
data names include symbols and structures that R does not allow as part of
a name, so some renaming is automatic regardless of \code{simplify.names}
setting.  For example, brackets or symbols are removed from names or
replaced with ".", and names in the form "1/x" may be returned as "X1.x" or
"recip.x".)}

\item{test.file.structure}{Option to test file structure before trying to
import. Default, \code{TRUE}, tests for expected file structure and halts
import operation if this is not found.  Alternative, \code{FALSE}, attempts
import regardless of structure.}

\item{drop.delim}{Option to remove delim columns from the data frame. ADMS
.mop files include two columns, "INPUT_DATA:" and "PROCESSED_DATA:", to
separate model input and output types.  Default, \code{TRUE}, removes
these. Alternative, \code{FALSE}, retains them as part of import.  (Note:
Option ignored when importing \code{.bgd}, \code{.met} or \code{.pst}
files.)}

\item{add.prefixes}{Option to add prefixes to data names. ADMS .mop files
include a number of input and process data types with shared names.
Prefixes can be automatically added to these so individual data can be
readily identified in the R/openair environment. Default, \code{TRUE}, adds
"process." as a prefix to processed data. Other options include:
\code{FALSE} which uses no prefixes and leave all name rationalisation to
R, and character vectors which are treated as the required prefixes. If one
vector is sent, this is treated as processed data prefix. If two (or more)
vectors are sent, the first and second are treated as the input and
processed data prefixes, respectively. For example, the argument
(\code{add.prefixes="out"}) would add the "out" prefix to processed data
names, while the argument (\code{add.prefixes=c("in","out")}) would add
"in" and "out" prefixes to input and output data names, respectively.
(Note: Option ignored when importing \code{.bgd}, \code{.met} or
\code{.pst} files.)}

\item{names}{Option applied by \code{simplifyNamesADMS} when
\code{simplify.names} is enabled.  All names are simplified for the default
setting, \code{NULL}.}

\item{all}{For .MOP files, return all variables or not. If \code{all = TRUE}
a large number of processed variables are returned.}

\item{...}{
  Arguments passed on to \code{\link[utils:read.csv]{utils::read.csv}}
  \describe{
    \item{\code{header}}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}
    \item{\code{sep}}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}
    \item{\code{quote}}{the set of quoting characters. To disable quoting
    altogether, use \code{quote = ""}.  See \code{\link{scan}} for the
    behaviour on quotes embedded in quotes.  Quoting is only considered
    for columns read as character, which is all of them unless
    \code{colClasses} is specified.}
    \item{\code{dec}}{the character used in the file for decimal points.}
    \item{\code{numerals}}{string indicating how to convert numbers whose conversion
    to double precision would lose accuracy, see \code{\link[utils]{type.convert}}.
    Can be abbreviated.  (Applies also to complex-number inputs.)}
    \item{\code{row.names}}{a vector of row names.  This can be a vector giving
    the actual row names, or a single number giving the column of the
    table which contains the row names, or character string giving the
    name of the table column containing the row names.

    If there is a header and the first row contains one fewer field than
    the number of columns, the first column in the input is used for the
    row names.  Otherwise if \code{row.names} is missing, the rows are
    numbered.

    Using \code{row.names = NULL} forces row numbering. Missing or
    \code{NULL} \code{row.names} generate row names that are considered
    to be \sQuote{automatic} (and not preserved by \code{\link{as.matrix}}).
  }
    \item{\code{col.names}}{a vector of optional names for the variables.
    The default is to use \code{"V"} followed by the column number.}
    \item{\code{as.is}}{controls conversion of character variables (insofar as
    they are not converted to logical, numeric or complex) to factors,
    if not otherwise specified by \code{colClasses}.
    Its value is either a vector of logicals (values are recycled if
    necessary), or a vector of numeric or character indices which
    specify which columns should not be converted to factors.

    Note: to suppress all conversions including those of numeric
    columns, set \code{colClasses = "character"}.

    Note that \code{as.is} is specified per column (not per
    variable) and so includes the column of row names (if any) and any
    columns to be skipped.
  }
    \item{\code{tryLogical}}{a \code{\link{logical}} determining if columns
    consisting entirely of \code{"F"}, \code{"T"}, \code{"FALSE"}, and
    \code{"TRUE"} should be converted to \code{\link{logical}}; passed to
    \code{\link[utils]{type.convert}}, true by default.}
    \item{\code{na.strings}}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after}
    white space is stripped from the input (if enabled), so \code{na.strings}
    values may need their own white space stripped in advance.}
    \item{\code{colClasses}}{character.  A vector of classes to be assumed for
    the columns.  If unnamed, recycled as necessary.  If named, names
    are matched with unspecified values being taken to be \code{NA}.

    Possible values are \code{NA} (the default, when
    \code{\link[utils]{type.convert}} is used), \code{"NULL"} (when the column
    is skipped), one of the atomic vector classes (logical, integer,
    numeric, complex, character, raw), or \code{"factor"}, \code{"Date"}
    or \code{"POSIXct"}.  Otherwise there needs to be an \code{as}
    method (from package \pkg{methods}) for conversion from
    \code{"character"} to the specified formal class.

    Note that \code{colClasses} is specified per column (not per
    variable) and so includes the column of row names (if any).
  }
    \item{\code{nrows}}{integer: the maximum number of rows to read in.  Negative
    and other invalid values are ignored.}
    \item{\code{skip}}{integer: the number of lines of the data file to skip before
    beginning to read data.}
    \item{\code{check.names}}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names.  If necessary they are adjusted
    (by \code{\link{make.names}}) so that they are, and also to ensure
    that there are no duplicates.}
    \item{\code{fill}}{logical. If \code{TRUE} then in case the rows have unequal
    length, blank fields are implicitly added.  See \sQuote{Details}.}
    \item{\code{strip.white}}{logical. Used only when \code{sep} has
    been specified, and allows the stripping of leading and trailing
    white space from unquoted \code{character} fields (\code{numeric} fields
    are always stripped).  See \code{\link{scan}} for further details
    (including the exact meaning of \sQuote{white space}),
    remembering that the columns may include the row names.}
    \item{\code{blank.lines.skip}}{logical: if \code{TRUE} blank lines in the
    input are ignored.}
    \item{\code{comment.char}}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}
    \item{\code{allowEscapes}}{logical.  Should C-style escapes such as
    \samp{\\n} be processed or read verbatim (the default)?   Note that if
    not within quotes these could be interpreted as a delimiter (but not
    as a comment character).  For more details see \code{\link{scan}}.}
    \item{\code{flush}}{logical: if \code{TRUE}, \code{scan} will flush to the
    end of the line after reading the last of the fields requested.
    This allows putting comments after the last field.}
    \item{\code{stringsAsFactors}}{logical: should character vectors be converted
    to factors?  Note that this is overridden by \code{as.is} and
    \code{colClasses}, both of which allow finer control.}
    \item{\code{fileEncoding}}{character string: if non-empty declares the
    encoding used on a file when given as a character string (not on an
    existing connection) so the character data can
    be re-encoded.  See the \sQuote{Encoding} section of the help for
    \code{\link{file}}, {\href{https://cloud.R-project.org/doc/manuals/R-data.html#Variations-on-read_002etable}{\dQuote{Variations on read.table}} in \cite{R Data Import/Export}}, and
    \sQuote{Note}.
  }
    \item{\code{encoding}}{encoding to be assumed for input strings.  It is
    used to mark character strings as known to be in
    Latin-1 or UTF-8 (see \code{\link{Encoding}}): it is not used to
    re-encode the input, but allows \R to handle encoded strings in
    their native encoding (if one of those two).  See \sQuote{Value}
    and \sQuote{Note}.
  }
    \item{\code{text}}{character string: if \code{file} is not supplied and this is,
    then data are read from the value of \code{text} via a text connection.
    Notice that a literal string can be used to include (small) data sets
    within R code.
  }
    \item{\code{skipNul}}{logical: should \abbr{NUL}s be skipped?}
  }}
}
\value{
In standard use \code{\link[=importADMS]{importADMS()}} returns a data frame for use in
openair. By comparison to the original file, the resulting data frame is
modified as follows:

Time and date information will combined in a single column "date",
formatted as a conventional timeseries (\verb{as.POSIX*}). If
\code{drop.input.dates} is enabled data series combined to generated the
new "date" data series will also be removed.

If \code{simplify.names} is enabled common chemical names may be
simplified, and some other parameters may be reset to openair standards
(e.g. "ws", "wd" and "temp") according to operations defined in
\code{simplifyNamesADMS}. A summary of simplification operations can be
obtained using, e.g., the call \code{importADMS(simplify.names)}.

If \code{drop.case} is enabled all upper case characters in names will be
converted to lower case.

If \code{keep.units} is enabled data units information may also be retained
as part of the data frame comment if available.

With \code{.mop} files, input and processed data series names may also been
modified on the basis of \code{drop.delim} and \code{add.prefixes} settings
}
\description{
Function(s) to import various ADMS file types into openair. Currently handles
".met", ".bgd", ".mop" and ".pst" file structures. Uses \code{\link[utils:read.csv]{utils::read.csv()}}
to read in data, format for R and openair and apply some file structure
testing.
}
\details{
The \code{importADMS} function were developed to help import various ADMS
file types into openair. In most cases the parent import function should work
in default configuration, e.g. \code{mydata <- importADMS()}. The function
currently recognises four file formats: \code{.bgd}, \code{.met}, \code{.mop}
and \code{.pst}.  Where other file extensions have been set but the file
structure is known, the import call can be forced by, e.g, \code{mydata <- importADMS(file.type="bgd")}. Other options can be adjusted to provide fine
control of the data structuring and renaming.
}
\note{
Times are assumed to be in GMT. Zero wind directions reset to 360 as
part of \code{.mop} file import.
}
\examples{
##########
# example 1
##########
# To be confirmed

# all current simplify.names operations
importADMS(simplify.names)

# to see what simplify.names does to adms data series name PHI
new.name <- importADMS(simplify.names, names = "PHI")
new.name
}
\seealso{
Other import functions:
\code{\link[=importAURN]{importAURN()}},
\code{\link[=importEurope]{importEurope()}},
\code{\link[=importImperial]{importImperial()}},
\code{\link[=importMeta]{importMeta()}},
\code{\link[=importTraj]{importTraj()}},
\code{\link[=importUKAQ]{importUKAQ()}}
}
\author{
Karl Ropkins, David Carslaw and Matthew Williams (CERC).
}
\concept{import functions}
