% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUKEAP.R
\name{importUKEAP}
\alias{importUKEAP}
\title{Import Data from the UK Eutrophying & Acidifying Network}
\usage{
importUKEAP(
  year = 2018,
  source = "namn",
  data_type = "annual",
  pollutant = "all",
  meta = TRUE,
  to_narrow = FALSE,
  progress = TRUE
)
}
\arguments{
\item{year}{Year(s) to import. To import a series of years use, e.g.,
\code{2000:2020}. To import several specific years use \code{year = c(2000, 2010, 2020)}.}

\item{source}{The network(s) from which to import site data. Providing a
single network will attempt to import all of the given sites from the
provided network. Available networks include:
\itemize{
\item \code{"namn"}, the UK National Ammonia Network.
}}

\item{data_type}{The type of data to be returned. There are two options:
\itemize{
\item \code{"multiday"}: Multi-day measurements, provided with separate "date" (start) and "date_end" (end) columns.
\item \code{"annual"}: Annual average data with data capture information for the whole network.
}}

\item{pollutant}{Pollutants to import. If omitted will import all pollutants
from a site. To import only NOx and NO2 for example use \code{pollutant =
  c("nox", "no2")}. Pollutant names can be upper or lower case.}

\item{meta}{Append the site type, latitude and longitude of each selected
\code{site}? Defaults to \code{FALSE}.}

\item{to_narrow}{Return the data in a "narrow"/"long"/"tidy" format? By
default the returned data is "wide" and has a column for each
pollutant/variable. When \code{to_narrow = TRUE} the data are returned with a
column identifying the pollutant name and a column containing the
corresponding concentration/statistic. Defaults to \code{FALSE}.}

\item{progress}{Show a progress bar when many sites/years are being imported?
Defaults to \code{TRUE}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Functions for importing air pollution data from the \href{https://uk-air.defra.gov.uk/networks/network-info?view=ukeap}{UK Eutrophying & Acidifying Network}. Files
are imported from a remote server operated by Ricardo that provides air
quality data files as R data objects. For an up to date list of available
sites that can be imported, see \code{\link[=importMeta]{importMeta()}}. Currently, only the NAMN
network is available, but it is intended that more data will be made
available in the future.
}
\examples{
\dontrun{
importNAMN(
  2020:2022,
  pollutant = c("delta_nh3_g", "a_nh4_p"),
  meta = T,
  to_narrow = F
)
}
}
\author{
Jack Davison, Trevor Davies, and David Carslaw
}
