% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tiktoken.R
\name{install_tiktoken}
\alias{install_tiktoken}
\title{Install the tiktoken Python package}
\usage{
install_tiktoken(method = "auto", conda = "auto")
}
\arguments{
\item{method}{The installation method to use. Can be one of "auto" (default), "conda", 
"virtualenv", "pip", or "windows".}

\item{conda}{The name or path of the Conda environment to use for the installation, 
or "auto" (default) to let reticulate automatically manage the environment.}
}
\description{
This function installs the tiktoken Python package using the specified installation method 
and Conda environment (if any).
}
