% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chaglog.R
\name{is_chatlog}
\alias{is_chatlog}
\title{Determine if an object is a chatlog}
\usage{
is_chatlog(object)
}
\arguments{
\item{object}{An R object to check}
}
\value{
TRUE if the object is of class "chatlog", FALSE otherwise
}
\description{
This function checks if an object is of class "chatlog".
}
\examples{
is_chatlog("Hello, World!")
# [1] FALSE

chat <- new("chatlog")
is_chatlog(chat)
# [1] TRUE
}
\author{
Ulrich Matter umatter@protonmail.com
}
