% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_python.R
\name{is_python}
\alias{is_python}
\title{Check if a Character String Contains Valid Python Code}
\usage{
is_python(code)
}
\arguments{
\item{code}{A character string containing code to be checked for Python code
validity.}
}
\value{
A logical value: TRUE if the input character string contains valid
Python code, and FALSE otherwise.
}
\description{
This function takes a character string as input and checks if it contains
valid Python code. It returns TRUE if the string contains valid Python code,
and FALSE otherwise.
}
\details{
This function presuposes that python is installed on the system.
}
\examples{
\dontrun{
# Check if the string contains valid Python code
is_python("print('Hello, World!')")

# Check if the string contains invalid Python code
is_python("prit('Hello, World!')")
}
}
