% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xml.R
\name{is_xml}
\alias{is_xml}
\title{Check if the provided string is in valid XML format.}
\usage{
is_xml(input_string)
}
\arguments{
\item{input_string}{A character string to be checked for XML format.}
}
\value{
A logical value. If the input string is in valid XML format, returns TRUE, otherwise returns FALSE.
}
\description{
Check if the provided string is in valid XML format.
}
\examples{
is_xml('<?xml version="1.0"?><root><element>value</element></root>')
# TRUE

is_xml('<root><element>value</element></root>')
# FALSE

is_xml('')
# FALSE
}
\author{
Ulrich Matter umatter@protonmail.com
}
