% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{java_to_r_prompt}
\alias{java_to_r_prompt}
\title{Input: message content template for java_to_r()}
\format{
## `java_to_r_prompt`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
java_to_r_prompt
}
\description{
This message content serves as user input for the java_to_r() function.
}
\keyword{datasets}
