% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_csv.R
\name{json_to_csv}
\alias{json_to_csv}
\title{Convert JSON input to CSV}
\usage{
json_to_csv(file)
}
\arguments{
\item{file}{A character string representing the path to the file with JSON input}
}
\value{
If the input is a character string, the function returns the output 
as a character string and prints it to the console. Otherwise, the function returns 
the output as a data frame and generates a CSV file with the same name as the input file 
but with a .csv extension.
}
\description{
This function takes a file with JSON input and processes it to generate 
corresponding output in CSV format. If the input is a character string, 
the output will be printed to the console. Otherwise, a CSV file will be 
generated with the same name as the input file but with a .csv extension.
}
\examples{

json_string <- '{"employees":[
{ "firstName":"John", "lastName":"Doe" },
{ "firstName":"Anna", "lastName":"Smith" },
{ "firstName":"Peter", "lastName":"Jones" }
]}'
json_to_csv(json_string)

}
