% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_openair.R
\name{label_openair}
\alias{label_openair}
\title{Automatic text formatting for openair}
\usage{
label_openair(x, style = "marquee", auto_text = TRUE)
}
\arguments{
\item{x}{A character vector.}

\item{style}{The type of formatting to use. Currently limited to \code{"marquee"}
for use with, e.g., \code{\link[marquee:marquee_grob]{marquee::marquee_grob()}}.}

\item{auto_text}{Perform the formatting of the data? Used internally by
\code{openair} functions to turn on and off label formatting.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This workhorse function automatically applies routing text formatting to
common words in air quality and meteorological science (for example,
subscripting the 'x' in NOx). Unlike \code{\link[=quickText]{quickText()}}, it is fully vectorised
and returns a character string.
}
\author{
Jack Davison
}
