% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{list_assistants}
\alias{list_assistants}
\title{List assistants (convenience function)}
\usage{
list_assistants(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
List of assistants
}
\description{
List assistants (convenience function)
}
