% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{list_batches}
\alias{list_batches}
\title{List batches (convenience function)}
\usage{
list_batches(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
List of batches
}
\description{
List batches (convenience function)
}
