% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{List files (convenience function)}
\usage{
list_files(purpose = NULL, ...)
}
\arguments{
\item{purpose}{Filter by purpose}

\item{...}{Additional parameters}
}
\value{
List of files
}
\description{
List files (convenience function)
}
