% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tuning.R
\name{list_fine_tuning_checkpoints}
\alias{list_fine_tuning_checkpoints}
\title{List fine-tuning checkpoints (convenience function)}
\usage{
list_fine_tuning_checkpoints(fine_tuning_job_id, ...)
}
\arguments{
\item{fine_tuning_job_id}{Job ID}

\item{...}{Additional parameters}
}
\value{
List of checkpoints
}
\description{
List fine-tuning checkpoints (convenience function)
}
