% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{list_messages}
\alias{list_messages}
\title{List messages (convenience function)}
\usage{
list_messages(thread_id, ...)
}
\arguments{
\item{thread_id}{Thread ID}

\item{...}{Additional parameters}
}
\value{
List of messages
}
\description{
List messages (convenience function)
}
