% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{list_vector_stores}
\alias{list_vector_stores}
\title{List vector stores (convenience function)}
\usage{
list_vector_stores(...)
}
\arguments{
\item{...}{Additional parameters}
}
\value{
List of vector stores
}
\description{
List vector stores (convenience function)
}
