% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages_content.R
\name{messages_content}
\alias{messages_content}
\title{Extract Messages Content from OpenAI API Response}
\usage{
messages_content(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a character string representing the choices/messages from the response
}
\description{
This function extracts the messages content from the HTTP response of an API call to the
OpenAI chat completions endpoint.
}
