% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Extract Model from OpenAI API response}
\usage{
model(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a character string representing the model from the response
}
\description{
This function extracts the model from the parsed HTTP response of an API call to the
OpenAI chat completions endpoint.
}
\author{
Ulrich Matter umatter@protonmail.com
}
