% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderations.R
\name{ModerationsClient}
\alias{ModerationsClient}
\title{Moderations Client}
\description{
Client for OpenAI Moderations API.
Classifies text as potentially harmful.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModerationsClient-new}{\code{ModerationsClient$new()}}
\item \href{#method-ModerationsClient-create}{\code{ModerationsClient$create()}}
\item \href{#method-ModerationsClient-clone}{\code{ModerationsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$create(input, model = "omni-moderation-latest")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModerationsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModerationsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModerationsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
