% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nested_to_pipe_prompt}
\alias{nested_to_pipe_prompt}
\title{Input: message content template for nested_to_pipe()}
\format{
## `nested_to_pipe_prompt`
A data frame with 4 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
nested_to_pipe_prompt
}
\description{
This message content serves as user input for the nested_to_pipe() function.
}
\keyword{datasets}
