% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_tokens.R
\name{num_tokens}
\alias{num_tokens}
\title{Get number of tokens in a string using OpenAI's tiktoken library}
\usage{
num_tokens(text, encoding = "cl100k_base")
}
\arguments{
\item{text}{a character string to count the number of tokens from}

\item{encoding}{a character string that specifies how text is converted into
tokens. The default is "cl100k_base" (for ChatGPT models; see
https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
for details)}
}
\value{
an integer indicating the number of tokens in the input string
}
\description{
This function uses the num_tokens_from_string function provided in OpenAI's
tiktoken python library to get the number of tokens in a string.
}
\references{
https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
}
