% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object}
\alias{object}
\title{Extract object information from OpenAI API response}
\usage{
object(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a data frame with the usage statistics of the API call (how many tokens used)
}
\description{
This function extracts the object data from the parsed HTTP response of an API call to the
OpenAI chat completions endpoint (provides information about the endpoint).
}
\author{
Ulrich Matter umatter@protonmail.com
}
