% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openColours.R
\docType{data}
\name{openColours}
\alias{openColours}
\alias{openSchemes}
\title{Pre-defined openair colours and definition of user-defined colours}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 111 rows and 5 columns.
}
\usage{
openColours(
  scheme = "default",
  n = 100,
  direction = 1L,
  too_few = c("interpolate", "repeat", "error")
)

openSchemes
}
\arguments{
\item{scheme}{Any one of the pre-defined \code{openair} schemes (e.g.,
\code{"increment"}) or a user-defined palette (e.g., \code{c("red", "orange", "gold")}). See \code{\link[=openColours]{openColours()}} for a full list of available schemes.
Starting a pre-defined scheme with \code{"-"} (e.g., \code{"-increment"}) will
reverse \code{direction}; this is mainly useful for use within other \code{openair}
functions (e.g., \code{polarPlot(mydata, cols = "-viridis")}).}

\item{n}{number of colours required. By default, this will return \code{100}
values for sequential palettes or the maximum number of pre-defined colours
in categorical palettes.}

\item{direction}{Sets the order of colours in the scale. If \code{1}, the default,
outputs the default order of the \code{scheme}. If \code{-1}, the order of colours is
reversed.}

\item{too_few}{If a categorical palette has too few colours for the given
\code{n}, how should \code{\link[=openColours]{openColours()}} deal with this? One of:
\itemize{
\item \code{"interpolate"} (the default), which causes new colours to be
interpolated based on the existing palette. Note that this may create ugly
colours or compromise colourblind-friendliness.
\item \code{"repeat"}, which cycles through the sequential scheme meaning colours
are repeated.
\item \code{"error"}, which causes an error if \code{n} exceeds available colours in \code{scheme}.
}}
}
\value{
A character vector of hex codes
}
\description{
\code{\link[=openColours]{openColours()}} is a convenience function for accessing pre-defined colours
or interpolating user-defined colours. Many useful and accessible themes are
defined within \code{openair} for convenience, all of which are categorised and
referenced within \link{openSchemes}. Further, \code{\link[=openColours]{openColours()}} is often used
within \emph{other} openair functions; for example \code{polarPlot(mydata, cols = "viridis")} will use \code{openColours("viridis")} to access a colour palette.
}
\section{Details}{


Because of the way many of the schemes have been developed they only exist
over certain number of colour gradations (typically 3--10). If the required
number of colours is outside of this range, \code{openair} will interpolate the
colours with a warning.

Each of the pre-defined schemes have merits and their use will depend on a
particular situation. For showing incrementing concentrations, e.g., high
concentrations emphasised, then \code{"default"}, \code{"turbo"}, \code{"viridis"}, and
\code{"increment"} are very useful.

To colour-code categorical-type problems, e.g., colours for different
pollutants, \code{"hue"} and \code{"brewer1"} are useful.

When publishing in black and white, \code{"greyscale"} is often convenient. With
most \code{openair} functions, as well as generating a greyscale colour
gradient, it also resets strip background and other coloured text and lines
to greyscale values.

Failing that, the user can define their own schemes based on R colour
names. To see the full list of names, see \code{\link[=colors]{colors()}}.
}

\section{Schemes}{


The following is an abridged list of schemes made available by
\code{\link[=openColours]{openColours()}}. A complete list is outlined in \link{openSchemes}.

\strong{Sequential Colours:}
\itemize{
\item \code{"default"}, \code{"increment"}, \code{"brewer1"}, \code{"heat"}, \code{"jet"}, \code{"turbo"},
\code{"hue"}, \code{"greyscale"}.
\item Simplified versions of the \code{viridis} colours: \code{"viridis"}, \code{"plasma"},
\code{"magma"}, \code{"inferno"}, and \code{"cividis"}.
\item Simplified versions of the \code{RColorBrewer} sequential palettes: \code{"Blues"}, \code{"BuGn"},
\code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"}, \code{"OrRd"}, \code{"PuBu"},
\code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"}, \code{"RdPu"}, \code{"Reds"}, \code{"YlGn"},
\code{"YlGnBu"}, \code{"YlOrBr"}, \code{"YlOrRd"}.
\item Simplified versions of Fabio Crameri's sequential color schemes (e.g., \code{"batlow"}).
}

\strong{Diverging Palettes:}
\itemize{
\item Simplified versions of the \code{RColorBrewer} diverging palettes: \code{"BrBG"},
\code{"PiYG"}, \code{"PRGn"}, \code{"PuOr"}, \code{"RdBu"}, \code{"RdGy"}, \code{"RdYlBu"}, \code{"RdYlGn"},
\code{"Spectral"}.
\item Simplified versions of Fabio Crameri's diverging color schemes (e.g., \code{"berlin"}).
}

\strong{Qualitative Palettes:}
\itemize{
\item Simplified versions of the \code{RColorBrewer} qualitative palettes:
\code{"Accent"}, \code{"Dark2"}, \code{"Paired"}, \code{"Pastel1"}, \code{"Pastel2"}, \code{"Set1"},
\code{"Set2"}, \code{"Set3"}.
\item \code{"okabeito"} (or \code{"cbPalette"}), a colour-blind safe palette based on
the work of Masataka Okabe and Kei Ito.
\item Paul Tol's palettes (e.g., \code{"tol.bright"}, \code{"tol.muted"}, \code{"tol.light"},
etc.), many of which are colour blind safe.
\item \code{"tableau"} and \code{"observable"}, aliases for the
"Tableau10" and "Observable10" colour palettes. These could be useful for
consistency between openair plots and with figures made in these software
programs.
}

\strong{UK Government Palettes:}
\itemize{
\item \code{"daqi"} and \code{"daqi.bands"}, the colours associated with the UK daily air
quality index; \code{"daqi"} (a palette of 10 colours, corresponding to each
index value) or \code{"daqi.bands"} (4 colours, corresponding to each band -
Low, Moderate, High, and Very High). These colours may be useful in figures
like \code{\link[=calendarPlot]{calendarPlot()}}.
\item \code{"gaf.cat"}, \code{"gaf.focus"} and \code{"gaf.seq"}, colours recommended by the UK Government Analysis function.
\code{"gaf.cat"} will return the 'categorical' palette (max 6 colours),
\code{"gaf.focus"} the 'focus' palette (max 2 colours), and \code{"gaf.seq"} the
'sequential' palette.
}
}

\section{The \code{openSchemes} dataset}{


To help understand the available palettes in \code{\link[=openColours]{openColours()}}, the
\link{openSchemes} dataset is provided. This is a \code{data.frame} listing all
schemes, categorising them into diverging, sequential, or qualitative. This
dataset can be filtered to find useful scales for your specific
visualisation.

\describe{
\item{category}{The category - one of \code{"div"} (diverging), \code{"qual"} (qualitative), or \code{"seq"} (sequential).}
\item{scheme}{The scheme name - to be provided to \code{\link[=openColours]{openColours()}}}
\item{max_n}{The maximum number of pre-defined colours in a qualitative palette. Setting \code{n} greater than this value is not recommended.}
\item{origin}{The original authors of the palette.}
\item{reference}{A URL reference for further reading about the colour schemes.}
}
}

\examples{

# to return 5 colours from the "jet" scheme:
cols <- openColours("jet", 5)
cols

# to interpolate between named colours e.g. 10 colours from yellow to
#  green to red:
cols <- openColours(c("yellow", "green", "red"), 10)
cols

openSchemes
}
\author{
David Carslaw

Jack Davison
}
\keyword{datasets}
