% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{OpenAI}
\alias{OpenAI}
\title{OpenAI Client Class}
\description{
Main client class for interacting with OpenAI API.
Compatible with Python OpenAI SDK interface.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpenAI-new}{\code{OpenAI$new()}}
\item \href{#method-OpenAI-build_headers}{\code{OpenAI$build_headers()}}
\item \href{#method-OpenAI-request}{\code{OpenAI$request()}}
\item \href{#method-OpenAI-request_multipart}{\code{OpenAI$request_multipart()}}
\item \href{#method-OpenAI-request_raw}{\code{OpenAI$request_raw()}}
\item \href{#method-OpenAI-clone}{\code{OpenAI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$new(
  api_key = NULL,
  base_url = NULL,
  organization = NULL,
  project = NULL,
  timeout = 600,
  max_retries = 2
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-build_headers"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-build_headers}{}}}
\subsection{Method \code{build_headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$build_headers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-request"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-request}{}}}
\subsection{Method \code{request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$request(
  method,
  path,
  body = NULL,
  query = NULL,
  stream = FALSE,
  callback = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-request_multipart"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-request_multipart}{}}}
\subsection{Method \code{request_multipart()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$request_multipart(method, path, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-request_raw"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-request_raw}{}}}
\subsection{Method \code{request_raw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$request_raw(method, path, body = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
