% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai_api_key.R
\name{openai_api_key}
\alias{openai_api_key}
\title{Set OpenAI API key as an environment variable}
\usage{
openai_api_key(api_key)
}
\arguments{
\item{api_key}{character string containing the OpenAI API key.}
}
\value{
Nothing is returned; the function is called for its side effects.
}
\description{
This function sets the OpenAI API key as an environment variable in the current R session.
}
\examples{
openai_api_key("your_api_key_here")

}
