% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openColours.R
\name{openSchemes}
\alias{openSchemes}
\title{List available colour schemes in \code{openair}}
\usage{
openSchemes(palette_type = c("seq", "div", "qual"), n = NULL)
}
\arguments{
\item{palette_type}{A character vector specifying which types of palettes to
include in the output. Options are \code{"seq"} for sequential, \code{"div"} for
diverging, and \code{"qual"} for qualitative. Multiple options can be selected
(e.g., \code{c("seq", "div")}). All three are returned by default.}

\item{n}{An optional whole number to filter the schemes by their maximum
number of colours. Only schemes that can provide at least \code{n} colours will
be included in the output. This only applies to qualitative schemes, as
sequential and diverging schemes can be interpolated to any number of
colours.}
}
\description{
This function returns a table of the available colour schemes in \code{openair},
along with their type (sequential, diverging, or qualitative) and the maximum
number of colours available for that scheme (where applicable). This can be
useful for users to explore the available schemes and to check which schemes
can provide a certain number of colours.
}
\author{
Jack Davison
}
