% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Parse a response from the API}
\usage{
parse_response(input_string)
}
\arguments{
\item{input_string}{The response from the API}
}
\value{
A list of blocks
}
\description{
Parse a response from the API
}
\examples{
parse_response("Hello world!")
# [[1]]
# $type
# [1] "text"
#
# $content
# [1] "Hello world!"
#
parse_response("```python\nprint('Hello world!')\n```")
# [[1]]
# $type
# [1] "code"
#
# $content
# [1] "print('Hello world!')"
#
# $language
# [1] "python"
#
parse_response("Hello world!\n\n```python\nprint('Hello world!')\n```")
# [[1]]
# $type
# [1] "text"
#
# $content
# [1] "Hello world!"
#
# [[2]]
# $type
# [1] "code"
#
# $content
# [1] "print('Hello world!')"
#
# $language
# [1] "python"
#
}
\author{
Jonathan Chassot
}
