% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentileRose.R
\name{percentileRose}
\alias{percentileRose}
\title{Function to plot percentiles by wind direction}
\usage{
percentileRose(
  mydata,
  pollutant = "nox",
  ws = "ws",
  wd = "wd",
  type = "default",
  percentile = c(25, 50, 75, 90, 95),
  smooth = FALSE,
  method = "default",
  cols = "default",
  angle = 10,
  mean = TRUE,
  mean.lty = 1,
  mean.lwd = 1,
  mean.col = "grey",
  fill = TRUE,
  intervals = NULL,
  angle.scale = 45,
  offset = 0,
  auto.text = TRUE,
  key.title = NULL,
  key.position = "bottom",
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame minimally containing a decimal wind direction and
a numeric field to plot.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}. More than one
pollutant can be supplied e.g. \code{pollutant = c("no2", "o3")} provided there
is only one \code{type}.}

\item{ws}{The name of the column in \code{mydata} representing the wind speed.
Defaults to \code{"ws"}.}

\item{wd}{The name of the column in \code{mydata} representing the decimal wind
direction, 0 to 360 where 0/360 are North and 180 is South. Defaults to
\code{"wd"}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{percentile}{The percentile value(s) to plot. Must be between 0--100. If
\code{percentile = NA} then only a mean line will be shown.}

\item{smooth}{Should the wind direction data be smoothed using a cyclic
spline?}

\item{method}{When \code{method = "default"} the supplied percentiles by wind
direction are calculated. When \code{method = "cpf"} the conditional probability
function (CPF) is plotted and a single (usually high) percentile level is
supplied. The CPF is defined as CPF = my/ny, where my is the number of
samples in the wind sector y with mixing ratios greater than the \emph{overall}
percentile concentration, and ny is the total number of samples in the same
wind sector (see Ashbaugh et al., 1985).}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{angle}{Default angle of \dQuote{spokes} is when \code{smooth = FALSE}.}

\item{mean}{Show the mean by wind direction as a line?}

\item{mean.lty}{Line type for mean line.}

\item{mean.lwd}{Line width for mean line.}

\item{mean.col}{Line colour for mean line.}

\item{fill}{Should the percentile intervals be filled (default) or should
lines be drawn (\code{fill = FALSE}).}

\item{intervals}{User-supplied intervals for the scale e.g. \code{intervals = c(0, 10, 30, 50)}.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{offset}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label, passed to
\code{\link[ggplot2:labs]{ggplot2::labs()}} via \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{scales} can be \code{"fixed"}, \code{"free_x"}, \code{"free_y"} or \code{"free"} to control
whether axes are shared across facets when using \code{type}. Also supported are
the legacy \code{x.relation} and \code{y.relation}, which can be either \code{"same"} or
\code{"free"} and get remapped to \code{scales} automatically.
\item Similarly, \code{space}, \code{axes}, \code{axis.labels}, \code{switch} and \code{strip.position}
can be used to customise the appearance of faceted plots. See
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} for the arguments these
take.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
\code{\link[=percentileRose]{percentileRose()}} plots percentiles by wind direction with flexible
conditioning. The plot can display multiple percentile lines or filled areas.
}
\details{
\code{\link[=percentileRose]{percentileRose()}} calculates percentile levels of a pollutant and plots them
by wind direction. One or more percentile levels can be calculated and these
are displayed as either filled areas or as lines.

The wind directions are rounded to the nearest 10 degrees, consistent with
surface data from the UK Met Office before a smooth is fitted. The levels by
wind direction are optionally calculated using a cyclic smooth cubic spline
using the option \code{smooth}. If \code{smooth = FALSE} then the data are shown in 10
degree sectors.

The \code{percentileRose} function compliments other similar functions including
\code{\link[=windRose]{windRose()}}, \code{\link[=pollutionRose]{pollutionRose()}}, \code{\link[=polarFreq]{polarFreq()}} or \code{\link[=polarPlot]{polarPlot()}}. It is most
useful for showing the distribution of concentrations by wind direction and
often can reveal different sources e.g. those that only affect high
percentile concentrations such as a chimney stack.

Similar to other functions, flexible conditioning is available through the
\code{type} option. It is easy for example to consider multiple percentile values
for a pollutant by season, year and so on. See examples below.

\code{percentileRose} also offers great flexibility with the scale used and the
user has fine control over both the range, interval and colour.
}
\examples{
# basic percentile plot
percentileRose(mydata, pollutant = "o3")

# 50/95th percentiles of ozone, with different colours
percentileRose(mydata, pollutant = "o3", percentile = c(50, 95), col = "brewer1")

\dontrun{
# percentiles of ozone by year, with different colours
percentileRose(
  mydata,
  type = "year",
  pollutant = "o3",
  col = "brewer1",
  ncol = 4,
  nrow = 2
)

# percentile concentrations by season and day/nighttime..
percentileRose(
  mydata,
  type = c("daylight", "season"),
  pollutant = "o3",
  col = "brewer1"
)
}
}
\references{
Ashbaugh, L.L., Malm, W.C., Sadeh, W.Z., 1985. A residence time
probability analysis of sulfur concentrations at ground canyon national
park. Atmospheric Environment 19 (8), 1263-1270.
}
\seealso{
Other polar directional analysis functions:
\code{\link[=polarAnnulus]{polarAnnulus()}},
\code{\link[=polarCluster]{polarCluster()}},
\code{\link[=polarDiff]{polarDiff()}},
\code{\link[=polarFreq]{polarFreq()}},
\code{\link[=polarPlot]{polarPlot()}},
\code{\link[=pollutionRose]{pollutionRose()}},
\code{\link[=windRose]{windRose()}}
}
\author{
David Carslaw

Jack Davison
}
\concept{polar directional analysis functions}
