% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Plot an air quality timeseries chart in a conventional calendar format}
\usage{
plot_calendar(
  data,
  pollutant = "nox",
  year = NULL,
  month = NULL,
  annotate = "date",
  digits = 0,
  statistic = "mean",
  data.thresh = 0,
  percentile = NA,
  cols = "heat",
  limits = c(0, 100),
  labels = NA,
  breaks = NA,
  w.shift = 0,
  w.abbr.len = 1,
  remove.empty = TRUE,
  show.year = TRUE,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \verb{pollutant = "nox". }}

\item{year}{Year to plot e.g. \code{year = 2003}. If not supplied all data
potentially spanning several years will be plotted.}

\item{month}{If only certain month are required. By default the function will
plot an entire year even if months are missing. To only plot certain months
use the \code{month} option where month is a numeric 1:12 e.g. \code{month = c(1, 12)} to only plot January and December.}

\item{annotate}{This option controls what appears on each day of the
calendar. Can be:
\itemize{
\item \code{"date"} --- shows day of the month
\item \code{"value"} --- shows the daily mean value
}}

\item{digits}{The number of digits used to display concentration values when
\code{annotate = "value"}.}

\item{statistic}{Statistic passed to \code{\link[=timeAverage]{timeAverage()}}. Note that if \code{statistic \%in\% c("max", "min")} and \code{annotate} is "ws" or "wd", the hour
corresponding to the maximum/minimum concentration of \code{polluant} is used to
provide the associated \code{ws} or \code{wd} and not the maximum/minimum daily \code{ws}
or \code{wd}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{percentile}{The percentile level in percent used when \code{statistic = "percentile"} and when aggregating the data with \code{avg.time}. More than one
percentile level is allowed for \code{type = "default"} e.g. \code{percentile = c(50, 95)}. Not used if \code{avg.time = "default"}.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{limits}{Use this option to manually set the colour scale limits. This
is useful in the case when there is a need for two or more plots and a
consistent scale is needed on each. Set the limits to cover the maximum
range of the data for all plots of interest. For example, if one plot had
data covering 0--60 and another 0--100, then set \code{limits = c(0, 100)}. Note
that data will be ignored if outside the limits range.}

\item{labels}{If a categorical scale is defined using \code{breaks}, then \code{labels}
can be used to override the default category labels, e.g., \code{labels = c("good", "bad", "very bad")}. Note there is one less label than break.}

\item{breaks}{If a categorical scale is required then these breaks will be
used. For example, \code{breaks = c(0, 50, 100, 1000)}. In this case "good"
corresponds to values between 0 and 50 and so on. Users should set the
maximum value of \code{breaks} to exceed the maximum data value to ensure it is
within the maximum final range e.g. 100--1000 in this case.}

\item{w.shift}{Controls the order of the days of the week. By default the
plot shows Saturday first (\code{w.shift = 0}). To change this so that it starts
on a Monday for example, set \code{w.shift = 2}, and so on.}

\item{w.abbr.len}{The default (\code{1}) abbreviates the days of the week to a
single letter (e.g., in English, S/S/M/T/W/T/F). \code{w.abbr.len} defines the
number of letters to abbreviate until. For example, \code{w.abbr.len = 3} will
abbreviate "Monday" to "Mon".}

\item{remove.empty}{Should months with no data present be removed? Default is
\code{TRUE}.}

\item{show.year}{If only a single year is being plotted, should the calendar
labels include the year label? \code{TRUE} creates labels like "January-2000",
\code{FALSE} labels just as "January". If multiple years of data are detected,
this option is forced to be \code{TRUE}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Not currently used.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function will plot data by month laid out in a conventional calendar
format. The main purpose is to help rapidly visualise potentially complex
data in a familiar way. Daily statistics are calculated using
\code{\link[=timeAverage]{timeAverage()}}, which by default will calculate the daily mean
concentration.
}
\details{
It is possible to plot categorical scales. This is useful where, for example,
an air quality index defines concentrations as bands, e.g., "good", "poor".
In these cases users must supply \code{labels} and corresponding \code{breaks}.

Note that is is possible to pre-calculate concentrations in some way before
passing the data to \code{\link[=calendarPlot]{calendarPlot()}}. For example \code{\link[=rollingMean]{rollingMean()}} could be
used to calculate rolling 8-hour mean concentrations. The data can then be
passed to \code{\link[=calendarPlot]{calendarPlot()}} and \code{statistic = "max"} chosen, which will plot
maximum daily 8-hour mean concentrations.
}
\examples{
# basic plot
plot_calendar(mydata, pollutant = "o3", year = 2003)

\dontrun{
# categorical scale example
plot_calendar(
  mydata,
  pollutant = "no2",
  year = 2003,
  breaks = c(0, 50, 100, 150, 1000),
  labels = c("Very low", "Low", "High", "Very High"),
  cols = c("lightblue", "green", "yellow", "red"),
  statistic = "max"
)

# UK daily air quality index
pm10_breaks <- c(0, 17, 34, 50, 59, 67, 75, 84, 92, 100, 1000)
plot_calendar(
  mydata,
  "pm10",
  year = 1999,
  breaks = pm10_breaks,
  labels = c(1:10),
  cols = "daqi",
  statistic = "mean"
)$plot + ggplot2::labs(fill = "DAQI")
}
}
\seealso{
Other ggplot2 time series and trend functions: 
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_smooth}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{ggplot2 time series and trend functions}
