% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_TSV.R
\name{plot_cluster_TSV}
\alias{plot_cluster_TSV}
\title{Plot calculated TSV and change-in-TSV-% values for a given cluster results grob}
\usage{
plot_cluster_TSV(TSV_res, threshold = 5, max_clusters = 30, max_change = 10)
}
\arguments{
\item{TSV_res}{Data.frame, an output from trajCluster_TSV.}

\item{threshold}{numeric, a percentage change threshold to highlight in the plot.}

\item{max_clusters}{maximum number of clusters to show on the graph. Will shrink if a smaller-than-specified TSV_res frame is supplied.}

\item{max_change}{y-axis on the percent change TSV plot.}
}
\description{
Takes results from \code{\link{trajCluster_TSV}} and plots the change-in-TSV percentage
between different n clusters. Options to set a threshold in % variance change to visualise
potentially desirable n cluster specifications
}
