% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditional_quantile.R
\name{plot_conditional_quantile}
\alias{plot_conditional_quantile}
\title{Conditional quantile estimates for model evaluation}
\usage{
plot_conditional_quantile(
  data,
  obs,
  mod,
  type = NULL,
  bins = 30,
  min_bin = 20,
  scale_y = openair::scale_opts(),
  scale_x = openair::scale_opts(),
  cols = "viridis",
  col_ideal = "grey50",
  col_histogram = "grey70",
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{obs, mod}{The names of the columns to summarise, one for observations
(\code{obs}) and one for modelled values (\code{mod}); these should be columns in
\code{data}.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{bins}{Number of bins to be used in calculating the different quantile
levels.}

\item{min_bin}{The minimum number of measurements to use to calculate a
binned statistic. If the number is below \code{min_bin}, the value will be
dropped from the returned plot and/or data.}

\item{scale_y}{Options for the y-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{scale_x}{Options for the x-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}. The first colour
will be used for the 10/90 percentile band, the second for the 25/75 band,
and the third for the median line.}

\item{col_ideal, col_histogram}{The colours to use for the 'ideal' case (the
1:1 diagonal line and observed value histogram) and the modelled value
histogram. Passed to \code{\link[=openColours]{openColours()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Conditional quantiles are a very useful way of considering model
performance against observations for continuous measurements (Wilks, 2005).
The conditional quantile plot splits the data into evenly spaced bins. For
each predicted value bin e.g. from 0 to 10~ppb the \emph{corresponding} values
of the observations are identified and the median, 25/75th and 10/90
percentile (quantile) calculated for that bin. The data are plotted to show
how these values vary across all bins. For a time series of observations
and predictions that agree precisely the median value of the predictions
will equal that for the observations for each bin.
}
\section{Controlling scales}{


In \code{openair}, a 'scale' is anything expressed in a numeric or date range.
This includes axes and colour bars. Most \code{openair} plotting functions
possess arguments in the pattern \verb{scale_*} to control these - e.g.,
\code{scale_x}, \code{scale_y}, and \code{scale_col}.

If these arguments are passed a single number, this will set the maximum of
the scale to be that number. If two numbers are provided, these will be
used as the scale range.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# x axis goes up to 100
plot_xy_scatter(mydata, "nox", "no2", scale_x = 100)

# x axis is between 50 and 150
plot_xy_scatter(mydata, "nox", "no2", scale_x = c(50, 150))
}\if{html}{\out{</div>}}

These \verb{scale_*} options can also be provided with the \code{\link[=scale_opts]{scale_opts()}}
function, which allows for further customisation, such as:
\itemize{
\item Where to add major axis/legend breaks, and how to label them.
\item Whether to transform the scale (e.g., a log or sqrt transform).
\item Where to place the axis, and whether to draw a second one (axes only).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_xy_scatter(
  mydata,
  "nox",
  "no2",
  scale_x = scale_opts(
    transform = "pseudo_log",
    breaks = c(1, 3, 10, 30, 100, 300, 1000, 3000),
    sec.axis = ggplot2::sec_axis(~.)
  )
)
}\if{html}{\out{</div>}}

Take care with date scales (e.g., \code{scale_x} in \code{\link[=plot_trend_bars]{plot_trend_bars()}});
\code{limits} should be provided as POSIXct objects. This is likely to be
refined in future. Note that, for date scales, \code{date_breaks} and
\code{date_labels} can be used to more easily refine the appearance of the
scale.
}

\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# create a small dataset
mod_data <- dplyr::select(mydata, date, nox)

# lets create some fake modelled values
mod_data <-
  dplyr::mutate(
    mod_data,
    mod = jitter(nox, factor = 1000)
  )

# plot this on a taylor diagram
plot_conditional_quantile(mod_data, "nox", "mod")
}
}
\references{
Murphy, A. H., B.G. Brown and Y. Chen. (1989) Diagnostic Verification of
Temperature Forecasts, Weather and Forecasting, Volume: 4, Issue: 4, Pages:
485-501.

Wilks, D. S., 2005. Statistical Methods in the Atmospheric Sciences, Volume
91, Second Edition (International Geophysics), 2nd Edition. Academic Press.
}
\seealso{
the legacy \code{\link[=conditionalQuantile]{conditionalQuantile()}} function

Other ggplot2 model evaluation functions: 
\code{\link{plot_taylor_diagram}()}
}
\author{
Jack Davison
}
\concept{ggplot2 model evaluation functions}
