% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot a correlation matrix with conditioning}
\usage{
plot_correlation(
  data,
  pollutant = NULL,
  type = "default",
  cluster = TRUE,
  method = "pearson",
  use = "pairwise.complete.obs",
  annotate = c("cor", "signif", "stars", "none"),
  triangle = c("all", "lower", "upper"),
  diagonal = TRUE,
  discretise = NULL,
  cols = "PRGn",
  cols_annotation = "black",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{If \code{NULL}, all numeric columns in \code{data} will be correlated
with one another. \code{pollutant}, as in other \code{openair} functions, may also be
a vector of column names to use instead.}

\item{cluster}{Should the data be ordered according to cluster analysis? If
\code{TRUE} hierarchical clustering is applied to the correlation matrices using
\code{hclust} to group similar variables together.}

\item{method}{The correlation method to use. Can be \code{"pearson"}, \code{"spearman"}
or \code{"kendall"}.}

\item{use}{How to handle missing values in the \code{cor} function. The default is
\code{"pairwise.complete.obs"}. Care should be taken with the choice of how to
handle missing data when considering pair-wise correlations.}

\item{annotate}{What to annotate each correlation tile with. One of:
\itemize{
\item \code{"cor"}, the correlation coefficient to 2 decimal places.
\item \code{"signif"}, an X marker if the correlation is significant.
\item \code{"stars"}, standard significance stars.
\item \code{"none"}, no annotation.
}}

\item{triangle}{One of \code{"all"} (to show the complete matrix), \code{"lower"} (to
show the bottom triangle) or \code{"upper"} (to show the upper triangle).}

\item{diagonal}{When \code{TRUE}, the 'diagonal' of the plot (where the
correlation is always \code{1}) will be shown. \code{FALSE} removes this.}

\item{discretise}{Options to discretise the continuous scale into discrete
bins. Providing an integer will split the scale into that number of
quantiles. Alternatively, \code{\link[=discretise_breaks]{discretise_breaks()}} allows for extensive
customisation (e.g., defining specific bin widths, breaks, labels, and so
on).}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}. Note that
\code{\link[=plot_correlation]{plot_correlation()}} fixes the colour scale between \code{-1} and \code{1}, so a
'diverging' palette is recommended.}

\item{cols_annotation}{The colour to use to colour text annotations. If two
colours are defined, the first will be used for positive correlations and
the second for negative correlations.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Not currently used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function calculates and visualises correlation matrices between air
quality pollutants (or any other numeric variables). The primary purpose is
as a tool for exploratory data analysis. Hierarchical clustering is used to
group variables that are most similar to one another, if \code{cluster = TRUE}.
It is also possible to use the \code{openair} type option to condition the data
in many flexible ways.
}
\examples{
# basic corrgram plot
plot_correlation(mydata)

# plot by season ... and so on
plot_correlation(mydata, type = "season")

# recover dendrogram when cluster = TRUE and plot it
res <- plot_correlation(mydata, plot = FALSE)
plot(res$hclust)

\dontrun{
# more interesting are hydrocarbon measurements
hc <- importAURN(site = "my1", year = 2005, hc = TRUE)

# now it is possible to see the hydrocarbons that behave most
# similarly to one another
plot_correlation(hc, annotate = "none", cols = "Spectral")
}
}
\seealso{
The legacy \code{\link[=corPlot]{corPlot()}} function.
}
\author{
Jack Davison
}
