% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polar_percentile.R
\name{plot_polar_percentile}
\alias{plot_polar_percentile}
\title{Function to plot percentiles by wind direction}
\usage{
plot_polar_percentile(
  data,
  pollutant,
  type = NULL,
  percentile = c(25, 50, 75, 90, 95),
  wd_angle = 10,
  r_axis_inside = 315,
  cols = "turbo",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{One or more pollutants in \code{data} to calculate percentiles
for. Multiple pollutants limit the number of allowed \code{type}s to one.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{percentile}{Any number of percentiles (\code{0} to \code{100}) to calculate.}

\item{wd_angle}{The angle to bin the wind direction to. Should divide into
360 with no remainders; e.g., \code{5}, \code{10}, \code{12}, \code{15}, etc.}

\item{r_axis_inside}{Either a numeric value, setting a theta axis value at
which the axis should be placed inside the panel, or \code{FALSE}, which will
place it outside of panel on a vertical axis.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
\code{\link[=plot_polar_percentile]{plot_polar_percentile()}} plots percentiles by wind direction with flexible
conditioning. The plot can display any number of percentile lines, though
useful defaults have been chosen. It is most useful for showing the
distribution of concentrations by wind direction and often can reveal
different sources, e.g., those that only affect high percentile
concentrations such as a chimney stack.
}
\seealso{
the legacy \code{\link[=percentileRose]{percentileRose()}} function
}
\author{
Jack Davison
}
\concept{ggplot2 directional analysis functions}
