% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taylor_diagram.R
\name{plot_taylor_diagram}
\alias{plot_taylor_diagram}
\title{Taylor Diagram for model evaluation}
\usage{
plot_taylor_diagram(
  data,
  obs,
  mod,
  group = NULL,
  type = NULL,
  show_negative_cor = FALSE,
  scale_y = openair::scale_opts(),
  cols = "tol",
  cols_obs = "black",
  cols_crmse = "goldenrod",
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{obs, mod}{The names of the columns to summarise, one for observations
(\code{obs}) and one for modelled values (\code{mod}); these should be columns in
\code{data}.}

\item{group}{The column of \code{data} to use to condition the data within each
panel, passed to \code{\link[=cutData]{cutData()}}. This is usually achieved using colour, so
will conflict with providing multiple \code{pollutant}s or other options which
use the colour aesthetic.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{show_negative_cor}{Should the negative correlation quadrant be
displayed? If \code{FALSE}, negative correlations will be set to \code{0} to still
appear on the single-quadrant Taylor Diagram.}

\item{scale_y}{Options for the y-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{cols_obs, cols_crmse}{Colours to use for the observed value marker and
radiating CRMSE axes. Passed to \code{\link[=openColours]{openColours()}}, though most usefully
provided as a single colour (e.g, \code{"black"}).}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
The Taylor Diagram is a very useful model evaluation tool. The diagram
provides a way of showing how three complementary model performance
statistics vary simultaneously. These statistics are the correlation
coefficient R, the standard deviation (sigma) and the (centred)
root-mean-square error. These three statistics can be plotted on one (2D)
graph because of the way they are related to one another which can be
represented through the Law of Cosines.
}
\section{Controlling scales}{


In \code{openair}, a 'scale' is anything expressed in a numeric or date range.
This includes axes and colour bars. Most \code{openair} plotting functions
possess arguments in the pattern \verb{scale_*} to control these - e.g.,
\code{scale_x}, \code{scale_y}, and \code{scale_col}.

If these arguments are passed a single number, this will set the maximum of
the scale to be that number. If two numbers are provided, these will be
used as the scale range.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# x axis goes up to 100
plot_xy_scatter(mydata, "nox", "no2", scale_x = 100)

# x axis is between 50 and 150
plot_xy_scatter(mydata, "nox", "no2", scale_x = c(50, 150))
}\if{html}{\out{</div>}}

These \verb{scale_*} options can also be provided with the \code{\link[=scale_opts]{scale_opts()}}
function, which allows for further customisation, such as:
\itemize{
\item Where to add major axis/legend breaks, and how to label them.
\item Whether to transform the scale (e.g., a log or sqrt transform).
\item Where to place the axis, and whether to draw a second one (axes only).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_xy_scatter(
  mydata,
  "nox",
  "no2",
  scale_x = scale_opts(
    transform = "pseudo_log",
    breaks = c(1, 3, 10, 30, 100, 300, 1000, 3000),
    sec.axis = ggplot2::sec_axis(~.)
  )
)
}\if{html}{\out{</div>}}

Take care with date scales (e.g., \code{scale_x} in \code{\link[=plot_trend_bars]{plot_trend_bars()}});
\code{limits} should be provided as POSIXct objects. This is likely to be
refined in future. Note that, for date scales, \code{date_breaks} and
\code{date_labels} can be used to more easily refine the appearance of the
scale.
}

\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# create a small dataset
mod_data <- dplyr::select(mydata, date, nox)

# lets create some fake modelled values
mod_data <-
  dplyr::mutate(
    mod_data,
    model1 = jitter(nox, factor = 100),
    model2 = jitter(nox, factor = 100) * 1.5,
    model3 = jitter(nox, factor = 1000)
  ) |>
  tidyr::pivot_longer(model1:model3)

# plot this on a taylor diagram
plot_taylor_diagram(mod_data, "nox", "value", group = "name")

}
}
\references{
Taylor, K.E.: Summarizing multiple aspects of model performance in a single
diagram. J.  Geophys. Res., 106, 7183-7192, 2001 (also see PCMDI Report 55).
}
\seealso{
the legacy \code{\link[=TaylorDiagram]{TaylorDiagram()}} function

Other ggplot2 model evaluation functions: 
\code{\link{plot_conditional_quantile}()}
}
\author{
Jack Davison
}
\concept{ggplot2 model evaluation functions}
