% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_bars.R
\name{plot_trend_bars}
\alias{plot_trend_bars}
\title{Plot air quality trends as stacked bar charts}
\usage{
plot_trend_bars(
  data,
  pollutant,
  group = NULL,
  type = NULL,
  avg.time = "month",
  normalise = FALSE,
  scale_x = openair::scale_opts(),
  scale_y = openair::scale_opts(),
  cols = "tol",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The names of the pollutants to plot; these should be columns
in \code{data}.}

\item{group}{The splitting variable that makes up the bars in the bar chart
e.g. \code{group = "cluster"} if the output from \code{polarCluster} is being
analysed. If \code{group} is a numeric variable it is split into 4 quantiles (by
default) by \code{\link[=cutData]{cutData()}}. If \code{group} is a factor or character variable then
the categories are used directly.}

\item{type}{The name of the data series to use as the conditioning variable,
passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with maximum length
2, which creates a 2D grid of plots.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be \code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in which case
3-month seasonal values are calculated with spring defined as March, April,
May and so on.

Note that \code{avg.time} when used in \code{timeProp} should be greater than the
time gap in the original data. For example, \code{avg.time = "day"} for hourly
data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{scale_x}{Options for the x-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{scale_y}{Options for the y-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the
plotting of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and
\code{\link[=trajCluster]{trajCluster()}} that consider local and regional (back trajectory) cluster
analysis respectively. However, the function has more general use for
understanding time series data. In order to plot time series in this way,
some sort of time aggregation is needed, which is controlled by the option
\code{avg.time}.
}
\section{Controlling scales}{


In \code{openair}, a 'scale' is anything expressed in a numeric or date range.
This includes axes and colour bars. Most \code{openair} plotting functions
possess arguments in the pattern \verb{scale_*} to control these - e.g.,
\code{scale_x}, \code{scale_y}, and \code{scale_col}.

If these arguments are passed a single number, this will set the maximum of
the scale to be that number. If two numbers are provided, these will be
used as the scale range.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# x axis goes up to 100
plot_xy_scatter(mydata, "nox", "no2", scale_x = 100)

# x axis is between 50 and 150
plot_xy_scatter(mydata, "nox", "no2", scale_x = c(50, 150))
}\if{html}{\out{</div>}}

These \verb{scale_*} options can also be provided with the \code{\link[=scale_opts]{scale_opts()}}
function, which allows for further customisation, such as:
\itemize{
\item Where to add major axis/legend breaks, and how to label them.
\item Whether to transform the scale (e.g., a log or sqrt transform).
\item Where to place the axis, and whether to draw a second one (axes only).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_xy_scatter(
  mydata,
  "nox",
  "no2",
  scale_x = scale_opts(
    transform = "pseudo_log",
    breaks = c(1, 3, 10, 30, 100, 300, 1000, 3000),
    sec.axis = ggplot2::sec_axis(~.)
  )
)
}\if{html}{\out{</div>}}

Take care with date scales (e.g., \code{scale_x} in \code{\link[=plot_trend_bars]{plot_trend_bars()}});
\code{limits} should be provided as POSIXct objects. This is likely to be
refined in future. Note that, for date scales, \code{date_breaks} and
\code{date_labels} can be used to more easily refine the appearance of the
scale.
}

\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
# monthly plot of SO2 showing the contribution by wind sector
timeProp(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
the legacy \code{\link[=timeProp]{timeProp()}} function

Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_smooth}()},
\code{\link{plot_trend_theilsen}()}

Other cluster analysis functions: 
\code{\link{polarCluster}()},
\code{\link{timeProp}()},
\code{\link{trajCluster}()}
}
\author{
Jack Davison

David Carslaw
}
\concept{cluster analysis functions}
\concept{ggplot2 time series and trend functions}
