% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_lines.R
\name{plot_trend_lines}
\alias{plot_trend_lines}
\title{Plot air quality trends using a tradition line chart}
\usage{
plot_trend_lines(
  data,
  pollutant,
  group = NULL,
  type = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  date_pad = FALSE,
  windflow = FALSE,
  scale_x = openair::scale_opts(),
  scale_y = openair::scale_opts(),
  cols = "tol",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The names of the pollutants to plot; these should be columns
in \code{data}.}

\item{group}{The name of the data series to use as a grouping variable within
each panel, passed to \code{\link[=cutData]{cutData()}}. The grouping variable is mapped to a
colour, and so cannot be used if the colour channel is already in use
(e.g., if multiple \code{pollutants} have been specified).}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

Type can be up length two e.g. \code{type = c("season", "weekday")} will produce
a 2x2 plot split by season and day of the week. Note, when two types are
provided the first forms the columns and the second the rows.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{Statistic used for calculating monthly values. Default is
\code{"mean"}, but can also be \code{"percentile"}. See \code{\link[=timeAverage]{timeAverage()}} for more
details.}

\item{percentile}{Percentile value(s) to use if \code{statistic = "percentile"} is
chosen. Can be a vector of numbers e.g. \code{percentile = c(5, 50, 95)} will
plot the 5th, 50th and 95th percentile values together on the same plot.}

\item{date_pad}{Should missing data be padded-out? This is useful where a
data frame consists of two or more "chunks" of data with time gaps between
them. By setting \code{date.pad = TRUE} the time gaps between the chunks are
shown properly, rather than with a line connecting each chunk. For
irregular data, set to \code{FALSE}. Note, this should not be set for \code{type}
other than \code{default}.}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}. Alternatively, \code{\link[=windflow_opts]{windflow_opts()}}
allows for greater customisation of the arrows (e.g., wind speed ranges,
relative size of arrows, arrow appearance).}

\item{scale_x}{Options for the x-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{scale_y}{Options for the y-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=plot_trend_lines]{plot_trend_lines()}} function simply plots a traditional line chart of
air pollution data. Some additional functionality is also included, such as
self-contained time-averaging and the addition of 'windflow' wind arrows.
}
\section{Controlling scales}{


In \code{openair}, a 'scale' is anything expressed in a numeric or date range.
This includes axes and colour bars. Most \code{openair} plotting functions
possess arguments in the pattern \verb{scale_*} to control these - e.g.,
\code{scale_x}, \code{scale_y}, and \code{scale_col}.

If these arguments are passed a single number, this will set the maximum of
the scale to be that number. If two numbers are provided, these will be
used as the scale range.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# x axis goes up to 100
plot_xy_scatter(mydata, "nox", "no2", scale_x = 100)

# x axis is between 50 and 150
plot_xy_scatter(mydata, "nox", "no2", scale_x = c(50, 150))
}\if{html}{\out{</div>}}

These \verb{scale_*} options can also be provided with the \code{\link[=scale_opts]{scale_opts()}}
function, which allows for further customisation, such as:
\itemize{
\item Where to add major axis/legend breaks, and how to label them.
\item Whether to transform the scale (e.g., a log or sqrt transform).
\item Where to place the axis, and whether to draw a second one (axes only).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_xy_scatter(
  mydata,
  "nox",
  "no2",
  scale_x = scale_opts(
    transform = "pseudo_log",
    breaks = c(1, 3, 10, 30, 100, 300, 1000, 3000),
    sec.axis = ggplot2::sec_axis(~.)
  )
)
}\if{html}{\out{</div>}}

Take care with date scales (e.g., \code{scale_x} in \code{\link[=plot_trend_bars]{plot_trend_bars()}});
\code{limits} should be provided as POSIXct objects. This is likely to be
refined in future. Note that, for date scales, \code{date_breaks} and
\code{date_labels} can be used to more easily refine the appearance of the
scale.
}

\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
# trend plot for nox
plot_trend_smooth(mydata, pollutant = "nox")

# trend plot by each of 8 wind sectors
\dontrun{
plot_trend_smooth(mydata, pollutant = "o3", type = "wd")

# several pollutants
plot_trend_smooth(mydata, pollutant = c("no2", "o3", "pm10", "pm25"))

# percentiles
plot_trend_smooth(
  mydata,
  pollutant = "o3",
  statistic = "percentile",
  percentile = 95
)
}
}
\seealso{
the legacy \code{\link[=smoothTrend]{smoothTrend()}} function

Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_smooth}()},
\code{\link{plot_trend_theilsen}()}
}
\author{
Jack Davison

David Carslaw
}
\concept{ggplot2 time series and trend functions}
