% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_smooth.R
\name{plot_trend_smooth}
\alias{plot_trend_smooth}
\title{Plot air quality trends with a non-parametric smooth curve}
\usage{
plot_trend_smooth(
  data,
  pollutant = "nox",
  group = NULL,
  type = NULL,
  avg.time = "month",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  smooth_method = "gam",
  smooth_formula = NULL,
  smooth_se = TRUE,
  smooth_level = 0.95,
  deseason = FALSE,
  windflow = FALSE,
  cols = "tol",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The name of the pollutant to summarise.}

\item{group}{The name of the data series to use as a grouping variable within
each panel, passed to \code{\link[=cutData]{cutData()}}. The grouping variable is mapped to a
colour, and so cannot be used if the colour channel is already in use
(e.g., if multiple \code{pollutants} have been specified).}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

Type can be up length two e.g. \code{type = c("season", "weekday")} will produce
a 2x2 plot split by season and day of the week. Note, when two types are
provided the first forms the columns and the second the rows.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{Statistic used for calculating monthly values. Default is
\code{"mean"}, but can also be \code{"percentile"}. See \code{\link[=timeAverage]{timeAverage()}} for more
details.}

\item{percentile}{Percentile value(s) to use if \code{statistic = "percentile"} is
chosen. Can be a vector of numbers e.g. \code{percentile = c(5, 50, 95)} will
plot the 5th, 50th and 95th percentile values together on the same plot.}

\item{smooth_method, smooth_formula, smooth_se, smooth_level}{Options to control
various parameters for the smooth trend function. Passed to the
corresponding arguments of \code{\link[ggplot2:geom_smooth]{ggplot2::stat_smooth()}}; see that function for
definitions.}

\item{deseason}{Should the data be de-deasonalized first? If \code{TRUE} the
function \code{stl} is used (seasonal trend decomposition using loess). Note
that if \code{TRUE} missing data are first imputed using a Kalman filter and
Kalman smooth.}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Not currently used.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=plot_trend_smooth]{plot_trend_smooth()}} function provides a flexible way of estimating
the trend in the concentration of a pollutant or other variable. Monthly
mean values are calculated from an hourly (or higher resolution) or daily
time series. There is the option to deseasonalise the data if there is
evidence of a seasonal cycle.
}
\examples{
# trend plot for nox
plot_trend_smooth(mydata, pollutant = "nox")

# trend plot by each of 8 wind sectors
\dontrun{
plot_trend_smooth(mydata, pollutant = "o3", type = "wd")

# several pollutants
plot_trend_smooth(mydata, pollutant = c("no2", "o3", "pm10", "pm25"))

# percentiles
plot_trend_smooth(
  mydata,
  pollutant = "o3",
  statistic = "percentile",
  percentile = 95
)
}
}
\seealso{
the legacy \code{\link[=smoothTrend]{smoothTrend()}} function

Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()}
}
\author{
Jack Davison

David Carslaw
}
\concept{ggplot2 time series and trend functions}
