% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_theilsen.R
\name{plot_trend_theilsen}
\alias{plot_trend_theilsen}
\title{Plot an air quality timeseries chart with Theil-Sen slope estimates}
\usage{
plot_trend_theilsen(
  data,
  pollutant = "nox",
  type = NULL,
  avg.time = "month",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  alpha = 0.05,
  deseason = FALSE,
  autocor = FALSE,
  cols = "tol",
  slope_unit = NULL,
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The names of the pollutants to plot; these should be columns
in \code{data}.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option --- see below. Not used if \code{avg.time = "default"}.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{alpha}{For the confidence interval calculations of the slope. The
default is 0.05. To show 99\\% confidence intervals for the value of the
trend, choose alpha = 0.01 etc.}

\item{deseason}{Should the data be de-deasonalized first? If \code{TRUE} the
function \code{stl} is used (seasonal trend decomposition using loess). Note
that if \code{TRUE} missing data are first imputed using a Kalman filter and
Kalman smooth.}

\item{autocor}{Should autocorrelation be considered in the trend uncertainty
estimates? The default is \code{FALSE}. Generally, accounting for
autocorrelation increases the uncertainty of the trend estimate ---
sometimes by a large amount.}

\item{cols}{The colour scheme to use in the plot. Passed to \code{\link[=openColours]{openColours()}}.
Likely more usefully, three colours can be defined in a single vector. For
example, when \code{cols = c("blue", "red", "green")}, the data will be coloured
blue, the trend lines red, and the annotations green.}

\item{slope_unit}{The text shown for the slope (default is
\sQuote{units/year}).}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=plot_trend_theilsen]{plot_trend_theilsen()}} function provides a collection of functions to
analyse trends in air pollution data. It is flexible in the sense that it
can be applied to data in many ways, e.g., by day of the week, hour of day
and wind direction. This flexibility makes it much easier to draw
inferences from data e.g. why is there a strong downward trend in
concentration from one wind sector and not another, or why trends on one
day of the week or a certain time of day are unexpected.

For data that are strongly seasonal, perhaps from a background site, or a
pollutant such as ozone, it will be important to deseasonalise the data
(using the option \code{deseason = TRUE}. Similarly, for data that increase,
then decrease, or show sharp changes it may be better to use
\code{\link[=plot_trend_smooth]{plot_trend_smooth()}}.

A minimum of 6 points are required for trend estimates to be made.

Note that the symbols shown next to each trend estimate relate to how
statistically significant the trend estimate is: p $<$ 0.001 =
***, p $<$ 0.01 = **, p $<$ 0.05 = * and p $<$ 0.1 = $+$.

Some of the code used in \code{\link[=TheilSen]{TheilSen()}} is based on that from Rand Wilcox.
This mostly relates to the Theil-Sen slope estimates and uncertainties.
Further modifications have been made to take account of correlated data
based on Kunsch (1989). The basic function has been adapted to take account
of auto-correlated data using block bootstrap simulations if \code{autocor = TRUE} (Kunsch, 1989). We follow the suggestion of Kunsch (1989) of setting
the block length to n(1/3) where n is the length of the time series.
}
\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
# trend plot for nox
plot_trend_theilsen(mydata, pollutant = "nox")

\dontrun{
# trend plot for ozone with p=0.01 i.e. uncertainty in slope shown at
# 99 \% confidence interval
plot_trend_theilsen(mydata, pollutant = "o3", alpha = 0.01)

# trend plot by each of 8 wind sectors
plot_trend_theilsen(mydata, pollutant = "o3", type = "wd")

# and for a subset of data (from year 2000 onwards)
plot_trend_theilsen(selectByDate(mydata, year = 2000:2005), pollutant = "o3")
}
}
\references{
Helsel, D., Hirsch, R., 2002. Statistical methods in water resources. US
Geological Survey. Note that this is a very good resource for statistics as
applied to environmental data.

Hirsch, R. M., Slack, J. R., Smith, R. A., 1982. Techniques of trend analysis
for monthly water-quality data. Water Resources Research 18 (1), 107-121.

Kunsch, H. R., 1989. The jackknife and the bootstrap for general stationary
observations. Annals of Statistics 17 (3), 1217-1241.

Sen, P. K., 1968. Estimates of regression coefficient based on Kendall's tau.
Journal of the American Statistical Association 63(324).

Theil, H., 1950. A rank invariant method of linear and polynomial regression
analysis, i, ii, iii. Proceedings of the Koninklijke Nederlandse Akademie
Wetenschappen, Series A - Mathematical Sciences 53, 386-392, 521-525,
1397-1412.

See also several of the Air Quality Expert Group (AQEG) reports for the use
of similar tests applied to UK/European air quality data.
}
\seealso{
Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_smooth}()}
}
\author{
David Carslaw

Jack Davison

Rand Wilcox (trend code)
}
\concept{ggplot2 time series and trend functions}
