% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation.R
\name{plot_variation}
\alias{plot_variation}
\title{Plot the variation in a variable with flexible conditioning}
\usage{
plot_variation(
  data,
  pollutant,
  x = "month",
  group = NULL,
  type = NULL,
  statistic = c("mean", "median"),
  stat_interval = TRUE,
  windflow = FALSE,
  cols = "tol",
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The name of the pollutant to summarise.}

\item{x, type}{The name of the data series to use as the x-axis or
conditioning variable, passed to \code{\link[=cutData]{cutData()}}. These are used before
applying \code{statistic}. \code{type} may be \code{NULL} or a vector with maximum length
2, which creates a 2D grid of plots.}

\item{group}{The name of the data series to use to colour traces within a
single panel, passed to \code{\link[=cutData]{cutData()}}.}

\item{statistic}{One of:
\itemize{
\item \code{"mean"}, which displays the bootstrapped mean confidence interval in the data.
\item \code{"median"}, which displays the median value along with a quantile range.

The \code{stat_interval} argument helps control how the intervals are
calculated.
}}

\item{stat_interval}{A single numeric value between \code{0} and \code{1}. When
\code{statistic = "mean"}, this is the confidence interval around the
bootstrapped mean. When \code{statistic = "median"}, this is the upper quantile
(the lower quantile is calculated using \code{1 - stat_interval}). \code{TRUE} will
use sensible defaults (\code{0.95} for \code{"mean"}, \code{0.75} for \code{"median"}) and
\code{FALSE} will remove intervals entirely.}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Not currently used.}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2} plot, or \code{data.frame} if \code{plot = FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function plots the variation of a pollutant or other variable (e.g.,
bootstrapped confidence interval around the mean, or a quantile range around
the median). These variations are grouped by up to three different
conditioning variables - \code{x} (x-axis), \code{group} (colour) and \code{type} (panels).
\code{x} may usefully be a temporal variable for time series data (e.g.,
\code{"month"}, \code{"year"}, etc.), but could be any factor (e.g., \code{"site"}) or
numeric (e.g., \code{"no2"}, which will be binned) variable.
}
\seealso{
The legacy \code{\link[=timeVariation]{timeVariation()}} function
}
\author{
Jack Davison
}
