% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_scatter.R
\name{plot_xy_scatter}
\alias{plot_xy_scatter}
\alias{plot_xy_bins}
\title{Plot the relationship between two (or more) variables using scatter plot
variations}
\usage{
plot_xy_scatter(
  data,
  x,
  y,
  z = NULL,
  group_col = NULL,
  group_shp = NULL,
  type = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  discretise = NULL,
  scale_x = openair::scale_opts(),
  scale_y = openair::scale_opts(),
  scale_col = openair::scale_opts(),
  cols = NULL,
  windflow = FALSE,
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)

plot_xy_bins(
  data,
  x,
  y,
  type = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  method = c("bin", "hexbin", "density"),
  binwidth = NULL,
  bins = 30,
  breaks = NULL,
  adjust = 1,
  dist = 0.02,
  discretise = NULL,
  scale_x = openair::scale_opts(),
  scale_y = openair::scale_opts(),
  scale_col = openair::scale_opts(),
  cols = "viridis",
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x, y}{The columns of \code{data} to use as the x and y variables.}

\item{z}{The column of \code{data} to colour by. This is not passed though
\code{\link[=cutData]{cutData()}}, so a colour}

\item{group_col, group_shp}{Columns of \code{data} to use to condition the data
within each panel, passed to \code{\link[=cutData]{cutData()}}. \code{group_col} maps to the colour
aesthetic and \code{group_shp} to the marker shape aesthetic.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option --- see below. Not used if \code{avg.time = "default"}.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{discretise}{Options to discretise the continuous scale into discrete
bins. Providing an integer will split the scale into that number of
quantiles. Alternatively, \code{\link[=discretise_breaks]{discretise_breaks()}} allows for extensive
customisation (e.g., defining specific bin widths, breaks, labels, and so
on).}

\item{scale_x}{Options for the x-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{scale_y}{Options for the y-axis scale of the plot. If a single number,
this will set an upper scale limit. If greater than one number, the range
of the input will be used as the axis limits. Can also be \code{\link[=scale_opts]{scale_opts()}},
which permit further customisation such as setting axis breaks, labels and
scale transforms.}

\item{scale_col}{Options for the colour scale of the plot. If a single
number, this will set an upper scale limit. If greater than one number, the
range of the input will be used as the axis limits. Can also be
\code{\link[=scale_opts]{scale_opts()}}, which permit further customisation such as setting axis
breaks, labels and scale transforms. Note that these options will not apply
if \code{discretise} is used, as this overrides the continuous scale.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}. Alternatively, \code{\link[=windflow_opts]{windflow_opts()}}
allows for greater customisation of the arrows (e.g., wind speed ranges,
relative size of arrows, arrow appearance).}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}

\item{method}{One of:
\itemize{
\item \code{"bin"}, which bins the data using square bins
\item \code{"hexbin"}, which uses regular hexagonal binning, or
\item \code{"density"}, which fits a 2D kernel density estimation.
}}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that takes x after scale transformation as input and
returns a single numeric value. When specifying a function along with a
grouping structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{breaks}{Alternatively, you can supply a numeric vector giving
the bin boundaries. Overrides \code{binwidth}, \code{bins}, \code{center},
and \code{boundary}. Can also be a function that takes group-wise values as input and returns bin boundaries.}

\item{adjust}{Adjustment factor to kernel density bandwidth. Used in
\code{\link[=plot_xy_bins]{plot_xy_bins()}} when \code{method = "density"}.}

\item{dist}{How far from the original data predictions should be made; passed
to \code{\link[mgcv:exclude.too.far]{mgcv::exclude.too.far()}}. Values should be between 0 and 1. Used in
\code{\link[=plot_xy_bins]{plot_xy_bins()}} when \code{method = "density"}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This family of functions allows for two or more pollutants to be directly
compared using a relational diagram, with convenient time averaging. This
family of plotting functions encompasses:
\itemize{
\item \code{\link[=plot_xy_scatter]{plot_xy_scatter()}}, which plots a conventional scatter plot
\item \code{\link[=plot_xy_bins]{plot_xy_bins()}}, which bins large amount of data either in square or hexagonal bins
}
}
\section{Controlling scales}{


In \code{openair}, a 'scale' is anything expressed in a numeric or date range.
This includes axes and colour bars. Most \code{openair} plotting functions
possess arguments in the pattern \verb{scale_*} to control these - e.g.,
\code{scale_x}, \code{scale_y}, and \code{scale_col}.

If these arguments are passed a single number, this will set the maximum of
the scale to be that number. If two numbers are provided, these will be
used as the scale range.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# x axis goes up to 100
plot_xy_scatter(mydata, "nox", "no2", scale_x = 100)

# x axis is between 50 and 150
plot_xy_scatter(mydata, "nox", "no2", scale_x = c(50, 150))
}\if{html}{\out{</div>}}

These \verb{scale_*} options can also be provided with the \code{\link[=scale_opts]{scale_opts()}}
function, which allows for further customisation, such as:
\itemize{
\item Where to add major axis/legend breaks, and how to label them.
\item Whether to transform the scale (e.g., a log or sqrt transform).
\item Where to place the axis, and whether to draw a second one (axes only).
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_xy_scatter(
  mydata,
  "nox",
  "no2",
  scale_x = scale_opts(
    transform = "pseudo_log",
    breaks = c(1, 3, 10, 30, 100, 300, 1000, 3000),
    sec.axis = ggplot2::sec_axis(~.)
  )
)
}\if{html}{\out{</div>}}

Take care with date scales (e.g., \code{scale_x} in \code{\link[=plot_trend_bars]{plot_trend_bars()}});
\code{limits} should be provided as POSIXct objects. This is likely to be
refined in future. Note that, for date scales, \code{date_breaks} and
\code{date_labels} can be used to more easily refine the appearance of the
scale.
}

\section{Conditioning with \code{type}}{


Many \code{openair} plotting functions can be conditioned by using the \code{type}
argument. This splits a single plot into a grid of smaller plots,
popularised under the name "small multiples" by Edward Tufte and more
recently referred to as "faceting" (e.g., in \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}).
\code{type} is always passed to \code{\link[=cutData]{cutData()}}, meaning various presets like
\code{"year"} or \code{"month"} can be used, and numeric variables are automatically
binned.

If a single \code{type} is given, the default behaviour is for each panel to
wrap into a roughly square grid. If two \code{type}s are given, the first will
be used for the rows of the resulting grid and the second for the columns.

Not all plots can take two \code{type}s. This is usually because one facet
dimension is already reserved for the specific plot type. For example, in
\code{\link[=plot_calendar]{plot_calendar()}}, the plot is always faceted by \code{"month"} even if \code{type = NULL}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# one type (square 2D grid)
plot_heatmap(mydata, "no2", type = "year")

# two types (two 'axes' of types)
plot_heatmap(mydata, "no2", type = c("o3", "weekend"))
}\if{html}{\out{</div>}}

Functions with \code{type} also possess the \code{facet_opts} argument, which takes
the \code{\link[=facet_opts]{facet_opts()}} function. This allows for finer control of how faceting
works. This includes:
\itemize{
\item The number of rows/columns when only one \code{type} has been given.
\item Whether axes should share the same x/y limits, or if each panel should
have unique scales.
\item If scales are free, whether each panel should have a different size
proportional to their scales.
\item On which side of the plot where the facet labels should be drawn.
\item Whether all axes and axis labels should be drawn, or whether they should
only be drawn on the margins.
}

Note that not all options will apply to all plots. For example, plots on a
polar axis always have fixed x-axes that are mapped to the compass
directions.

\if{html}{\out{<div class="sourceCode">}}\preformatted{plot_heatmap(
  mydata,
  "no2",
  type = "year",
  facet_opts = facet_opts(nrow = 2, axes = "all_y", strip.position = "right")
)
}\if{html}{\out{</div>}}
}

\examples{
dat2004 <- selectByDate(mydata, year = 2004)

# basic use, single pollutant
plot_xy_scatter(dat2004, x = "nox", y = "no2")

\dontrun{
# scatterPlot by year
plot_xy_scatter(mydata, x = "nox", y = "no2", type = "year")
}

# examine group and type arguments
plot_xy_scatter(
  dat2004,
  x = "nox",
  y = "no2",
  group_col = "weekday",
  group_shp = "season",
  type = "weekend"
)

\dontrun{
# colour by CO
plot_xy_scatter(dat2004, x = "nox", y = "no2", z = "co", avg.time = "day")

# multiple types and grouping variable and continuous colour scale
plot_xy_scatter(
  mydata,
  x = "nox",
  y = "no2",
  z = "o3",
  type = c("season", "weekend")
)

# use binning
plot_xy_bins(mydata, x = "nox", y = "no2", method = "bin")
plot_xy_bins(mydata, x = "nox", y = "no2", method = "hexbin")
plot_xy_bins(mydata, x = "nox", y = "no2", method = "density")
}
}
\seealso{
the legacy \code{\link[=scatterPlot]{scatterPlot()}} function
}
