% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarAnnulus.R
\name{polarAnnulus}
\alias{polarAnnulus}
\title{Bivariate polarAnnulus plot}
\usage{
polarAnnulus(
  mydata,
  pollutant = "nox",
  resolution = "fine",
  local.tz = NULL,
  period = "hour",
  type = "default",
  statistic = "mean",
  percentile = NA,
  limits = NULL,
  cols = "default",
  col.na = "white",
  offset = 50,
  angle.scale = 0,
  min.bin = 1,
  exclude.missing = TRUE,
  date.pad = FALSE,
  force.positive = TRUE,
  k = c(20, 10),
  normalise = FALSE,
  strip.position = "top",
  key = TRUE,
  key.title = paste(statistic, pollutant, sep = "\\n"),
  key.position = "right",
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame minimally containing \code{date}, \code{wd} and a
pollutant.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}. There can also be
more than one pollutant specified e.g. \code{pollutant = c("nox", "no2")}. The
main use of using two or more pollutants is for model evaluation where two
species would be expected to have similar concentrations. This saves the
user stacking the data and it is possible to work with columns of data
directly. A typical use would be \code{pollutant = c("obs", "mod")} to compare
two columns \dQuote{obs} (the observations) and \dQuote{mod} (modelled
values).}

\item{resolution}{Two plot resolutions can be set: \dQuote{normal} and
\dQuote{fine} (the default).}

\item{local.tz}{Should the results be calculated in local time that includes
a treatment of daylight savings time (DST)? The default is not to consider
DST issues, provided the data were imported without a DST offset. Emissions
activity tends to occur at local time e.g. rush hour is at 8 am every day.
When the clocks go forward in spring, the emissions are effectively
released into the atmosphere typically 1 hour earlier during the summertime
i.e. when DST applies. When plotting diurnal profiles, this has the effect
of \dQuote{smearing-out} the concentrations. Sometimes, a useful approach
is to express time as local time. This correction tends to produce
better-defined diurnal profiles of concentration (or other variables) and
allows a better comparison to be made with emissions/activity data. If set
to \code{FALSE} then GMT is used. Examples of usage include \code{local.tz = "Europe/London"}, \code{local.tz = "America/New_York"}. See \code{cutData} and
\code{import} for more details.}

\item{period}{This determines the temporal period to consider. Options are
\dQuote{hour} (the default, to plot diurnal variations), \dQuote{season} to
plot variation throughout the year, \dQuote{weekday} to plot day of the
week variation and \dQuote{trend} to plot the trend by wind direction.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{statistic}{The statistic that should be applied to each wind
speed/direction bin. Can be \dQuote{mean} (default), \dQuote{median},
\dQuote{max} (maximum), \dQuote{frequency}. \dQuote{stdev} (standard
deviation), \dQuote{weighted.mean} or \dQuote{cpf} (Conditional Probability
Function). Because of the smoothing involved, the colour scale for some of
these statistics is only to provide an indication of overall pattern and
should not be interpreted in concentration units e.g. for \code{statistic = "weighted.mean"} where the bin mean is multiplied by the bin frequency and
divided by the total frequency. In many cases using \code{polarFreq} will be
better. Setting \code{statistic = "weighted.mean"} can be useful because it
provides an indication of the concentration * frequency of occurrence and
will highlight the wind speed/direction conditions that dominate the
overall mean.}

\item{percentile}{If \code{statistic = "percentile"} or \code{statistic = "cpf"} then
\code{percentile} is used, expressed from 0 to 100. Note that the percentile
value is calculated in the wind speed, wind direction \sQuote{bins}. For
this reason it can also be useful to set \code{min.bin} to ensure there are a
sufficient number of points available to estimate a percentile. See
\code{quantile} for more details of how percentiles are calculated.}

\item{limits}{The function does its best to choose sensible limits
automatically. However, there are circumstances when the user will wish to
set different ones. An example would be a series of plots showing each year
of data separately. The limits are set in the form \code{c(lower, upper)}, so
\code{limits = c(0, 100)} would force the plot limits to span 0-100.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{col.na}{Colour to be used to show missing data.}

\item{offset}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{min.bin}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points. It is recommended to consider your data with
care. Also, the \code{polarFreq} function can be of use in such circumstances.}

\item{exclude.missing}{Setting this option to \code{TRUE} (the default) removes
points from the plot that are too far from the original data. The smoothing
routines will produce predictions at points where no data exist i.e. they
predict. By removing the points too far from the original data produces a
plot where it is clear where the original data lie. If set to \code{FALSE}
missing data will be interpolated.}

\item{date.pad}{For \code{type = "trend"} (default), \code{date.pad = TRUE} will
pad-out missing data to the beginning of the first year and the end of the
last year. The purpose is to ensure that the trend plot begins and ends at
the beginning or end of year.}

\item{force.positive}{The default is \code{TRUE}. Sometimes if smoothing data with
steep gradients it is possible for predicted values to be negative.
\code{force.positive = TRUE} ensures that predictions remain positive. This is
useful for several reasons. First, with lots of missing data more
interpolation is needed and this can result in artefacts because the
predictions are too far from the original data. Second, if it is known
beforehand that the data are all positive, then this option carries that
assumption through to the prediction. The only likely time where setting
\code{force.positive = FALSE} would be if background concentrations were first
subtracted resulting in data that is legitimately negative. For the vast
majority of situations it is expected that the user will not need to alter
the default option.}

\item{k}{The smoothing value supplied to \code{gam} for the temporal and wind
direction components, respectively. In some cases e.g. a trend plot with
less than 1-year of data the smoothing with the default values may become
too noisy and affected more by outliers. Choosing a lower value of \code{k} (say
10) may help produce a better plot.}

\item{normalise}{If \code{TRUE} concentrations are normalised by dividing by their
mean value. This is done \emph{after} fitting the smooth surface. This option is
particularly useful if one is interested in the patterns of concentrations
for several pollutants on different scales e.g. NOx and CO. Often useful if
more than one \code{pollutant} is chosen.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{annotate = FALSE} will not plot the N/E/S/W labels.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
Typically plots the concentration of a pollutant by wind direction and as a
function of time as an annulus. The function is good for visualising how
concentrations of pollutants vary by wind direction and a time period e.g. by
month, day of week.
}
\details{
The \code{polarAnnulus} function shares many of the properties of the
\code{polarPlot}. However, \code{polarAnnulus} is focussed on displaying
information on how concentrations of a pollutant (values of another variable)
vary with wind direction and time. Plotting as an annulus helps to reduce
compression of information towards the centre of the plot. The circular plot
is easy to interpret because wind direction is most easily understood in
polar rather than Cartesian coordinates.

The inner part of the annulus represents the earliest time and the outer part
of the annulus the latest time. The time dimension can be shown in many ways
including "trend", "hour" (hour or day), "season" (month of the year) and
"weekday" (day of the week). Taking hour as an example, the plot will show
how concentrations vary by hour of the day and wind direction. Such plots can
be very useful for understanding how different source influences affect a
location.

For \code{type = "trend"} the amount of smoothing does not vary linearly with
the length of the time series i.e. a certain amount of smoothing per unit
interval in time. This is a deliberate choice because should one be
interested in a subset (in time) of data, more detail will be provided for
the subset compared with the full data set. This allows users to investigate
specific periods in more detail. Full flexibility is given through the
smoothing parameter \code{k}.
}
\examples{
# diurnal plot for PM10 at Marylebone Rd
\dontrun{
polarAnnulus(mydata,
  pollutant = "pm10",
  main = "diurnal variation in pm10 at Marylebone Road"
)
}

# seasonal plot for PM10 at Marylebone Rd
\dontrun{
polarAnnulus(mydata, poll = "pm10", period = "season")
}

# trend in coarse particles (PMc = PM10 - PM2.5), calculate PMc first

mydata$pmc <- mydata$pm10 - mydata$pm25
\dontrun{
polarAnnulus(mydata,
  poll = "pmc", period = "trend",
  main = "trend in pmc at Marylebone Road"
)
}
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarCluster}()},
\code{\link{polarDiff}()},
\code{\link{polarFreq}()},
\code{\link{polarPlot}()},
\code{\link{pollutionRose}()},
\code{\link{windRose}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{polar directional analysis functions}
