% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarCluster.R
\name{polarCluster}
\alias{polarCluster}
\title{K-means clustering of bivariate polar plots}
\usage{
polarCluster(
  mydata,
  pollutant = "nox",
  x = "ws",
  wd = "wd",
  n.clusters = 6,
  after = NA,
  cols = "Paired",
  angle.scale = 315,
  units = x,
  auto.text = TRUE,
  plot = TRUE,
  plot.data = FALSE,
  ...
)
}
\arguments{
\item{mydata}{A data frame minimally containing a decimal wind direction,
another variable to plot in polar coordinates (the default is a column
\code{"ws"} --- wind speed) and a pollutant. Should also contain \code{date} if plots
by time period are required.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}. Only one pollutant
can be chosen.}

\item{x}{Name of variable to plot against wind direction in polar
coordinates, the default is wind speed, \dQuote{ws}.}

\item{wd}{The name of the column in \code{mydata} representing the decimal wind
direction, 0 to 360 where 0/360 are North and 180 is South. Defaults to
\code{"wd"}.}

\item{n.clusters}{Number of clusters to use. If \code{n.clusters} is more than
length 1, then a faceted plot will be output showing the clusters
identified for each one of \code{n.clusters}.}

\item{after}{The function can be applied to differences between polar plot
surfaces (see \link{polarDiff} for details). If an \code{after} data frame is
supplied, the clustering will be carried out on the differences between
\code{after} and \code{mydata} in the same way as \link{polarDiff}.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{units}{The units shown on the polar axis scale.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{plot.data}{By default, the \code{data} component of \code{polarCluster()}
contains the original data frame appended with a new "cluster" column. When
\code{plot.data = TRUE}, the \code{data} component instead contains data to reproduce
the clustered polar plot itself (similar to \code{data} returned by
\code{\link[=polarPlot]{polarPlot()}}). This may be useful for re-plotting the \code{polarCluster()}
plot in other ways.}

\item{...}{Other arguments passed on to \code{\link[=polarPlot]{polarPlot()}}.}
}
\value{
an \link[=openair-package]{openair} object. The object includes four main
components: \code{call}, the command used to generate the plot; \code{data}, by
default the original data frame with a new field \code{cluster} identifying the
cluster, \code{clust_stats} giving the contributions made by each cluster to
number of measurements, their percentage and the percentage by pollutant;
and \code{plot}, the plot itself. Note that any rows where the value of
\code{pollutant} is \code{NA} are ignored so that the returned data frame may have
fewer rows than the original.

If the clustering is carried out considering differences, i.e., an \code{after}
data frame is supplied, the output also includes the \code{after} data frame
with cluster identified.
}
\description{
Function for identifying clusters in bivariate polar plots (\code{\link[=polarPlot]{polarPlot()}});
identifying clusters in the original data for subsequent processing.
}
\details{
Bivariate polar plots generated using the \code{polarPlot} function provide a very
useful graphical technique for identifying and characterising different air
pollution sources. While bivariate polar plots provide a useful graphical
indication of potential sources, their location and wind-speed or other
variable dependence, they do have several limitations. Often, a `feature'
will be detected in a plot but the subsequent analysis of data meeting
particular wind speed/direction criteria will be based only on the judgement
of the investigator concerning the wind speed-direction intervals of
interest. Furthermore, the identification of a feature can depend on the
choice of the colour scale used, making the process somewhat arbitrary.

\code{polarCluster} applies Partition Around Medoids (PAM) clustering techniques
to \code{\link[=polarPlot]{polarPlot()}} surfaces to help identify potentially interesting features
for further analysis. Details of PAM can be found in the \code{cluster} package (a
core R package that will be pre-installed on all R systems). PAM clustering
is similar to k-means but has several advantages e.g. is more robust to
outliers. The clustering is based on the equal contribution assumed from the
u and v wind components and the associated concentration. The data are
standardized before clustering takes place.

The function works best by first trying different numbers of clusters and
plotting them. This is achieved by setting \code{n.clusters} to be of length more
than 1. For example, if \code{n.clusters = 2:10} then a plot will be output
showing the 9 cluster levels 2 to 10.

The clustering can also be applied to differences in polar plot surfaces (see
\code{\link[=polarDiff]{polarDiff()}}). On this case a second data frame (\code{after}) should be
supplied.

Note that clustering is computationally intensive and the function can take a
long time to run --- particularly when the number of clusters is increased.
For this reason it can be a good idea to run a few clusters first to get a
feel for it e.g. \code{n.clusters = 2:5}.

Once the number of clusters has been decided, the user can then run
\code{polarCluster} to return the original data frame together with a new column
\code{cluster}, which gives the cluster number as a character (see example). Note
that any rows where the value of \code{pollutant} is \code{NA} are ignored so that the
returned data frame may have fewer rows than the original.

Note that there are no automatic ways in ensuring the most appropriate number
of clusters as this is application dependent. However, there is often
a-priori information available on what different features in polar plots
correspond to. Nevertheless, the appropriateness of different clusters is
best determined by post-processing the data. The Carslaw and Beevers (2012)
paper discusses these issues in more detail.

Note that unlike most other \code{openair} functions only a single \code{type}
\dQuote{default} is allowed.
}
\examples{
\dontrun{
# plot 2-8 clusters. Warning! This can take several minutes...
polarCluster(mydata, pollutant = "nox", n.clusters = 2:8)

# basic plot with 6 clusters
results <- polarCluster(mydata, pollutant = "nox", n.clusters = 6)

# get results, could read into a new data frame to make it easier to refer to
# e.g. results <- results$data...
head(results$data)

# how many points are there in each cluster?
table(results$data$cluster)

# plot clusters 3 and 4 as a timeVariation plot using SAME colours as in
# cluster plot
timeVariation(subset(results$data, cluster \%in\% c("3", "4")),
  pollutant = "nox",
  group = "cluster", col = openColours("Paired", 6)[c(3, 4)]
)
}

}
\references{
Carslaw, D.C., Beevers, S.D, Ropkins, K and M.C. Bell (2006). Detecting and
quantifying aircraft and other on-airport contributions to ambient nitrogen
oxides in the vicinity of a large international airport.  Atmospheric
Environment. 40/28 pp 5424-5434.

Carslaw, D.C., & Beevers, S.D. (2013). Characterising and understanding
emission sources using bivariate polar plots and k-means clustering.
Environmental Modelling & Software, 40, 325-329.
doi:10.1016/j.envsoft.2012.09.005
}
\seealso{
Other polar directional analysis functions:
\code{\link[=percentileRose]{percentileRose()}},
\code{\link[=polarAnnulus]{polarAnnulus()}},
\code{\link[=polarDiff]{polarDiff()}},
\code{\link[=polarFreq]{polarFreq()}},
\code{\link[=polarPlot]{polarPlot()}},
\code{\link[=pollutionRose]{pollutionRose()}},
\code{\link[=windRose]{windRose()}}

Other cluster analysis functions:
\code{\link[=timeProp]{timeProp()}},
\code{\link[=trajCluster]{trajCluster()}}
}
\author{
David Carslaw
}
\concept{cluster analysis functions}
\concept{polar directional analysis functions}
