% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarDiff.R
\name{polarDiff}
\alias{polarDiff}
\title{Polar plots considering changes in concentrations between two time periods}
\usage{
polarDiff(
  before,
  after,
  pollutant = "nox",
  x = "ws",
  wd = "wd",
  type = "default",
  limits = NULL,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{before, after}{Data frames representing the "before" and "after" cases.
See \code{\link[=polarPlot]{polarPlot()}} for details of different input requirements.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}. There can also be
more than one pollutant specified e.g. \code{pollutant = c("nox", "no2")}. The
main use of using two or more pollutants is for model evaluation where two
species would be expected to have similar concentrations. This saves the
user stacking the data and it is possible to work with columns of data
directly. A typical use would be \code{pollutant = c("obs", "mod")} to compare
two columns \dQuote{obs} (the observations) and \dQuote{mod} (modelled
values). When pair-wise statistics such as Pearson correlation and
regression techniques are to be plotted, \code{pollutant} takes two elements
too. For example, \code{pollutant = c("bc", "pm25")} where \code{"bc"} is a function
of \code{"pm25"}.}

\item{x}{Name of variable to plot against wind direction in polar
coordinates, the default is wind speed, \dQuote{ws}.}

\item{wd}{The name of the column in \code{mydata} representing the decimal wind
direction, 0 to 360 where 0/360 are North and 180 is South. Defaults to
\code{"wd"}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{limits}{The function does its best to choose sensible limits
automatically. However, there are circumstances when the user will wish to
set different ones. An example would be a series of plots showing each year
of data separately. The limits are set in the form \code{c(lower, upper)}, so
\code{limits = c(0, 100)} would force the plot limits to span 0-100.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{
  Arguments passed on to \code{\link[=polarPlot]{polarPlot}}
  \describe{
    \item{\code{statistic}}{The statistic that should be applied to each wind
speed/direction bin. Because of the smoothing involved, the colour scale
for some of these statistics is only to provide an indication of overall
pattern and should not be interpreted in concentration units e.g. for
\code{statistic = "weighted.mean"} where the bin mean is multiplied by the bin
frequency and divided by the total frequency. In many cases using
\code{polarFreq} will be better. Setting \code{statistic = "weighted.mean"} can be
useful because it provides an indication of the concentration * frequency
of occurrence and will highlight the wind speed/direction conditions that
dominate the overall mean.Can be:

\itemize{ \item  \dQuote{mean} (default), \dQuote{median}, \dQuote{max}
(maximum), \dQuote{frequency}. \dQuote{stdev} (standard deviation),
\dQuote{weighted.mean}.

\item \code{statistic = "nwr"} Implements the Non-parametric Wind
Regression approach of Henry et al. (2009) that uses kernel smoothers. The
\code{openair} implementation is not identical because Gaussian kernels are
used for both wind direction and speed. The smoothing is controlled by
\code{ws_spread} and \code{wd_spread}.

\item \code{statistic = "cpf"} the conditional probability function (CPF)
is plotted and a single (usually high) percentile level is supplied. The
CPF is defined as CPF = my/ny, where my is the number of samples in the y
bin (by default a wind direction, wind speed interval) with mixing ratios
greater than the \emph{overall} percentile concentration, and ny is the
total number of samples in the same wind sector (see Ashbaugh et al.,
1985). Note that percentile intervals can also be considered; see
\code{percentile} for details.

\item When \code{statistic = "r"} or \code{statistic = "Pearson"}, the
Pearson correlation coefficient is calculated for \emph{two} pollutants.
The calculation involves a weighted Pearson correlation coefficient, which
is weighted by Gaussian kernels for wind direction an the radial variable
(by default wind speed). More weight is assigned to values close to a wind
speed-direction interval. Kernel weighting is used to ensure that all data
are used rather than relying on the potentially small number of values in a
wind speed-direction interval.

\item When \code{statistic = "Spearman"}, the Spearman correlation
coefficient is calculated for \emph{two} pollutants. The calculation
involves a weighted Spearman correlation coefficient, which is weighted by
Gaussian kernels for wind direction an the radial variable (by default wind
speed). More weight is assigned to values close to a wind speed-direction
interval. Kernel weighting is used to ensure that all data are used rather
than relying on the potentially small number of values in a wind
speed-direction interval.

\item \code{"robust_slope"} is another option for pair-wise statistics and
\code{"quantile.slope"}, which uses quantile regression to estimate the
slope for a particular quantile level (see also \code{tau} for setting the
quantile level).

\item \code{"york_slope"} is another option for pair-wise statistics which
uses the \emph{York regression method} to estimate the slope. In this
method the uncertainties in \code{x} and \code{y} are used in the
determination of the slope. The uncertainties are provided by
\code{x_error} and \code{y_error} --- see below.}}
    \item{\code{exclude.missing}}{Setting this option to \code{TRUE} (the default) removes
points from the plot that are too far from the original data. The smoothing
routines will produce predictions at points where no data exist i.e. they
predict. By removing the points too far from the original data produces a
plot where it is clear where the original data lie. If set to \code{FALSE}
missing data will be interpolated.}
    \item{\code{uncertainty}}{Should the uncertainty in the calculated surface be shown?
If \code{TRUE} three plots are produced on the same scale showing the predicted
surface together with the estimated lower and upper uncertainties at the
95\% confidence interval. Calculating the uncertainties is useful to
understand whether features are real or not.  For example, at high wind
speeds where there are few data there is greater uncertainty over the
predicted values. The uncertainties are calculated using the GAM and
weighting is done by the frequency of measurements in each wind
speed-direction bin. Note that if uncertainties are calculated then the
type is set to "default".}
    \item{\code{percentile}}{If \code{statistic = "percentile"} then \code{percentile} is used,
expressed from 0 to 100. Note that the percentile value is calculated in
the wind speed, wind direction \sQuote{bins}. For this reason it can also
be useful to set \code{min.bin} to ensure there are a sufficient number of
points available to estimate a percentile. See \code{quantile} for more details
of how percentiles are calculated.

\code{percentile} is also used for the Conditional Probability Function (CPF)
plots. \code{percentile} can be of length two, in which case the percentile
\emph{interval} is considered for use with CPF. For example, \code{percentile = c(90, 100)} will plot the CPF for concentrations between the 90 and 100th
percentiles. Percentile intervals can be useful for identifying specific
sources. In addition, \code{percentile} can also be of length 3. The third value
is the \sQuote{trim} value to be applied. When calculating percentile
intervals many can cover very low values where there is no useful
information. The trim value ensures that values greater than or equal to
the trim * mean value are considered \emph{before} the percentile intervals are
calculated. The effect is to extract more detail from many source
signatures. See the manual for examples. Finally, if the trim value is less
than zero the percentile range is interpreted as absolute concentration
values and subsetting is carried out directly.}
    \item{\code{weights}}{At the edges of the plot there may only be a few data points
in each wind speed-direction interval, which could in some situations
distort the plot if the concentrations are high. \code{weights} applies a
weighting to reduce their influence. For example and by default if only a
single data point exists then the weighting factor is 0.25 and for two
points 0.5. To not apply any weighting and use the data as is, use \code{weights = c(1, 1, 1)}.

An alternative to down-weighting these points they can be removed
altogether using \code{min.bin}.}
    \item{\code{min.bin}}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points. It is recommended to consider your data with
care. Also, the \code{polarFreq} function can be of use in such circumstances.}
    \item{\code{col.na}}{When \code{min.bin} is > 1 it can be useful to show where data are
removed on the plots. This is done by shading the missing data in \code{col.na}.
To not highlight missing data when \code{min.bin} > 1 choose \code{col.na = "transparent"}.}
    \item{\code{upper}}{This sets the upper limit wind speed to be used. Often there are
only a relatively few data points at very high wind speeds and plotting all
of them can reduce the useful information in the plot.}
    \item{\code{units}}{The units shown on the polar axis scale.}
    \item{\code{force.positive}}{The default is \code{TRUE}. Sometimes if smoothing data with
steep gradients it is possible for predicted values to be negative.
\code{force.positive = TRUE} ensures that predictions remain positive. This is
useful for several reasons. First, with lots of missing data more
interpolation is needed and this can result in artefacts because the
predictions are too far from the original data. Second, if it is known
beforehand that the data are all positive, then this option carries that
assumption through to the prediction. The only likely time where setting
\code{force.positive = FALSE} would be if background concentrations were first
subtracted resulting in data that is legitimately negative. For the vast
majority of situations it is expected that the user will not need to alter
the default option.}
    \item{\code{k}}{This is the smoothing parameter used by the \code{gam} function in
package \code{mgcv}. Typically, value of around 100 (the default) seems to be
suitable and will resolve important features in the plot. The most
appropriate choice of \code{k} is problem-dependent; but extensive testing of
polar plots for many different problems suggests a value of \code{k} of about
100 is suitable. Setting \code{k} to higher values will not tend to affect the
surface predictions by much but will add to the computation time. Lower
values of \code{k} will increase smoothing. Sometimes with few data to plot
\code{polarPlot} will fail. Under these circumstances it can be worth lowering
the value of \code{k}.}
    \item{\code{normalise}}{If \code{TRUE} concentrations are normalised by dividing by their
mean value. This is done \emph{after} fitting the smooth surface. This option is
particularly useful if one is interested in the patterns of concentrations
for several pollutants on different scales e.g. NOx and CO. Often useful if
more than one \code{pollutant} is chosen.}
    \item{\code{ws_spread}}{The value of sigma used for Gaussian kernel weighting of
wind speed when \code{statistic = "nwr"} or when correlation and regression
statistics are used such as \emph{r}. Default is \code{0.5}.}
    \item{\code{wd_spread}}{The value of sigma used for Gaussian kernel weighting of
wind direction when \code{statistic = "nwr"} or when correlation and regression
statistics are used such as \emph{r}. Default is \code{4}.}
    \item{\code{x_error}}{The \code{x} error / uncertainty used when \code{statistic = "york_slope"}.}
    \item{\code{y_error}}{The \code{y} error / uncertainty used when \code{statistic = "york_slope"}.}
    \item{\code{kernel}}{Type of kernel used for the weighting procedure for when
correlation or regression techniques are used. Only \code{"gaussian"} is
supported but this may be enhanced in the future.}
    \item{\code{formula.label}}{When pair-wise statistics such as regression slopes are
calculated and plotted, should a formula label be displayed?
\code{formula.label} will also determine whether concentration information is
printed when \code{statistic = "cpf"}.}
    \item{\code{tau}}{The quantile to be estimated when \code{statistic} is set to
\code{"quantile.slope"}. Default is \code{0.5} which is equal to the median and will
be ignored if \code{"quantile.slope"} is not used.}
    \item{\code{cols}}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}
    \item{\code{breaks,labels}}{If a categorical colour scale is required, \code{breaks}
should be specified. This can be either of:
\itemize{
\item A single value, which will divide the scale into \code{breaks} levels using the
same logic as \code{\link[=cutData]{cutData()}}. For example, \code{breaks = 5} will split the scale
into five quantiles.
\item A numeric vector, which will define the specific breakpoints. For
example, \code{c(0, 50, 100)} will bin the data into \verb{0 to 50}, \verb{50 to 100}, and
so on. If \code{breaks} does not cover the full range of the data, the outer
limits will be extended so that the full colour scale is covered while
retaining the desired number of breaks.
}

By default, \code{breaks} will generate nicely formatted labels for each
category. The \code{labels} argument overrides this - for example, a user could
define \verb{breaks = 3, labels = c("low", "medium", "high")}. Care should be
taken to provide the appropriate number of \code{labels} - it should be equal to
\code{breaks} if a single value is given, or equal to \code{length(breaks)-1} if
\code{breaks} is a vector.}
    \item{\code{angle.scale}}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}
    \item{\code{key.position}}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}
    \item{\code{key.title}}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}
    \item{\code{key}}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}
  }}
}
\value{
an \link[=openair-package]{openair} plot.
}
\description{
This function provides a way of showing the differences in concentrations
between two time periods as a polar plot. There are several uses of this
function, but the most common will be to see how source(s) may have changed
between two periods.
}
\details{
While the function is primarily intended to compare two time periods at the
same location, it can be used for any two data sets that contain the same
pollutant. For example, data from two sites that are close to one another, or
two co-located instruments.

The analysis works by calculating the polar plot surface for the \code{before} and
\code{after} periods and then subtracting the \code{before} surface from the \code{after}
surface.
}
\examples{
\dontrun{
before_data <- selectByDate(mydata, year = 2002)
after_data <- selectByDate(mydata, year = 2003)

polarDiff(before_data, after_data, pollutant = "no2")

# with some options
polarDiff(
  before_data,
  after_data,
  pollutant = "no2",
  cols = "RdYlBu",
  limits = c(-20, 20)
)
}
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarAnnulus}()},
\code{\link{polarCluster}()},
\code{\link{polarFreq}()},
\code{\link{polarPlot}()},
\code{\link{pollutionRose}()},
\code{\link{windRose}()}
}
\concept{polar directional analysis functions}
