% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarFreq.R
\name{polarFreq}
\alias{polarFreq}
\title{Function to plot wind speed/direction frequencies and other statistics}
\usage{
polarFreq(
  mydata,
  pollutant = NULL,
  ws = "ws",
  wd = "wd",
  statistic = "frequency",
  ws.int = 1,
  wd.nint = 36,
  grid.line = 5,
  limits = NULL,
  breaks = NULL,
  labels = NULL,
  cols = "default",
  trans = TRUE,
  type = "default",
  min.bin = 1,
  ws.upper = NA,
  angle.scale = 45,
  offset = 10,
  border.col = "transparent",
  key.title = paste(statistic, pollutant, sep = " "),
  key.position = "right",
  strip.position = "top",
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame minimally containing a wind speed, a decimal wind
direction, and \code{date}.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}}

\item{ws}{The name of the column in \code{mydata} representing the wind speed.
Defaults to \code{"ws"}.}

\item{wd}{The name of the column in \code{mydata} representing the decimal wind
direction, 0 to 360 where 0/360 are North and 180 is South. Defaults to
\code{"wd"}.}

\item{statistic}{The statistic that should be applied to each wind
speed/direction bin. Can be one of:
\itemize{
\item \code{"frequency"}: the simplest and plots the frequency of wind speed/direction in
different bins. The scale therefore shows the counts in each bin.
\item \code{"mean"}, \code{"median"}, \code{"max"} (maximum), \code{"stdev"} (standard deviation):
Plots the relevant summary statistic of a pollutant in wind speed/direction
bins.
\item \code{"weighted.mean"} will plot the concentration of a pollutant
weighted by wind speed/direction. Each segment therefore provides the
percentage overall contribution to the total concentration.
}

Note that for options other than \code{"frequency"}, it is necessary to also
provide the name of a \code{pollutant}.}

\item{ws.int}{Wind speed interval assumed. In some cases e.g. a low met mast,
an interval of 0.5 may be more appropriate.}

\item{wd.nint}{Number of intervals of wind direction.}

\item{grid.line}{Radial spacing of grid lines.}

\item{limits}{The limits of the colour bar (e.g., \code{c(0, 100)}).}

\item{breaks, labels}{If a categorical colour scale is required then \code{breaks}
should be specified. These should be provided as a numeric vector, e.g.,
\code{breaks = c(0, 50, 100, 1000)}. Users should set the maximum value of
\code{breaks} to exceed the maximum data value to ensure it is within the
maximum final range, e.g., 100--1000 in this case. Labels will
automatically be generated, but can be customised by passing a character
vector to \code{labels}, e.g., \code{labels = c("good", "bad", "very bad")}. In this
example, \code{0 - 50} will be \code{"good"} and so on. Note there is one less label
than break.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{trans}{Should a transformation be applied? Sometimes when producing
plots of this kind they can be dominated by a few high points. The default
therefore is \code{TRUE} and a square-root transform is applied. This results in
a non-linear scale and (usually) a better representation of the
distribution. If set to \code{FALSE} a linear scale is used.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{min.bin}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points. It is recommended to consider your data with
care. Also, the \code{polarFreq} function can be of use in such circumstances.}

\item{ws.upper}{A user-defined upper wind speed to use. This is useful for
ensuring a consistent scale between different plots. For example, to always
ensure that wind speeds are displayed between 1-10, set \code{ws.int = 10}.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{offset}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}

\item{border.col}{The colour of the boundary of each wind speed/direction
bin. The default is transparent. Another useful choice sometimes is
"white".}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label. All of these
are passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
\code{polarFreq} primarily plots wind speed-direction frequencies in
\sQuote{bins}. Each bin is colour-coded depending on the frequency of
measurements. Bins can also be used to show the concentration of pollutants
using a range of commonly used statistics.
}
\details{
\code{polarFreq} is its default use provides details of wind speed and direction
frequencies. In this respect it is similar to \code{\link[=windRose]{windRose()}}, but considers
wind direction intervals of 10 degrees and a user-specified wind speed
interval. The frequency of wind speeds/directions formed by these
\sQuote{bins} is represented on a colour scale.

The \code{polarFreq} function is more flexible than either \code{\link[=windRose]{windRose()}} or
\code{\link[=polarPlot]{polarPlot()}}. It can, for example, also consider pollutant concentrations
(see examples below). Instead of the number of data points in each bin, the
concentration can be shown. Further, a range of statistics can be used to
describe each bin - see \code{statistic} above. Plotting mean concentrations is
useful for source identification and is the same as \code{\link[=polarPlot]{polarPlot()}} but without
smoothing, which may be preferable for some data. Plotting with \code{statistic = "weighted.mean"} is particularly useful for understanding the relative
importance of different source contributions. For example, high mean
concentrations may be observed for high wind speed conditions, but the
weighted mean concentration may well show that the contribution to overall
concentrations is very low.

\code{polarFreq} also offers great flexibility with the scale used and the user
has fine control over both the range, interval and colour.
}
\examples{
# basic wind frequency plot
polarFreq(mydata)

# wind frequencies by year
\dontrun{
polarFreq(mydata, type = "year")
}


# mean SO2 by year, showing only bins with at least 2 points
\dontrun{
polarFreq(mydata, pollutant = "so2", type = "year", statistic = "mean", min.bin = 2)
}

# weighted mean SO2 by year, showing only bins with at least 2 points
\dontrun{
polarFreq(mydata,
  pollutant = "so2", type = "year", statistic = "weighted.mean",
  min.bin = 2
)
}

# windRose for just 2000 and 2003 with different colours
\dontrun{
polarFreq(subset(mydata, format(date, "\%Y") \%in\% c(2000, 2003)),
  type = "year", cols = "turbo"
)
}

# user defined breaks from 0-700 in intervals of 100 (note linear scale)
\dontrun{
polarFreq(mydata, breaks = seq(0, 700, 100))
}

# more complicated user-defined breaks - useful for highlighting bins
# with a certain number of data points
\dontrun{
polarFreq(mydata, breaks = c(0, 10, 50, 100, 250, 500, 700))
}

# source contribution plot and use of offset option
\dontrun{
polarFreq(mydata,
  pollutant = "pm25",
  statistic = "weighted.mean", offset = 50, ws.int = 25, trans = FALSE
)
}
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarAnnulus}()},
\code{\link{polarCluster}()},
\code{\link{polarDiff}()},
\code{\link{polarPlot}()},
\code{\link{pollutionRose}()},
\code{\link{windRose}()}
}
\author{
David Carslaw
}
\concept{polar directional analysis functions}
