% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarPlot.R
\name{polarPlot}
\alias{polarPlot}
\title{Function for plotting bivariate polar plots with smoothing.}
\usage{
polarPlot(
  mydata,
  pollutant = "nox",
  x = "ws",
  wd = "wd",
  type = "default",
  statistic = "mean",
  limits = NULL,
  exclude.missing = TRUE,
  uncertainty = FALSE,
  percentile = NA,
  cols = "default",
  weights = c(0.25, 0.5, 0.75),
  min.bin = 1,
  mis.col = "grey",
  upper = NA,
  angle.scale = 315,
  units = x,
  force.positive = TRUE,
  k = 100,
  normalise = FALSE,
  key = TRUE,
  key.title = paste(statistic, pollutant, sep = "\\n"),
  key.position = "right",
  strip.position = "top",
  auto.text = TRUE,
  ws_spread = 1.5,
  wd_spread = 5,
  x_error = NA,
  y_error = NA,
  kernel = "gaussian",
  formula.label = TRUE,
  tau = 0.5,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame minimally containing \code{wd}, another variable to
plot in polar coordinates (the default is a column \dQuote{ws} --- wind
speed) and a pollutant. Should also contain \code{date} if plots by time period
are required.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}. There can also be
more than one pollutant specified e.g. \code{pollutant = c("nox", "no2")}. The
main use of using two or more pollutants is for model evaluation where two
species would be expected to have similar concentrations. This saves the
user stacking the data and it is possible to work with columns of data
directly. A typical use would be \code{pollutant = c("obs", "mod")} to compare
two columns \dQuote{obs} (the observations) and \dQuote{mod} (modelled
values). When pair-wise statistics such as Pearson correlation and
regression techniques are to be plotted, \code{pollutant} takes two elements
too. For example, \code{pollutant = c("bc", "pm25")} where \code{"bc"} is a function
of \code{"pm25"}.}

\item{x}{Name of variable to plot against wind direction in polar
coordinates, the default is wind speed, \dQuote{ws}.}

\item{wd}{Name of wind direction field.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{statistic}{The statistic that should be applied to each wind
speed/direction bin. Because of the smoothing involved, the colour scale
for some of these statistics is only to provide an indication of overall
pattern and should not be interpreted in concentration units e.g. for
\code{statistic = "weighted.mean"} where the bin mean is multiplied by the bin
frequency and divided by the total frequency. In many cases using
\code{polarFreq} will be better. Setting \code{statistic = "weighted.mean"} can be
useful because it provides an indication of the concentration * frequency
of occurrence and will highlight the wind speed/direction conditions that
dominate the overall mean.Can be:

\itemize{ \item  \dQuote{mean} (default), \dQuote{median}, \dQuote{max}
(maximum), \dQuote{frequency}. \dQuote{stdev} (standard deviation),
\dQuote{weighted.mean}.

\item \code{statistic = "nwr"} Implements the Non-parametric Wind
Regression approach of Henry et al. (2009) that uses kernel smoothers. The
\code{openair} implementation is not identical because Gaussian kernels are
used for both wind direction and speed. The smoothing is controlled by
\code{ws_spread} and \code{wd_spread}.

\item \code{statistic = "cpf"} the conditional probability function (CPF)
is plotted and a single (usually high) percentile level is supplied. The
CPF is defined as CPF = my/ny, where my is the number of samples in the y
bin (by default a wind direction, wind speed interval) with mixing ratios
greater than the \emph{overall} percentile concentration, and ny is the
total number of samples in the same wind sector (see Ashbaugh et al.,
1985). Note that percentile intervals can also be considered; see
\code{percentile} for details.

\item When \code{statistic = "r"} or \code{statistic = "Pearson"}, the
Pearson correlation coefficient is calculated for \emph{two} pollutants.
The calculation involves a weighted Pearson correlation coefficient, which
is weighted by Gaussian kernels for wind direction an the radial variable
(by default wind speed). More weight is assigned to values close to a wind
speed-direction interval. Kernel weighting is used to ensure that all data
are used rather than relying on the potentially small number of values in a
wind speed-direction interval.

\item When \code{statistic = "Spearman"}, the Spearman correlation
coefficient is calculated for \emph{two} pollutants. The calculation
involves a weighted Spearman correlation coefficient, which is weighted by
Gaussian kernels for wind direction an the radial variable (by default wind
speed). More weight is assigned to values close to a wind speed-direction
interval. Kernel weighting is used to ensure that all data are used rather
than relying on the potentially small number of values in a wind
speed-direction interval.

\item \code{"robust_slope"} is another option for pair-wise statistics and
\code{"quantile.slope"}, which uses quantile regression to estimate the
slope for a particular quantile level (see also \code{tau} for setting the
quantile level).

\item \code{"york_slope"} is another option for pair-wise statistics which
uses the \emph{York regression method} to estimate the slope. In this
method the uncertainties in \code{x} and \code{y} are used in the
determination of the slope. The uncertainties are provided by
\code{x_error} and \code{y_error} --- see below.}}

\item{limits}{The function does its best to choose sensible limits
automatically. However, there are circumstances when the user will wish to
set different ones. An example would be a series of plots showing each year
of data separately. The limits are set in the form \code{c(lower, upper)}, so
\code{limits = c(0, 100)} would force the plot limits to span 0-100.}

\item{exclude.missing}{Setting this option to \code{TRUE} (the default) removes
points from the plot that are too far from the original data. The smoothing
routines will produce predictions at points where no data exist i.e. they
predict. By removing the points too far from the original data produces a
plot where it is clear where the original data lie. If set to \code{FALSE}
missing data will be interpolated.}

\item{uncertainty}{Should the uncertainty in the calculated surface be shown?
If \code{TRUE} three plots are produced on the same scale showing the predicted
surface together with the estimated lower and upper uncertainties at the
95\% confidence interval. Calculating the uncertainties is useful to
understand whether features are real or not.  For example, at high wind
speeds where there are few data there is greater uncertainty over the
predicted values. The uncertainties are calculated using the GAM and
weighting is done by the frequency of measurements in each wind
speed-direction bin. Note that if uncertainties are calculated then the
type is set to "default".}

\item{percentile}{If \code{statistic = "percentile"} then \code{percentile} is used,
expressed from 0 to 100. Note that the percentile value is calculated in
the wind speed, wind direction \sQuote{bins}. For this reason it can also
be useful to set \code{min.bin} to ensure there are a sufficient number of
points available to estimate a percentile. See \code{quantile} for more details
of how percentiles are calculated.

\code{percentile} is also used for the Conditional Probability Function (CPF)
plots. \code{percentile} can be of length two, in which case the percentile
\emph{interval} is considered for use with CPF. For example, \code{percentile = c(90, 100)} will plot the CPF for concentrations between the 90 and 100th
percentiles. Percentile intervals can be useful for identifying specific
sources. In addition, \code{percentile} can also be of length 3. The third value
is the \sQuote{trim} value to be applied. When calculating percentile
intervals many can cover very low values where there is no useful
information. The trim value ensures that values greater than or equal to
the trim * mean value are considered \emph{before} the percentile intervals are
calculated. The effect is to extract more detail from many source
signatures. See the manual for examples. Finally, if the trim value is less
than zero the percentile range is interpreted as absolute concentration
values and subsetting is carried out directly.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{weights}{At the edges of the plot there may only be a few data points
in each wind speed-direction interval, which could in some situations
distort the plot if the concentrations are high. \code{weights} applies a
weighting to reduce their influence. For example and by default if only a
single data point exists then the weighting factor is 0.25 and for two
points 0.5. To not apply any weighting and use the data as is, use \code{weights = c(1, 1, 1)}.

An alternative to down-weighting these points they can be removed
altogether using \code{min.bin}.}

\item{min.bin}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points. It is recommended to consider your data with
care. Also, the \code{polarFreq} function can be of use in such circumstances.}

\item{mis.col}{When \code{min.bin} is > 1 it can be useful to show where data are
removed on the plots. This is done by shading the missing data in
\code{mis.col}. To not highlight missing data when \code{min.bin} > 1 choose \code{mis.col = "transparent"}.}

\item{upper}{This sets the upper limit wind speed to be used. Often there are
only a relatively few data points at very high wind speeds and plotting all
of them can reduce the useful information in the plot.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{units}{The units shown on the polar axis scale.}

\item{force.positive}{The default is \code{TRUE}. Sometimes if smoothing data with
steep gradients it is possible for predicted values to be negative.
\code{force.positive = TRUE} ensures that predictions remain positive. This is
useful for several reasons. First, with lots of missing data more
interpolation is needed and this can result in artefacts because the
predictions are too far from the original data. Second, if it is known
beforehand that the data are all positive, then this option carries that
assumption through to the prediction. The only likely time where setting
\code{force.positive = FALSE} would be if background concentrations were first
subtracted resulting in data that is legitimately negative. For the vast
majority of situations it is expected that the user will not need to alter
the default option.}

\item{k}{This is the smoothing parameter used by the \code{gam} function in
package \code{mgcv}. Typically, value of around 100 (the default) seems to be
suitable and will resolve important features in the plot. The most
appropriate choice of \code{k} is problem-dependent; but extensive testing of
polar plots for many different problems suggests a value of \code{k} of about
100 is suitable. Setting \code{k} to higher values will not tend to affect the
surface predictions by much but will add to the computation time. Lower
values of \code{k} will increase smoothing. Sometimes with few data to plot
\code{polarPlot} will fail. Under these circumstances it can be worth lowering
the value of \code{k}.}

\item{normalise}{If \code{TRUE} concentrations are normalised by dividing by their
mean value. This is done \emph{after} fitting the smooth surface. This option is
particularly useful if one is interested in the patterns of concentrations
for several pollutants on different scales e.g. NOx and CO. Often useful if
more than one \code{pollutant} is chosen.}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{ws_spread}{The value of sigma used for Gaussian kernel weighting of
wind speed when \code{statistic = "nwr"} or when correlation and regression
statistics are used such as \emph{r}. Default is \code{0.5}.}

\item{wd_spread}{The value of sigma used for Gaussian kernel weighting of
wind direction when \code{statistic = "nwr"} or when correlation and regression
statistics are used such as \emph{r}. Default is \code{4}.}

\item{x_error}{The \code{x} error / uncertainty used when \code{statistic = "york_slope"}.}

\item{y_error}{The \code{y} error / uncertainty used when \code{statistic = "york_slope"}.}

\item{kernel}{Type of kernel used for the weighting procedure for when
correlation or regression techniques are used. Only \code{"gaussian"} is
supported but this may be enhanced in the future.}

\item{formula.label}{When pair-wise statistics such as regression slopes are
calculated and plotted, should a formula label be displayed?
\code{formula.label} will also determine whether concentration information is
printed when \code{statistic = "cpf"}.}

\item{tau}{The quantile to be estimated when \code{statistic} is set to
\code{"quantile.slope"}. Default is \code{0.5} which is equal to the median and will
be ignored if \code{"quantile.slope"} is not used.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{annotate = FALSE} will not plot the N/E/S/W labels.
}}
}
\value{
an \link[=openair-package]{openair} object. \code{data} contains four set
columns: \code{cond}, conditioning based on \code{type}; \code{u} and \code{v}, the
translational vectors based on \code{ws} and \code{wd}; and the local \code{pollutant}
estimate.
}
\description{
Function for plotting pollutant concentration in polar coordinates showing
concentration by wind speed (or another numeric variable) and direction. Mean
concentrations are calculated for wind speed-direction \sQuote{bins} (e.g.
0-1, 1-2 m/s,...  and 0-10, 10-20 degrees etc.).  To aid interpretation,
\code{gam} smoothing is carried out using \code{mgcv}.
}
\details{
The bivariate polar plot is a useful diagnostic tool for quickly gaining an
idea of potential sources. Wind speed is one of the most useful variables to
use to separate source types (see references). For example, ground-level
concentrations resulting from buoyant plumes from chimney stacks tend to peak
under higher wind speed conditions. Conversely, ground-level, non-buoyant
plumes such as from road traffic, tend to have highest concentrations under
low wind speed conditions. Other sources such as from aircraft engines also
show differing characteristics by wind speed.

The function has been developed to allow variables other than wind speed to
be plotted with wind direction in polar coordinates. The key issue is that
the other variable plotted against wind direction should be discriminating in
some way. For example, temperature can help reveal high-level sources brought
down to ground level in unstable atmospheric conditions, or show the effect a
source emission dependent on temperature e.g. biogenic isoprene.

The plots can vary considerably depending on how much smoothing is done.  The
approach adopted here is based on the very flexible and capable \code{mgcv}
package that uses \emph{Generalized Additive Models}. While methods do exist to
find an optimum level of smoothness, they are not necessarily useful. The
principal aim of \code{polarPlot} is as a graphical analysis rather than for
quantitative purposes. In this respect the smoothing aims to strike a balance
between revealing interesting (real) features and overly noisy data. The
defaults used in \code{\link[=polarPlot]{polarPlot()}} are based on the analysis of data from many
different sources. More advanced users may wish to modify the code and adopt
other smoothing approaches.

Various statistics are possible to consider e.g. mean, maximum, median.
\code{statistic = "max"} is often useful for revealing sources. Pair-wise
statistics between two pollutants can also be calculated.

The function can also be used to compare two pollutant species through a
range of pair-wise statistics (see help on \code{statistic}) and Grange et al.
(2016) (open-access publication link below).

Wind direction is split up into 10 degree intervals and the other variable
(e.g. wind speed) 30 intervals. These 2D bins are then used to calculate the
statistics.

These plots often show interesting features at higher wind speeds (see
references below). For these conditions there can be very few measurements
and therefore greater uncertainty in the calculation of the surface. There
are several ways in which this issue can be tackled. First, it is possible to
avoid smoothing altogether and use \code{\link[=polarFreq]{polarFreq()}}. Second, the effect of
setting a minimum number of measurements in each wind speed-direction bin can
be examined through \code{min.bin}. It is possible that a single point at high
wind speed conditions can strongly affect the surface prediction. Therefore,
setting \code{min.bin = 3}, for example, will remove all wind speed-direction bins
with fewer than 3 measurements \emph{before} fitting the surface. Third, consider
setting \code{uncertainty = TRUE}. This option will show the predicted surface
together with upper and lower 95\% confidence intervals, which take account of
the frequency of measurements.
}
\examples{

# basic plot
polarPlot(mydata, pollutant = "nox")
\dontrun{

# polarPlots by year on same scale
polarPlot(mydata, pollutant = "so2", type = "year", main = "polarPlot of so2")

# set minimum number of bins to be used to see if pattern remains similar
polarPlot(mydata, pollutant = "nox", min.bin = 3)

# plot by day of the week
polarPlot(mydata, pollutant = "pm10", type = "weekday")

# show the 95\% confidence intervals in the surface fitting
polarPlot(mydata, pollutant = "so2", uncertainty = TRUE)


# Pair-wise statistics
# Pearson correlation
polarPlot(mydata, pollutant = c("pm25", "pm10"), statistic = "r")

# Robust regression slope, takes a bit of time
polarPlot(mydata, pollutant = c("pm25", "pm10"), statistic = "robust.slope")

# Least squares regression works too but it is not recommended, use robust
# regression
# polarPlot(mydata, pollutant = c("pm25", "pm10"), statistic = "slope")
}

}
\references{
Ashbaugh, L.L., Malm, W.C., Sadeh, W.Z., 1985. A residence time probability
analysis of sulfur concentrations at ground canyon national park. Atmospheric
Environment 19 (8), 1263-1270.

Carslaw, D.C., Beevers, S.D, Ropkins, K and M.C. Bell (2006). Detecting and
quantifying aircraft and other on-airport contributions to ambient nitrogen
oxides in the vicinity of a large international airport.  Atmospheric
Environment. 40/28 pp 5424-5434.

Carslaw, D.C., & Beevers, S.D. (2013). Characterising and understanding
emission sources using bivariate polar plots and k-means clustering.
Environmental Modelling & Software, 40, 325-329. DOI:
10.1016/j.envsoft.2012.09.005.

Henry, R.C., Chang, Y.S., Spiegelman, C.H., 2002. Locating nearby sources of
air pollution by nonparametric regression of atmospheric concentrations on
wind direction. Atmospheric Environment 36 (13), 2237-2244.

Henry, R., Norris, G.A., Vedantham, R., Turner, J.R., 2009. Source region
identification using Kernel smoothing. Environ. Sci. Technol. 43 (11),
4090e4097. DOI: 10.1021/es8011723.

Uria-Tellaetxe, I. and D.C. Carslaw (2014). Source identification using a
conditional bivariate Probability function. Environmental Modelling &
Software, Vol. 59, 1-9.

Westmoreland, E.J., N. Carslaw, D.C. Carslaw, A. Gillah and E. Bates (2007).
Analysis of air quality within a street canyon using statistical and
dispersion modelling techniques. Atmospheric Environment. Vol.  41(39), pp.
9195-9205.

Yu, K.N., Cheung, Y.P., Cheung, T., Henry, R.C., 2004. Identifying the impact
of large urban airports on local air quality by nonparametric regression.
Atmospheric Environment 38 (27), 4501-4507.

Grange, S. K., Carslaw, D. C., & Lewis, A. C. 2016. Source apportionment
advances with bivariate polar plots, correlation, and regression techniques.
Atmospheric Environment. 145, 128-134. DOI: 10.1016/j.atmosenv.2016.09.016.
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarAnnulus}()},
\code{\link{polarCluster}()},
\code{\link{polarDiff}()},
\code{\link{polarFreq}()},
\code{\link{pollutionRose}()},
\code{\link{windRose}()}
}
\author{
David Carslaw
}
\concept{polar directional analysis functions}
