% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windRose.R
\name{pollutionRose}
\alias{pollutionRose}
\title{Pollution rose variation of the traditional wind rose plot}
\usage{
pollutionRose(
  mydata,
  pollutant = "nox",
  key.title = pollutant,
  key.position = "right",
  breaks = 6,
  paddle = FALSE,
  seg = 0.9,
  normalise = FALSE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing fields \code{ws} and \code{wd}}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{breaks}{Most commonly, the number of break points for pollutant
concentrations. The default, 6, attempts to breaks the supplied data at
approximately 6 sensible break points. However, \code{breaks} can also be
used to set specific break points. For example, the argument \code{breaks = c(0, 1, 10, 100)} breaks the data into segments <1, 1-10, 10-100, >100.}

\item{paddle}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE} plots rose using 'paddle'
style spokes. If \code{FALSE} plots rose using 'wedge' style spokes.}

\item{seg}{\code{seg} determines with width of the segments. For example, \code{seg = 0.5} will produce segments 0.5 * \code{angle}.}

\item{normalise}{If \code{TRUE} each wind direction segment is normalised to equal
one. This is useful for showing how the concentrations (or other
parameters) contribute to each wind sector when the proportion of time the
wind is from that direction is low. A line showing the probability that the
wind directions is from a particular wind sector is also shown.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Other arguments passed on to \code{\link[=windRose]{windRose()}}.}
}
\value{
an \link[=openair-package]{openair} object. Summarised proportions can be
extracted directly using the \verb{$data} operator, e.g.
\code{object$data} for \code{output <- windRose(mydata)}. This returns a
data frame with three set columns: \code{cond}, conditioning based on
\code{type}; \code{wd}, the wind direction; and \code{calm}, the
\code{statistic} for the proportion of data unattributed to any specific
wind direction because it was collected under calm conditions; and then
several (one for each range binned for the plot) columns giving proportions
of measurements associated with each \code{ws} or \code{pollutant} range
plotted as a discrete panel.
}
\description{
The traditional wind rose plot that plots wind speed and wind direction by
different intervals. The pollution rose applies the same plot structure but
substitutes other measurements, most commonly a pollutant time series, for
wind speed.
}
\details{
\code{\link[=pollutionRose]{pollutionRose()}} is a \code{\link[=windRose]{windRose()}} wrapper which brings \code{pollutant}
forward in the argument list, and attempts to sensibly rescale break points
based on the \code{pollutant} data range by by-passing \code{ws.int}.

By default, \code{\link[=pollutionRose]{pollutionRose()}} will plot a pollution rose of \code{nox} using
"wedge" style segments and placing the scale key to the right of the plot.

It is possible to compare two wind speed-direction data sets using
\code{\link[=pollutionRose]{pollutionRose()}}. There are many reasons for doing so e.g. to see how one
site compares with another or for meteorological model evaluation. In this
case, \code{ws} and \code{wd} are considered to the the reference data sets
with which a second set of wind speed and wind directions are to be compared
(\code{ws2} and \code{wd2}). The first set of values is subtracted from the
second and the differences compared. If for example, \code{wd2} was biased
positive compared with \code{wd} then \code{pollutionRose} will show the bias
in polar coordinates. In its default use, wind direction bias is colour-coded
to show negative bias in one colour and positive bias in another.
}
\examples{
# pollutionRose of nox
pollutionRose(mydata, pollutant = "nox")

# source apportionment plot - contribution to mean
\dontrun{
pollutionRose(mydata, pollutant = "pm10", type = "year", statistic = "prop.mean")

# example of comparing 2 met sites
# first we will make some new ws/wd data with a postive bias
mydata$ws2 <- mydata$ws + 2 * rnorm(nrow(mydata)) + 1
mydata$wd2 <- mydata$wd + 30 * rnorm(nrow(mydata)) + 30

# need to correct negative wd
id <- which(mydata$wd2 < 0)
mydata$wd2[id] <- mydata$wd2[id] + 360

# results show postive bias in wd and ws
pollutionRose(mydata, ws = "ws", wd = "wd", ws2 = "ws2", wd2 = "wd2")

## add some wd bias to some nighttime hours
id <- which(as.numeric(format(mydata$date, "\%H")) \%in\% c(23, 1, 2, 3, 4, 5))
mydata$wd2[id] <- mydata$wd[id] + 30 * rnorm(length(id)) + 120
id <- which(mydata$wd2 < 0)
mydata$wd2[id] <- mydata$wd2[id] + 360

pollutionRose(
  mydata,
  ws = "ws",
  wd = "wd",
  ws2 = "ws2",
  wd2 = "wd2",
  breaks = c(-11, -2, -1, -0.5, 0.5, 1, 2, 11),
  cols = c("dodgerblue4", "white", "firebrick"),
  type = "daylight"
)
}
}
\seealso{
Other polar directional analysis functions:
\code{\link[=percentileRose]{percentileRose()}},
\code{\link[=polarAnnulus]{polarAnnulus()}},
\code{\link[=polarCluster]{polarCluster()}},
\code{\link[=polarDiff]{polarDiff()}},
\code{\link[=polarFreq]{polarFreq()}},
\code{\link[=polarPlot]{polarPlot()}},
\code{\link[=windRose]{windRose()}}
}
\concept{polar directional analysis functions}
