% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt}
\alias{prompt}
\title{Prompt the chat model with user input}
\usage{
prompt(x, text, show = FALSE)
}
\arguments{
\item{x}{Either a character string indicating the name of the chat log object or a messages object (data.frame).}

\item{text}{A character string indicating the user input text.}

\item{show}{Logical, if TRUE, the current chat log is displayed via View(). Default is FALSE.}
}
\value{
A response from the chat model as a character string.
}
\description{
This function prompts the chat model with new user input.
}
\examples{
# Start a new chat session
chatlog_id <- start_chat()

# Prompt the chat model with user input
resp <- prompt(chatlog_id, "Hello!")
print(resp)

}
