% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_embeddings.R
\name{prune_embeddings}
\alias{prune_embeddings}
\title{reducing embedding dimensions}
\usage{
prune_embeddings(embeddings, ndim = 256)
}
\arguments{
\item{embeddings}{A vector, your embeddings}

\item{ndim}{the number of dimensions you wanted.(default is 256)}
}
\value{
An unit vector.
}
\description{
embeddings can be shortened to a size of ndim while still outperforming an unshortened version!
}
\examples{
x <- prune_embeddings(embeddings = x, ndim = 256)
}
