% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{python_to_r_prompt}
\alias{python_to_r_prompt}
\title{Input: message content template for python_to_r_prompt()}
\format{
## `python_to_r_prompt`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
python_to_r_prompt
}
\description{
This message content serves as user input for the python_to_r_prompt() function.
}
\keyword{datasets}
