% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_to_python_prompt}
\alias{r_to_python_prompt}
\title{Input: message content template for r_to_python()}
\format{
## `r_to_python_prompt`
A data frame with 2 rows and 2 columns representing the messages object in
API calls.
}
\source{
Contributed by umatter
}
\usage{
r_to_python_prompt
}
\description{
This message content serves as user input for the r_to_python() function.
}
\keyword{datasets}
