% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Refactor R Code with AI Assistance}
\usage{
refactor(file, ...)
}
\arguments{
\item{file}{file name containing R code to be refactored}

\item{...}{optional arguments to be passed to the chat_completion function}
}
\value{
The refactored code that can either be printed on the console or written to a file.
}
\description{
This function refactors R code with the assistance of an AI chatbot. 
It takes as input the file name containing R code that requires refactoring. 
After processing the content, it prompts the user to start a conversation with the chatbot. The conversation results in the refactored R code. The output can either be printed on the console or written to a file.
}
\examples{
\dontrun{
# Create a sample R function file
cat("my_sum <- function(a, b) {", "return(a + b)", "}", file = "sample_function.R")

# Refactor the R function and return the output
refactored_function <- refactor(file = "sample_function.R")

# Refactor the R function and write the output to the same file
refactor(file = "sample_function.R")
}
}
