% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refactor_prompt}
\alias{refactor_prompt}
\title{Input: message content template for refactor()}
\format{
## `nested_to_pipe_prompt`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
refactor_prompt
}
\description{
This message content serves as user input for the refactor() function.
}
\keyword{datasets}
