% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_to_bibtex.R
\name{references_to_bibtex}
\alias{references_to_bibtex}
\title{Convert references in plain text to BibTeX format}
\usage{
references_to_bibtex(references)
}
\arguments{
\item{references}{A character string or file path to a file containing the plain text references to convert.}
}
\value{
A character string containing the references in BibTeX format.
}
\description{
This function takes a character string or file path to plain text references and 
converts them to BibTeX format.
}
\examples{
\dontrun{
# Convert plain text references to BibTeX format
references <- "Doe, J., & Smith, J. (2020). The title of the paper.
Journal of Scientific Computing, 12, 45-67."
references_to_bibtex(references)
}
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
