% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{references_to_bibtex_input}
\alias{references_to_bibtex_input}
\title{Input: message content template for references_to_bibtex()}
\format{
## `references_to_bibtex_input`
A data frame with 4 rows and 2 columns representing the messages object in
API calls.
}
\source{
Contributed by umatter
}
\usage{
references_to_bibtex_input
}
\description{
This message content serves as user input for the references_to_bibtex
function.
}
\keyword{datasets}
