% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{retrieve_assistant}
\alias{retrieve_assistant}
\title{Retrieve an assistant (convenience function)}
\usage{
retrieve_assistant(assistant_id)
}
\arguments{
\item{assistant_id}{Assistant ID}
}
\value{
Assistant object
}
\description{
Retrieve an assistant (convenience function)
}
