% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{retrieve_batch}
\alias{retrieve_batch}
\title{Retrieve a batch (convenience function)}
\usage{
retrieve_batch(batch_id)
}
\arguments{
\item{batch_id}{Batch ID}
}
\value{
Batch object
}
\description{
Retrieve a batch (convenience function)
}
