% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{retrieve_file}
\alias{retrieve_file}
\title{Retrieve a file (convenience function)}
\usage{
retrieve_file(file_id)
}
\arguments{
\item{file_id}{File ID}
}
\value{
File object
}
\description{
Retrieve a file (convenience function)
}
