% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{retrieve_file_content}
\alias{retrieve_file_content}
\title{Retrieve file content (convenience function)}
\usage{
retrieve_file_content(file_id)
}
\arguments{
\item{file_id}{File ID}
}
\value{
Raw file content
}
\description{
Retrieve file content (convenience function)
}
