% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{retrieve_response}
\alias{retrieve_response}
\title{Retrieve a response (convenience function)}
\usage{
retrieve_response(response_id)
}
\arguments{
\item{response_id}{Response ID}
}
\value{
Response object
}
\description{
Retrieve a response (convenience function)
}
