% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{retrieve_run}
\alias{retrieve_run}
\title{Retrieve a run (convenience function)}
\usage{
retrieve_run(thread_id, run_id)
}
\arguments{
\item{thread_id}{Thread ID}

\item{run_id}{Run ID}
}
\value{
Run object
}
\description{
Retrieve a run (convenience function)
}
