% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{retrieve_thread}
\alias{retrieve_thread}
\title{Retrieve a thread (convenience function)}
\usage{
retrieve_thread(thread_id)
}
\arguments{
\item{thread_id}{Thread ID}
}
\value{
Thread object
}
\description{
Retrieve a thread (convenience function)
}
