% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{retrieve_vector_store}
\alias{retrieve_vector_store}
\title{Retrieve a vector store (convenience function)}
\usage{
retrieve_vector_store(vector_store_id)
}
\arguments{
\item{vector_store_id}{Vector store ID}
}
\value{
Vector store object
}
\description{
Retrieve a vector store (convenience function)
}
