% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rolling_average_cpp}
\alias{rolling_average_cpp}
\title{Rolling Average with Threshold}
\usage{
rolling_average_cpp(x, width, alignment, threshold)
}
\arguments{
\item{x}{A numeric vector}

\item{width}{Integer window width}

\item{alignment}{String: "left", "centre", or "right"}

\item{threshold}{Numeric fraction (0-1) of non-NA data required}
}
\description{
Rolling Average with Threshold
}
