% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollingQuantile.R
\name{rollingQuantile}
\alias{rollingQuantile}
\title{Calculate rolling quantile pollutant values}
\usage{
rollingQuantile(
  mydata,
  pollutant = "o3",
  width = 8L,
  type = "default",
  data.thresh = 75,
  align = c("centre", "center", "left", "right"),
  probs = 0.5,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing a \code{date} field. \code{mydata} must contain a
\code{date} field in \code{Date} or \code{POSIXct} format. The input time series must be
regular, e.g., hourly, daily.}

\item{pollutant}{The name of a pollutant, e.g., \code{pollutant = "o3"}.}

\item{width}{The averaging period (rolling window width) to use, e.g., \code{width = 8} will generate 8-hour rolling mean values when hourly data are
analysed.}

\item{type}{Used for splitting the data further. Passed to \code{\link[=cutData]{cutData()}}.}

\item{data.thresh}{The \% data capture threshold. No values are calculated if
data capture over the period of interest is less than this value. For
example, with \code{width = 8} and \code{data.thresh = 75} at least 6 hours are
required to calculate the mean, else \code{NA} is returned.}

\item{align}{Specifies how the moving window should be aligned. \code{"right"}
means that the previous hours (including the current) are averaged.
\code{"left"} means that the forward hours are averaged. \code{"centre"} (or
\code{"center"} - the default) centres the current hour in the window.}

\item{probs}{Probability for quantile calculate. A number between 0 and 1. Can be more than length one e.g. \code{probs = c(0.05, 0.95)}.}

\item{...}{Additional parameters passed to \code{\link[=cutData]{cutData()}}. For use with \code{type}.}
}
\value{
A tibble with new columns for the rolling quantile value and the number of valid values used.
}
\description{
This is a utility function mostly designed to calculate rolling quantile
statistics. The function will try
and fill in missing time gaps to get a full time sequence but return a data
frame with the same number of rows supplied.
}
\examples{
# rolling 24-hour 0.05 and 0.95 quantile for ozone
mydata <- rollingQuantile(mydata,
  pollutant = "o3", width = 24, data.thresh = 75, align = "right", probs = c(0.05, 0.95)
)
}
\author{
David Carslaw
}
