% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{scale_opts}
\alias{scale_opts}
\title{Define scale options for openair \code{ggplot2} plots}
\usage{
scale_opts(
  limits = NULL,
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  date_breaks = ggplot2::waiver(),
  date_labels = ggplot2::waiver(),
  transform = scales::transform_identity(),
  position = NULL,
  sec.axis = ggplot2::waiver()
)
}
\arguments{
\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}).
}}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:new_transform]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Note that for position scales, limits are provided after scale expansion.
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{labels}{One of the options below. Please note that when \code{labels} is a
vector, it is highly recommended to also set the \code{breaks} argument as a
vector to protect against unintended mismatches.
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{date_breaks}{A string giving the distance between breaks like "2
weeks", or "10 years". If both \code{breaks} and \code{date_breaks} are
specified, \code{date_breaks} wins. Valid specifications are 'sec', 'min',
'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.}

\item{date_labels}{A string giving the formatting specification for the
labels. Codes are defined in \code{\link[=strftime]{strftime()}}. If both \code{labels}
and \code{date_labels} are specified, \code{date_labels} wins.}

\item{transform}{For continuous scales, the name of a transformation object
or the object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \verb{transform_<name>}. If
transformations require arguments, you can call them from the scales
package, e.g. \code{\link[scales:transform_boxcox]{scales::transform_boxcox(p = 2)}}.
You can create your own transformation with \code{\link[scales:new_transform]{scales::new_transform()}}.}

\item{position}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}

\item{sec.axis}{\code{\link[ggplot2:sec_axis]{sec_axis()}} is used to specify a secondary axis.}
}
\description{
This function conveniently allows for options to be defined in relation to
the various scales of \code{openair} plots. These arguments are automatically
passed to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}. Note that, for certain plots, not
all of these may be controllable for the user (for example, if a scale
transformation would not make sense as in the y-axis of
\code{\link[=plot_polar_heatmap]{plot_polar_heatmap()}}). \code{date_breaks} and \code{date_labels} only apply to
date(time) axes (e.g., in \code{\link[=plot_trend_bars]{plot_trend_bars()}}).
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{disc_breaks}()},
\code{\link{facet_opts}()},
\code{\link{windflow_opts}()}
}
\author{
Jack Davison
}
\concept{ggplot2 plot utilities}
