% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_chatlog.R
\name{set_chatlog}
\alias{set_chatlog}
\title{Set up a new chatlog}
\usage{
set_chatlog(initial_system_content = "You are a helpful assistant.")
}
\arguments{
\item{initial_system_content}{A character string representing the initial message from the system}
}
\value{
A new chatlog object
}
\description{
This function sets up a new chatlog object for a conversation.
}
\examples{
chat <- set_chatlog("Welcome to our chat!")
is_chatlog(chat)
chat
}
