% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-ggplot.R
\name{shared_ggplot_params}
\alias{docs-ggplot-internal}
\title{Shared ggplot parameters}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The names of the pollutants to plot; these should be columns
in \code{data}.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{group}{The column of \code{data} to use to condition the data within each
panel, passed to \code{\link[=cutData]{cutData()}}. This is usually achieved using colour, so
will conflict with providing multiple \code{pollutant}s or other options which
use the colour aesthetic.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{discretise}{Options to discretise the continuous scale into discrete
bins. Providing an integer will split the scale into that number of
quantiles. Alternatively, \code{\link[=discretise_breaks]{discretise_breaks()}} allows for extensive
customisation (e.g., defining specific bin widths, breaks, labels, and so
on).}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}. Alternatively, \code{\link[=windflow_opts]{windflow_opts()}}
allows for greater customisation of the arrows (e.g., wind speed ranges,
relative size of arrows, arrow appearance).}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{r_axis_inside}{Either a numeric value, setting a theta axis value at
which the axis should be placed inside the panel, or \code{FALSE}, which will
place it outside of panel on a vertical axis.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
This is a central place for describing typical parameters. It
ensures consistency throughout openair.
}
\keyword{internal}
