% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{split_text}
\alias{split_text}
\title{Split Text into Chunks}
\usage{
split_text(text, N)
}
\arguments{
\item{text}{A character vector containing the text to be split.}

\item{N}{An integer specifying the number of tokens per chunk.}
}
\value{
A character vector containing the chunks of text with N tokens each.
}
\description{
This function splits a text string into a vector of strings with a
 specified number of tokens each.
}
\examples{
large_text <- "This is an example of a large text string
that will be split into chunks of N tokens each by our custom R function."
num_tokens_per_chunk <- 5
split_text(large_text, num_tokens_per_chunk)
}
