% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_chat.R
\name{start_chat}
\alias{start_chat}
\title{Start a new chat session}
\usage{
start_chat(
  initial_system_content = "You are a helpful assistant.",
  show = FALSE
)
}
\arguments{
\item{initial_system_content}{A character string indicating the initial content of the system message. The default value is "You are a helpful assistant.".}

\item{show}{Logical, if TRUE, the current chat log is displayed via View(). Default is FALSE.}
}
\value{
A character string indicating the name of the log environment created for the chat session.
}
\description{
This function starts a new chat session by initializing a messages object with an initial system message,
creating a new log environment to store the messages, and adding the messages object to the log environment.
}
\examples{
# Start a new chat session with the default system message
chatlog_id <- start_chat()

# Start a new chat session with a custom system message
chatlog_id <- start_chat("Hello! How can I assist you today?")

}
