\name{stlPlot}
\alias{stlPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ stlPlot }
\description{
  plot for `stl' function
}
\usage{
stlPlot(mydata, pollutant = "nox", s.wind = 17, xlab = "year", ...)
}
\arguments{
  \item{mydata}{ openair data frame}
  \item{pollutant}{ \code{pollutant} to use in \code{mydata}. }
  \item{s.wind}{ \code{s.wind} input }
  \item{xlab}{ The x axis label }
  \item{\dots}{ Any extra options to pass on. }
}
\details{
  Function in development
}
\value{
  Function in development
}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
%\note{ ~~further notes~~

% ~Make other sections like Warning with \section{Warning }{....} ~
%}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
