% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{StreamIterator}
\alias{StreamIterator}
\title{Stream Iterator Class}
\description{
R6 class for iterating over streaming response chunks.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StreamIterator-new}{\code{StreamIterator$new()}}
\item \href{#method-StreamIterator-next_chunk}{\code{StreamIterator$next_chunk()}}
\item \href{#method-StreamIterator-has_more}{\code{StreamIterator$has_more()}}
\item \href{#method-StreamIterator-reset}{\code{StreamIterator$reset()}}
\item \href{#method-StreamIterator-as_list}{\code{StreamIterator$as_list()}}
\item \href{#method-StreamIterator-get_full_text}{\code{StreamIterator$get_full_text()}}
\item \href{#method-StreamIterator-clone}{\code{StreamIterator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-new"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$new(chunks)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-next_chunk"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-next_chunk}{}}}
\subsection{Method \code{next_chunk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$next_chunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-has_more"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-has_more}{}}}
\subsection{Method \code{has_more()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$has_more()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-reset"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-as_list}{}}}
\subsection{Method \code{as_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$as_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-get_full_text"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-get_full_text}{}}}
\subsection{Method \code{get_full_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$get_full_text()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StreamIterator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StreamIterator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StreamIterator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
