% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAIR-package.R
\docType{package}
\name{TheOpenAIR-package}
\alias{TheOpenAIR}
\alias{TheOpenAIR-package}
\title{TheOpenAIR: Integrate OpenAI's Large Language Models into Your R Workflows}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

TheOpenAIR builds on OpenAI's API as backend and provides R wrapper functions for the ChatGPT endpoint as well as various high-level functions to integrate ChatGPT's functionality in a variety of data handling and data analytics tasks, ranging from data cleaning to the automated generation of analytics scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://openair-lib.org/}
}

}
\author{
\strong{Maintainer}: Ulrich Matter \email{umatter@protonmail.com} [copyright holder]

Authors:
\itemize{
  \item Jonathan Chassot \email{jonathan.chassot@unisg.ch}
}

}
\keyword{internal}
