% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidyair_input}
\alias{tidyair_input}
\title{Input: message content template for tidyair()}
\format{
## `tidyair`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
tidyair_input
}
\description{
This message content serves as user input for the tidyair function.
}
\keyword{datasets}
