% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeProp.R
\name{timeProp}
\alias{timeProp}
\title{Time series plot with categories shown as a stacked bar chart}
\usage{
timeProp(
  mydata,
  pollutant = "nox",
  proportion = "wd",
  avg.time = "day",
  type = "default",
  cols = "Set1",
  theme = "default",
  normalise = FALSE,
  ref.x = NULL,
  ref.y = NULL,
  key.columns = NULL,
  key.rows = NULL,
  key.position = "right",
  key.title = proportion,
  date.breaks = NULL,
  date.format = NULL,
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the fields \code{date}, \code{pollutant} and a
splitting variable \code{proportion}}

\item{pollutant}{Name of the pollutant to plot contained in \code{mydata}.}

\item{proportion}{The splitting variable that makes up the bars in the bar
chart, defaulting to \code{"wd"}. Could be \code{"cluster"} if the output from
\code{\link[=polarCluster]{polarCluster()}} or \code{\link[=trajCluster]{trajCluster()}} is being analysed. If \code{proportion} is a
numeric variable it is split into 4 quantiles (by default) by \code{\link[=cutData]{cutData()}}.
If \code{proportion} is a factor or character variable then the categories are
used directly.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be \code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in which case
3-month seasonal values are calculated with spring defined as March, April,
May and so on.

Note that \code{avg.time} when used in \code{\link[=timeProp]{timeProp()}} should be greater than the
time gap in the original data. For example, \code{avg.time = "day"} for hourly
data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{theme}{A string representing an overall plot theme, defaulting to
\code{"default"}. This option makes sweeping changes to non-data plot features
such as fonts, colours, line widths, and so on, and may also change default
arguments like \code{cols} if not set by the user. Can also take a
\code{\link[ggplot2:theme]{ggplot2::theme()}} object, which will be used to modify the \code{"default"}
theme. Pre-set options include:
\itemize{
\item \code{"default"}, a lattice-inspired theme resembling the traditional
\code{openair} look, with structured panels and visible gridlines.
\item \code{"dark"}, a dark-background variant of the default theme, designed for
presentations and low-light viewing, using high-contrast text and colour
palettes optimised for visibility against dark panels.
\item \code{"modern"}, a minimalist, contemporary theme inspired by tools such as
Plotly and Observable Plot, with reduced visual clutter, horizontal
emphasis in gridlines, a clean legend style, and typography suited to
dashboards and reports.
\item \code{"soft"}, a low-contrast, 'editorial' theme with warm
background tones, subtle gridlines, and gently desaturated colours,
designed for reports and publication-style figures, particularly where a
calmer appearance improves readability.
\item \code{"print"}, a strictly greyscale theme optimised for black-and-white
reproduction, with stronger structural elements such as clearer gridlines
and axis definitions to ensure good contrast and readability in printed or
photocopied outputs.
}

Please note that if a global theme is set with \code{\link[ggplot2:theme_set]{ggplot2::theme_set()}} to
anything other than the default \code{\link[ggplot2:theme_grey]{ggplot2::theme_grey()}}, the selected
openair theme will not be fully applied; instead, only minimal adjustments
(such as legend positioning) will be made.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{ref.x}{Either a single value or values representing the x axis
intercepts to draw lines, or a list such as that provided by \code{\link[=refOpts]{refOpts()}} to
customise the colour/width/type/etc. of each line. See \code{\link[=refOpts]{refOpts()}} for more
details.}

\item{ref.y}{Either a single value or values representing the y axis
intercepts to draw lines, or a list such as that provided by \code{\link[=refOpts]{refOpts()}} to
customise the colour/width/type/etc. of each line. See \code{\link[=refOpts]{refOpts()}} for more
details.}

\item{key.columns, key.rows}{Number of columns or rows to be used in a
categorical legend. With many categories a single column can make to key
too wide. The user can thus choose to use several columns by setting
\code{key.columns} to be less than the number of categories.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{date.breaks}{Either a number of major x-axis intervals to use, or a
character string giving the distance between breaks like \code{"2 weeks"}, or
\code{"10 years"}. Valid specifications are \code{"sec"}, \code{"min"}, \code{"hour"}, \code{"day"},
\code{"week"}, \code{"month"} or \code{"year"}, optionally followed by \code{"s"}. If
\code{date.breaks} is given a character string, it is recommended to also set
\code{date.format} as this will automatically default to \code{"\%Y-\%m-\%d"}.}

\item{date.format}{This option controls the date format on the x-axis. A
sensible format is chosen by default, but the user can set \code{date.format} to
override this. For format types see \code{\link[=strptime]{strptime()}}. For example, to format
the date like "Jan-2012" set \code{date.format = "\\\%b-\\\%Y"}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label, passed to
\code{\link[ggplot2:labs]{ggplot2::labs()}} via \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{scales} can be \code{"fixed"}, \code{"free_x"}, \code{"free_y"} or \code{"free"} to control
whether axes are shared across facets when using \code{type}. Also supported are
the legacy \code{x.relation} and \code{y.relation}, which can be either \code{"same"} or
\code{"free"} and get remapped to \code{scales} automatically.
\item Similarly, \code{space}, \code{axes}, \code{axis.labels}, \code{switch} and \code{strip.position}
can be used to customise the appearance of faceted plots. See
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} for the arguments these
take.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the plotting
of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and \code{\link[=trajCluster]{trajCluster()}} that
consider local and regional (back trajectory) cluster analysis respectively.
However, the function has more general use for understanding time series
data.
}
\details{
In order to plot time series in this way, some sort of time aggregation is
needed, which is controlled by the option \code{avg.time}.

The plot shows the value of \code{pollutant} on the y-axis (averaged according to
\code{avg.time}). The time intervals are made up of bars split according to
\code{proportion}. The bars therefore show how the total value of \code{pollutant} is
made up for any time interval.
}
\examples{
# monthly plot of SO2 showing the contribution by wind sector
timeProp(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
Other time series and trend functions:
\code{\link[=TheilSen]{TheilSen()}},
\code{\link[=calendarPlot]{calendarPlot()}},
\code{\link[=smoothTrend]{smoothTrend()}},
\code{\link[=timePlot]{timePlot()}},
\code{\link[=timeVariation]{timeVariation()}}

Other cluster analysis functions:
\code{\link[=polarCluster]{polarCluster()}},
\code{\link[=trajCluster]{trajCluster()}}
}
\author{
David Carslaw

Jack Davison
}
\concept{cluster analysis functions}
\concept{time series and trend functions}
