% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajCluster.R
\name{trajCluster}
\alias{trajCluster}
\title{Calculate clusters for back trajectories}
\usage{
trajCluster(
  traj,
  method = "Euclid",
  n.cluster = 5,
  type = "default",
  split.after = FALSE,
  by.type = FALSE,
  crs = 4326,
  cols = "Set1",
  plot = TRUE,
  ...
)
}
\arguments{
\item{traj}{An openair trajectory data frame resulting from the use of
\code{\link[=importTraj]{importTraj()}}.}

\item{method}{Method used to calculate the distance matrix for the back
trajectories. There are two methods available: \dQuote{Euclid} and
\dQuote{Angle}.}

\item{n.cluster}{Number of clusters to calculate.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{split.after}{For \code{type} other than \dQuote{default} e.g.
\dQuote{season}, the trajectories can either be calculated for each level
of \code{type} independently or extracted after the cluster calculations have
been applied to the whole data set.}

\item{by.type}{The percentage of the total number of trajectories is given
for all data by default. Setting \code{by.type = TRUE} will make each panel add
up to 100.}

\item{crs}{The coordinate reference system to use for plotting. Defaults to
\code{4326}, which is the WGS84 geographic coordinate system, the standard,
unprojected latitude/longitude system used in GPS, Google Earth, and GIS
mapping. Other \code{crs} values are available - for example, \code{27700} will use
the the OSGB36/British National Grid.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Passed to \code{\link[=trajPlot]{trajPlot()}}.}
}
\value{
an \link[=openair-package]{openair} object. The \code{data} component contains
both \code{traj} (the original data appended with its cluster) and \code{results}
(the average trajectory path per cluster, shown in the \code{trajCluster()}
plot.)
}
\description{
This function carries out cluster analysis of HYSPLIT back trajectories. The
function is specifically designed to work with the trajectories imported
using the \code{openair} \code{\link[=importTraj]{importTraj()}} function, which provides pre-calculated
back trajectories at specific receptor locations.
}
\details{
Two main methods are available to cluster the back trajectories using two
different calculations of the distance matrix. The default is to use the
standard Euclidian distance between each pair of trajectories. Also available
is an angle-based distance matrix based on Sirois and Bottenheim (1995). The
latter method is useful when the interest is the direction of the
trajectories in clustering.

The distance matrix calculations are made in C++ for speed. For data sets of
up to 1 year both methods should be relatively fast, although the \code{method = "Angle"} does tend to take much longer to calculate. Further details of these
methods are given in the openair manual.
}
\examples{
\dontrun{
## import trajectories
traj <- importTraj(site = "london", year = 2009)
## calculate clusters
clust <- trajCluster(traj, n.cluster = 5)
head(clust$data) ## note new variable 'cluster'
## use different distance matrix calculation, and calculate by season
traj <- trajCluster(traj, method = "Angle", type = "season", n.cluster = 4)
}
}
\references{
Sirois, A. and Bottenheim, J.W., 1995. Use of backward trajectories to
interpret the 5-year record of PAN and O3 ambient air concentrations at
Kejimkujik National Park, Nova Scotia. Journal of Geophysical Research, 100:
2867-2881.
}
\seealso{
Other trajectory analysis functions:
\code{\link[=importTraj]{importTraj()}},
\code{\link[=trajLevel]{trajLevel()}},
\code{\link[=trajPlot]{trajPlot()}}

Other cluster analysis functions:
\code{\link[=polarCluster]{polarCluster()}},
\code{\link[=timeProp]{timeProp()}}
}
\author{
David Carslaw

Jack Davison
}
\concept{cluster analysis functions}
\concept{trajectory analysis functions}
