% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajCluster_TSV.R
\name{trajCluster_TSV}
\alias{trajCluster_TSV}
\title{Calculate the TSV values for a set of cluster analyses.}
\usage{
trajCluster_TSV(traj, verbose = TRUE, nclusters = 5, normalise = TRUE)
}
\arguments{
\item{traj}{A trajectory results data.frame, formatted for openair.}

\item{verbose}{TRUE or FALSE to display progress text.}

\item{nclusters}{a numeric vector, containing the number of clusters to run. E.g. seq(2,10,1) will calculate TSV for 2 through 10 clusters.}

\item{normalise}{TRUE or FALSE to normalise calculated TSV values. Non-normalised values are left as total summed haversine distances.}
}
\description{
Uses \code{\link{trajCluster}} and \code{\link{cluster_TSV}} to generate clusters and
their corresponding total spatial variance (TSV), giving an indication as to how well a given
clustering attempt reduces relative spatial variance. Generally it is best to calculate the
TSV for a series of clusters, e.g. 2 through 20, and then identify the points at which
Variance increases suddenly. See e.g. Kassomenos et al., 2010.
}
