% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistants.R
\name{update_assistant}
\alias{update_assistant}
\title{Update an assistant (convenience function)}
\usage{
update_assistant(assistant_id, ...)
}
\arguments{
\item{assistant_id}{Assistant ID}

\item{...}{Fields to update}
}
\value{
Updated assistant
}
\description{
Update an assistant (convenience function)
}
