% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{update_thread}
\alias{update_thread}
\title{Update a thread (convenience function)}
\usage{
update_thread(thread_id, ...)
}
\arguments{
\item{thread_id}{Thread ID}

\item{...}{Fields to update}
}
\value{
Updated thread
}
\description{
Update a thread (convenience function)
}
