% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file (convenience function)}
\usage{
upload_file(file, purpose)
}
\arguments{
\item{file}{File path or raw content}

\item{purpose}{Purpose of the file}
}
\value{
File object
}
\description{
Upload a file (convenience function)
}
