% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{UploadsClient}
\alias{UploadsClient}
\title{Uploads Client}
\description{
Client for OpenAI Uploads API.
Upload large files in parts.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UploadsClient-new}{\code{UploadsClient$new()}}
\item \href{#method-UploadsClient-create}{\code{UploadsClient$create()}}
\item \href{#method-UploadsClient-add_part}{\code{UploadsClient$add_part()}}
\item \href{#method-UploadsClient-complete}{\code{UploadsClient$complete()}}
\item \href{#method-UploadsClient-cancel}{\code{UploadsClient$cancel()}}
\item \href{#method-UploadsClient-clone}{\code{UploadsClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$create(purpose, filename, bytes, mime_type = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-add_part"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-add_part}{}}}
\subsection{Method \code{add_part()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$add_part(upload_id, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-complete"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-complete}{}}}
\subsection{Method \code{complete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$complete(upload_id, part_ids, md5 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-cancel}{}}}
\subsection{Method \code{cancel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$cancel(upload_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadsClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UploadsClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UploadsClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
