% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage}
\alias{usage}
\title{Extract usage data from OpenAI API response}
\usage{
usage(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a data frame with the usage statistics of the API call (how many tokens used)
}
\description{
This function extracts the usage data from the parsed HTTP response of an API call to the
OpenAI chat completions endpoint.
}
