% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stores.R
\name{VectorStoresClient}
\alias{VectorStoresClient}
\title{Vector Stores Client (Beta)}
\description{
Client for OpenAI Vector Stores API v2.
Vector stores are used for file search with assistants.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VectorStoresClient-new}{\code{VectorStoresClient$new()}}
\item \href{#method-VectorStoresClient-create}{\code{VectorStoresClient$create()}}
\item \href{#method-VectorStoresClient-list}{\code{VectorStoresClient$list()}}
\item \href{#method-VectorStoresClient-retrieve}{\code{VectorStoresClient$retrieve()}}
\item \href{#method-VectorStoresClient-update}{\code{VectorStoresClient$update()}}
\item \href{#method-VectorStoresClient-delete}{\code{VectorStoresClient$delete()}}
\item \href{#method-VectorStoresClient-search}{\code{VectorStoresClient$search()}}
\item \href{#method-VectorStoresClient-clone}{\code{VectorStoresClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-create"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$create(
  name = NULL,
  file_ids = NULL,
  expires_after = NULL,
  chunking_strategy = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-list"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$list(
  limit = NULL,
  order = NULL,
  after = NULL,
  before = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$retrieve(vector_store_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-update"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$update(vector_store_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-delete"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$delete(vector_store_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-search"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-search}{}}}
\subsection{Method \code{search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$search(
  vector_store_id,
  query,
  filter = NULL,
  max_num_results = NULL,
  ranking_options = NULL,
  rewrite_query = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VectorStoresClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VectorStoresClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VectorStoresClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
