% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{windflow_opts}
\alias{windflow_opts}
\title{Define 'windflow' options for openair \code{ggplot2} plots}
\usage{
windflow_opts(
  limits = c(NA, NA),
  range = c(0.1, 1),
  arrow = ggplot2::arrow(angle = 15, length = ggplot2::unit(0.5, "lines"), ends = "last",
    type = "closed"),
  windflow = TRUE
)
}
\arguments{
\item{limits}{The limits of wind speed to include in the windflow annotation.
Any wind speeds outside of this range will be set to the maximum/minimum
value, as appropriate. Useful for ensuring arrows are consistent between
plots, or - for example - extending the wind speed limit to zero using
\code{range = c(0, NA)}.}

\item{range}{The range of possible sizes of the windflow arrows. The default
is broadly appropriate throughout \code{openair}, but if plots are being saved
at different resolutions it may be useful to tweak this.}

\item{arrow}{See \code{\link[ggplot2:reexports]{ggplot2::arrow()}}.}

\item{windflow}{Include a windflow layer on the plot? Defaults to \code{TRUE}, and
used internally by \code{openair} functions that may wish to override the user's
options.}
}
\description{
Various \code{openair} functions include the \code{windflow} argument to include wind
speed & direction arrows. This function can be provided to the \code{windflow}
argument in place of \code{TRUE}/\code{FALSE} to better control the appearance of the
arrows.
}
\seealso{
Other ggplot2 plot utilities: 
\code{\link{discretise_breaks}()},
\code{\link{facet_opts}()}
}
\author{
Jack Davison
}
\concept{ggplot2 plot utilities}
