% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windflowOpts.R
\name{windflowOpts}
\alias{windflowOpts}
\title{Define \code{windflow} options for \code{openair} plots}
\usage{
windflowOpts(
  limits = c(NA, NA),
  range = c(0.1, 1),
  arrow.angle = 15,
  arrow.length = ggplot2::unit(0.5, "lines"),
  arrow.ends = "last",
  arrow.type = "closed",
  lineend = "butt",
  alpha = 1,
  colour = "black",
  linetype = 1,
  linewidth = 0.5,
  color = NULL,
  windflow = TRUE
)
}
\arguments{
\item{limits}{Numeric vector of length 2 specifying the limits for wind
speed. By default, it is set to \code{c(NA, NA)}, which means the limits will be
determined automatically based on the data.}

\item{range}{Numeric vector of length 2 specifying the range of possible
sizes of the windflow arrows. The default is broadly appropriate throughout
\code{openair}, but if plots are beings saved at different resolutions it may be
appropriate to to tweak this.}

\item{arrow.angle, arrow.length, arrow.ends, arrow.type}{Passed to the
respective arguments of \code{\link[grid:arrow]{grid::arrow()}}; used to control various arrow
styling options.}

\item{lineend}{The style of line endings. Options include \code{"butt"},
\code{"round"}, and \code{"square"}. Default is \code{"butt"}.}

\item{alpha}{Numeric value between 0 and 1 specifying the transparency of the
lines. Default is 1 (fully opaque).}

\item{colour, color}{Colour of the lines. Default is \code{"black"}. \code{colour} and
\code{color} are interchangeable, but \code{colour} is used preferentially if both
are given.}

\item{linetype}{Line type. Can be an integer (e.g., 1 for solid, 2 for
dashed) or a string (e.g., "solid", "dashed"). Default is 1 (solid).}

\item{linewidth}{Numeric value specifying the width of the lines. Default is
0.5.}

\item{windflow}{Logical value indicating whether to include the \code{windflow}
layer. Default is \code{TRUE}. Used internally by \code{openair} functions.}

\item{arrow}{A \code{\link[grid:arrow]{grid::arrow()}} object specifying the appearance of the
arrows.}
}
\value{
A list of options that can be passed to the \code{windflow} argument of
functions like \code{\link[=trendLevel]{trendLevel()}}.
}
\description{
This function provides a convenient way to set default options for \code{windflow}
layers in \code{openair} plots, which typically show the mean wind speed and
direction as compass arrows. It returns a list of options that can be passed
to \code{layer_windflow()}.
}
\examples{

# `windflow` can be `TRUE` to use defaults
trendLevel(mydata, type = "default", cols = "greyscale", windflow = TRUE)

# use the `windflowOpts()` function to customise arrows
trendLevel(
  mydata,
  type = "default",
  cols = "greyscale",
  windflow = windflowOpts(
    colour = "white",
    alpha = 0.8,
    linewidth = 0.25,
    linetype = 2
  )
)
}
