% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windRose.R
\name{windRose}
\alias{windRose}
\title{Traditional wind rose plot}
\usage{
windRose(
  mydata,
  ws = "ws",
  wd = "wd",
  ws2 = NA,
  wd2 = NA,
  ws.int = 2,
  angle = 30,
  type = "default",
  calm.thresh = 0,
  bias.corr = TRUE,
  cols = "default",
  grid.line = NULL,
  width = 0.9,
  seg = 0.9,
  auto.text = TRUE,
  breaks = 4,
  offset = 10,
  normalise = FALSE,
  max.freq = NULL,
  paddle = TRUE,
  key = TRUE,
  key.title = "(m/s)",
  key.position = "bottom",
  strip.position = "top",
  dig.lab = 5,
  include.lowest = FALSE,
  statistic = "prop.count",
  pollutant = NULL,
  annotate = TRUE,
  angle.scale = 315,
  border = NA,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing fields \code{ws} and \code{wd}}

\item{ws}{Name of the column representing wind speed.}

\item{wd}{Name of the column representing wind direction.}

\item{ws2, wd2}{The user can supply a second set of wind speed and wind
direction values with which the first can be compared. See
\code{\link[=pollutionRose]{pollutionRose()}} for more details.}

\item{ws.int}{The Wind speed interval. Default is 2 m/s but for low met masts
with low mean wind speeds a value of 1 or 0.5 m/s may be better.}

\item{angle}{Default angle of \dQuote{spokes} is 30. Other potentially useful
angles are 45 and 10. Note that the width of the wind speed interval may
need adjusting using \code{width}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{calm.thresh}{By default, conditions are considered to be calm when the
wind speed is zero. The user can set a different threshold for calms be
setting \code{calm.thresh} to a higher value. For example, \code{calm.thresh = 0.5}
will identify wind speeds \strong{below} 0.5 as calm.}

\item{bias.corr}{When \code{angle} does not divide exactly into 360 a bias is
introduced in the frequencies when the wind direction is already supplied
rounded to the nearest 10 degrees, as is often the case. For example, if
\code{angle = 22.5}, N, E, S, W will include 3 wind sectors and all other angles
will be two. A bias correction can made to correct for this problem. A
simple method according to Applequist (2012) is used to adjust the
frequencies.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{grid.line}{Grid line interval to use. If \code{NULL}, as in default, this is
assigned based on the available data range. However, it can also be forced
to a specific value, e.g. \code{grid.line = 10}. \code{grid.line} can also be a list
to control the interval, line type and colour. For example \code{grid.line = list(value = 10, lty = 5, col = "purple")}.}

\item{width}{For paddle = TRUE, the adjustment factor for width of wind speed
intervals. For example, width = 1.5 will make the paddle width 1.5 times
wider.}

\item{seg}{\code{seg} determines with width of the segments. For example, \code{seg = 0.5} will produce segments 0.5 * \code{angle}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{breaks}{Most commonly, the number of break points for wind speed. With
the \code{ws.int} default of 2 m/s, the \code{breaks} default, 4, generates the break
points 2, 4, 6, 8 m/s. However, \code{breaks} can also be used to set specific
break points. For example, the argument \code{breaks = c(0, 1, 10, 100)} breaks
the data into segments <1, 1-10, 10-100, >100.}

\item{offset}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}

\item{normalise}{If \code{TRUE} each wind direction segment is normalised to equal
one. This is useful for showing how the concentrations (or other
parameters) contribute to each wind sector when the proportion of time the
wind is from that direction is low. A line showing the probability that the
wind directions is from a particular wind sector is also shown.}

\item{max.freq}{Controls the scaling used by setting the maximum value for
the radial limits. This is useful to ensure several plots use the same
radial limits.}

\item{paddle}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE} plots rose using 'paddle'
style spokes. If \code{FALSE} plots rose using 'wedge' style spokes.}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{dig.lab}{The number of significant figures at which scientific number
formatting is used in break point and key labelling. Default 5.}

\item{include.lowest}{Logical. If \code{FALSE} (the default), the first interval
will be left exclusive and right inclusive. If \code{TRUE}, the first interval
will be left and right inclusive. Passed to the \code{include.lowest} argument
of \code{\link[=cut]{cut()}}.}

\item{statistic}{The \code{statistic} to be applied to each data bin in the plot.
Options currently include \dQuote{prop.count}, \dQuote{prop.mean} and
\dQuote{abs.count}. The default \dQuote{prop.count} sizes bins according to
the proportion of the frequency of measurements.  Similarly,
\dQuote{prop.mean} sizes bins according to their relative contribution to
the mean. \dQuote{abs.count} provides the absolute count of measurements in
each bin.}

\item{pollutant}{Alternative data series to be sampled instead of wind speed.
The \code{\link[=windRose]{windRose()}} default NULL is equivalent to \code{pollutant = "ws"}. Use in
\code{\link[=pollutionRose]{pollutionRose()}}.}

\item{annotate}{If \code{TRUE} then the percentage calm and mean values are
printed in each panel together with a description of the statistic below
the plot. If \code{FALSE} then only the statistic will be printed.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{border}{Border colour for shaded areas. Default is no border.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
}}
}
\value{
an \link[=openair-package]{openair} object. Summarised proportions can be
extracted directly using the \verb{$data} operator, e.g. \code{object$data} for
\code{output <- windRose(mydata)}. This returns a data frame with three set
columns: \code{cond}, conditioning based on \code{type}; \code{wd}, the wind direction;
and \code{calm}, the \code{statistic} for the proportion of data unattributed to any
specific wind direction because it was collected under calm conditions; and
then several (one for each range binned for the plot) columns giving
proportions of measurements associated with each \code{ws} or \code{pollutant} range
plotted as a discrete panel.
}
\description{
The traditional wind rose plot that plots wind speed and wind direction by
different intervals. The pollution rose applies the same plot structure but
substitutes other measurements, most commonly a pollutant time series, for
wind speed.
}
\details{
For \code{windRose} data are summarised by direction, typically by 45 or 30 (or
10) degrees and by different wind speed categories. Typically, wind speeds
are represented by different width "paddles". The plots show the proportion
(here represented as a percentage) of time that the wind is from a certain
angle and wind speed range.

By default \code{windRose} will plot a windRose in using "paddle" style segments
and placing the scale key below the plot.

The argument \code{pollutant} uses the same plotting structure but substitutes
another data series, defined by \code{pollutant}, for wind speed. It is
recommended to use \code{\link[=pollutionRose]{pollutionRose()}} for plotting pollutant concentrations.

The option \code{statistic = "prop.mean"} provides a measure of the relative
contribution of each bin to the panel mean, and is intended for use with
\code{pollutionRose}.
}
\examples{
# basic plot
windRose(mydata)

# one windRose for each year
windRose(mydata, type = "year")

# windRose in 10 degree intervals with gridlines and width adjusted
\dontrun{
windRose(mydata, angle = 10, width = 0.2, grid.line = 1)
}
}
\references{
Applequist, S, 2012: Wind Rose Bias Correction. J. Appl. Meteor.
Climatol., 51, 1305-1309.

Droppo,  J.G. and B.A. Napier (2008) Wind Direction Bias in Generating Wind
Roses and Conducting Sector-Based Air Dispersion Modeling, Journal of the
Air & Waste Management Association, 58:7, 913-918.
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarAnnulus}()},
\code{\link{polarCluster}()},
\code{\link{polarDiff}()},
\code{\link{polarFreq}()},
\code{\link{polarPlot}()},
\code{\link{pollutionRose}()}
}
\author{
David Carslaw

Karl Ropkins

Jack Davison
}
\concept{polar directional analysis functions}
