% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_test.R
\name{write_test}
\alias{write_test}
\title{Write Test}
\usage{
write_test(file)
}
\arguments{
\item{file}{A file to test documentation for}
}
\value{
The processed r_function with the Roxygen2 documentation
}
\description{
Writes a test file with Roxygen2 documentation
}
\examples{
write_test("inst/text/func.R")


}
\author{
Ulrich Matter umatter@protonmail.com
}
