% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{write_test_input}
\alias{write_test_input}
\title{Input: message content template for write_test()}
\format{
## `write_test_input`
A data frame with 2 rows and 2 columns representing the messages object in API calls.
}
\source{
Contributed by umatter
}
\usage{
write_test_input
}
\description{
This message content serves as user input for the write_test function.
}
\keyword{datasets}
