% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yorkFit.R
\name{YorkFit}
\alias{YorkFit}
\title{York regression function}
\usage{
YorkFit(
  input_data,
  X = "X",
  Y = "Y",
  Xstd = "Xstd",
  Ystd = "Ystd",
  weight = NA,
  Ri = 0,
  tol = 1e-07,
  max_iterations = 100
)
}
\arguments{
\item{input_data}{A data frame minimally containing 2 pollutants and at least one column with variable errors.}

\item{X}{Name of first variable}

\item{Y}{Name of second variable}

\item{Xstd}{Name of errors in first variable (X)}

\item{Ystd}{Name of errors in second variable (Y)}

\item{weight}{Can be supplied to apply to errors}

\item{Ri}{Correlation coefficient between errors in X and Y. Values must be within the range -1 to +1.}

\item{tol}{Threshold of differences of estimates of the slope to determine convergence. Should not need to be changed.}

\item{max_iterations}{Maximum number of iterations to identify the slope.}
}
\description{
Function for calculating the slope of variables using the York regression,
which accounts for errors provided for both x and y.
}
\details{
The York regression (YR) was suggested by Derek York from the University of Toronto in 1968 (CITE).
YR incorporates the correlation coefficient of errors in the minimising function.
The distance minimised for YR has an angle which depends on the weight of x and y errors.
When errors are higher for a variable, the relevant data points will weigh less, due to their bigger uncertainty.

The method employed uses an iterative approach to estimate the slope.

The York approach is common among the geophysical community, but very limited for other sciences (CITE WEHR).

The slope, slope error and intercept are calculated using the YorkFit() function.

YR can also be incorporated in polar plots, where the inverse effect of x/y and y/x is more distinct that Ordinary Least Squares.
}
\examples{

# Simulate data with measurement errors.
true_x <- 1:50
true_y <- 1:50

set.seed(100)
sigma_x <- runif(length(true_x), 1, 20)
sigma_y <- runif(length(true_y), 1, 2)

noisy_x <- rnorm(length(true_x), true_x, sigma_x)
noisy_y <- rnorm(length(true_y), true_y, sigma_y)

york_data <- data.frame(
  X = noisy_x,
  Y = noisy_y,
  Xstd = sigma_x,
  Ystd = sigma_y
)

YorkFit(york_data, X = "X", Y = "Y",
        Xstd = "Xstd", Ystd = "Ystd")
}
\author{
Magdalini Alexiadou, Stuart Lacy
}
