% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_roxygen.R
\name{add_roxygen}
\alias{add_roxygen}
\title{Add Roxygen2 Documentation to R Function}
\usage{
add_roxygen(file)
}
\arguments{
\item{file}{character string with the name of the R file to process}
}
\description{
This function reads in an R file, adds Roxygen2 documentation using chat input, and outputs the processed R file with Roxygen2 documentation.
}
\examples{
add_roxygen("myRfile.R")


}
\author{
Ulrich Matter umatter@protonmail.com
}
