% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_roxygen.R
\name{add_roxygen}
\alias{add_roxygen}
\title{Add Roxygen2 documentation to an R function}
\usage{
add_roxygen(file, ...)
}
\arguments{
\item{file}{A character string indicating the path to the file containing the
R function.}

\item{...}{Additional arguments to pass to the chat_completion() function.}
}
\value{
If the path provided is a character string, this function returns the
documented function as a character string. If the input is a file path, this
function returns the path of the file to which documentation was added to the
file.
}
\description{
This function adds Roxygen2 documentation to an R function.
}
\author{
Ulrich Matter umatter@protonmail.com
}
