% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_chatlog.R
\name{add_to_chatlog}
\alias{add_to_chatlog}
\title{Add data to a chat log}
\usage{
add_to_chatlog(msgs, chatlog_id)
}
\arguments{
\item{msgs}{a chatlog object}

\item{chatlog_id}{The id of the chatlog object}
}
\value{
The resulting environment with the added data.
}
\description{
This function adds data to the in-memory chat log, which is in the global environment. The data
can be any R object, such as a vector, list, or data frame.
}
