% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqStats.R
\name{aqStats}
\alias{aqStats}
\title{Calculate summary statistics for air pollution data by year}
\usage{
aqStats(mydata, pollutant = "no2", type = "default", data.thresh = 0,
  percentile = c(95, 99), transpose = FALSE, ...)
}
\arguments{
\item{mydata}{A data frame containing a \code{date} field of hourly data.}

\item{pollutant}{The name of a pollutant e.g. \code{pollutant = c("o3",
"pm10")}.}

\item{type}{\code{type} allows \code{timeAverage} to be applied to
cases where there are groups of data that need to be split and
the function applied to each group. The most common example is
data with multiple sites identified with a column representing
site name e.g. \code{type = "site"}. More generally, \code{type}
should be used where the date repeats for a particular grouping
variable.}

\item{data.thresh}{The data capture threshold in %. No values are calculated
if data capture over the period of interest is less than this value.
\code{data.thresh} is used for example in the calculation of daily mean
values from hourly data. If there are less than \code{data.thresh}
percentage of measurements available in a period, \code{NA} is returned.}

\item{percentile}{Percentile values to calculate for each pollutant.}

\item{transpose}{The default is to return a data frame with columns
representing the statistics. If \code{transpose = TRUE} then the results
have columns for each pollutant-site combination.}

\item{...}{Other arguments, currently unused.}
}
\description{
Calculate a range of air pollution-relevant statistics by year.
}
\examples{

## Statistics for 2004. NOTE! these data are in ppb/ppm so the
## example is for illustrative purposes only
aqStats(selectByDate(mydata, year = 2004), pollutant = "no2")


}
\author{
David Carslaw
}
\keyword{methods}
