% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAQStats.R
\name{calcAQStats}
\alias{calcAQStats}
\alias{aqStat}
\title{Calculate user-defined air quality statistics}
\usage{
calcAQStats(mydata, aqstats, period = c("year"), progress = TRUE)

aqStat(
  pollutant = "no2",
  stat = c("mean", "min", "median", "max", "percentile", "limit"),
  avg.time = "hour",
  limit = NA,
  percentile = NA,
  roll = NA,
  roll.avg.time = NA,
  name = NULL
)
}
\arguments{
\item{mydata}{A data frame of time series. Must include a \code{date} field and
any variables required by the provided \code{aqstats}.}

\item{aqstats}{A list of \code{\link[=aqStat]{aqStat()}} objects to calculate.}

\item{period}{A time period to calculate statistics for. Currently, only
\code{"year"} is supported, which will calculate annual statistics.}

\item{progress}{Show a progress bar when large numbers of statistics are
being calculated? Defaults to \code{TRUE}.}

\item{pollutant}{The pollutant (or other named variable) of interest, e.g.,
\code{"no2"}. This should align with a column name in the \code{data.frame} passed to
\code{\link[=calcAQStats]{calcAQStats()}}.}

\item{stat}{The grand statistic to calculate. Simple statistics include
\code{"mean"} (the default), \code{"min"}, \code{"median"} or \code{"max"}. Can also be
\code{"percentile"}, the percentile being defined using the \code{percentile} option.
Finally, can be \code{"limit"}, which compares values against the \code{limit} value
and returns the number of instances where the value exceeds the limit.}

\item{avg.time}{An initial averaging time to calculate \emph{before} the \code{stat} is
calculated; passed directly to \code{\link[=timeAverage]{timeAverage()}}.}

\item{limit}{The limit value, for \code{stat = "limit"}.}

\item{percentile}{The percentile value, for \code{stat = "percentile"}.}

\item{roll}{If an integer value is provided, a rolling mean will be
calculated \emph{after} time averaging and before \emph{stat}. For example, to
calculate an 8-hour rolling mean set \verb{avg.time = "1 hour", roll = 8L}.}

\item{roll.avg.time}{If \code{roll} is specified, users can set \code{roll.avg.time} to
time average the rolling mean values. This could be useful to calculate a
\emph{max} daily running 8 hour mean to compare with an ozone limit, for example.}

\item{name}{Optionally change the output column name for this air quality
statistic.}
}
\description{
\code{\link[=calcAQStats]{calcAQStats()}} calculates user-defined annual air-quality statistics
constructed using \code{\link[=aqStat]{aqStat()}}. This pair of functions allows users to flexibly
define air quality statistics and limits relevant to their local air quality
legislation and quickly track compliance.
}
\examples{
# calculate some UK AQ limits
calcAQStats(
  mydata = openair::mydata,
  aqstats = list(
    # PM10 limits
    aqStat("pm10", "limit", "day", limit = 50),
    aqStat("pm10", "limit", "year", limit = 40),
    # NO2 limits
    aqStat("no2", "limit", "hour", limit = 200),
    aqStat("no2", "limit", "year", limit = 40),
    # O3 limit
    aqStat("o3", "limit", "hour", roll = 8L, limit = 100),
    # SO2 limits
    aqStat("so2", "limit", "15 min", limit = 266),
    aqStat("so2", "limit", "hour", limit = 350),
    aqStat("so2", "limit", "day", limit = 125),
    # CO limits
    aqStat("co", "limit", "hour", roll = 8L, roll.avg.time = "day", limit = 10)
  ),
  period = "year"
)
}
\seealso{
\code{\link[=aqStats]{aqStats()}}, for a simpler but more prescriptive way of calculating
air quality statistics
}
\author{
Jack Davison
}
