% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarPlot.R
\name{calendarPlot}
\alias{calendarPlot}
\title{Plot time series values in a conventional calendar format}
\usage{
calendarPlot(
  mydata,
  pollutant = "nox",
  year = NULL,
  month = NULL,
  annotate = "date",
  statistic = "mean",
  data.thresh = 0,
  percentile = NA,
  cols = "heat",
  limits = NULL,
  lim = NULL,
  col.lim = c("grey30", "black"),
  col.arrow = "black",
  col.na = "white",
  font.lim = c(1, 2),
  cex.lim = c(0.6, 0.9),
  cex.date = 0.6,
  digits = 0,
  labels = NULL,
  breaks = NULL,
  w.shift = 0,
  w.abbr.len = 1,
  remove.empty = TRUE,
  show.year = TRUE,
  key.header = statistic,
  key.footer = pollutant,
  key.position = "right",
  key = TRUE,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame of time series. Must include a \code{date} field and at
least one variable to plot.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \verb{pollutant = "nox". }}

\item{year}{Year to plot e.g. \code{year = 2003}. If not supplied and \code{mydata}
contains more than one year, the first year of the data will be
automatically selected. Manually setting \code{year} to \code{NULL} will use all
available years.}

\item{month}{If only certain month are required. By default the function will
plot an entire year even if months are missing. To only plot certain months
use the \code{month} option where month is a numeric 1:12 e.g. \code{month = c(1, 12)} to only plot January and December.}

\item{annotate}{This option controls what appears on each day of the
calendar. Can be:
\itemize{
\item \code{"date"} --- shows day of the month
\item \code{"wd"} --- shows vector-averaged wind direction
\item \code{"ws"} --- shows vector-averaged wind direction scaled by wind speed
\item \code{"value"} --- shows the daily mean value
}}

\item{statistic}{Statistic passed to \code{\link[=timeAverage]{timeAverage()}}. Note that if \code{statistic \%in\% c("max", "min")} and \code{annotate} is "ws" or "wd", the hour
corresponding to the maximum/minimum concentration of \code{polluant} is used to
provide the associated \code{ws} or \code{wd} and not the maximum/minimum daily \code{ws}
or \code{wd}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{percentile}{The percentile level in percent used when \code{statistic = "percentile"} and when aggregating the data with \code{avg.time}. More than one
percentile level is allowed for \code{type = "default"} e.g. \code{percentile = c(50, 95)}. Not used if \code{avg.time = "default"}.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{limits}{Use this option to manually set the colour scale limits. This
is useful in the case when there is a need for two or more plots and a
consistent scale is needed on each. Set the limits to cover the maximum
range of the data for all plots of interest. For example, if one plot had
data covering 0--60 and another 0--100, then set \code{limits = c(0, 100)}. Note
that data will be ignored if outside the limits range.}

\item{lim}{A threshold value to help differentiate values above and below
\code{lim}. It is used when \code{annotate = "value"}. See next few options for
control over the labels used.}

\item{col.lim}{For the annotation of concentration labels on each day. The
first sets the colour of the text below \code{lim} and the second sets the
colour of the text above \code{lim}.}

\item{col.arrow}{The colour of the annotated wind direction / wind speed
arrows.}

\item{col.na}{Colour to be used to show missing data.}

\item{font.lim}{For the annotation of concentration labels on each day. The
first sets the font of the text below \code{lim} and the second sets the font of
the text above \code{lim}. Note that font = 1 is normal text and font = 2 is
bold text.}

\item{cex.lim}{For the annotation of concentration labels on each day. The
first sets the size of the text below \code{lim} and the second sets the size of
the text above \code{lim}.}

\item{cex.date}{The base size of the annotation text for the date.}

\item{digits}{The number of digits used to display concentration values when
\code{annotate = "value"}.}

\item{labels}{If a categorical scale is defined using \code{breaks}, then \code{labels}
can be used to override the default category labels, e.g., \code{labels = c("good", "bad", "very bad")}. Note there is one less label than break.}

\item{breaks}{If a categorical scale is required then these breaks will be
used. For example, \code{breaks = c(0, 50, 100, 1000)}. In this case "good"
corresponds to values between 0 and 50 and so on. Users should set the
maximum value of \code{breaks} to exceed the maximum data value to ensure it is
within the maximum final range e.g. 100--1000 in this case.}

\item{w.shift}{Controls the order of the days of the week. By default the
plot shows Saturday first (\code{w.shift = 0}). To change this so that it starts
on a Monday for example, set \code{w.shift = 2}, and so on.}

\item{w.abbr.len}{The default (\code{1}) abbreviates the days of the week to a
single letter (e.g., in English, S/S/M/T/W/T/F). \code{w.abbr.len} defines the
number of letters to abbreviate until. For example, \code{w.abbr.len = 3} will
abbreviate "Monday" to "Mon".}

\item{remove.empty}{Should months with no data present be removed? Default is
\code{TRUE}.}

\item{show.year}{If only a single year is being plotted, should the calendar
labels include the year label? \code{TRUE} creates labels like "January-2000",
\code{FALSE} labels just as "January". If multiple years of data are detected,
this option is forced to be \code{TRUE}.}

\item{key.header, key.footer}{Used to control the title of the plot key. By
default, \code{key.header} is the \code{statistic} and \code{key.footer} is the
\code{pollutant}. These are pasted together to form the key title.}

\item{key.position}{Location where the scale key is to plotted. Allowed
arguments currently include \code{"top"}, \code{"right"}, \code{"bottom"} and \code{"left"}.}

\item{key}{Fine control of the scale key via \code{\link[=drawOpenKey]{drawOpenKey()}}. See
\code{\link[=drawOpenKey]{drawOpenKey()}} for further details.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{The following additional arguments are available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{border} sets the border colour of each tile.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function will plot data by month laid out in a conventional calendar
format. The main purpose is to help rapidly visualise potentially complex
data in a familiar way. Users can also choose to show daily mean wind vectors
if wind speed and direction are available.
}
\details{
\code{\link[=calendarPlot]{calendarPlot()}} will plot data in a conventional calendar format, i.e., by
month and day of the week. Daily statistics are calculated using
\code{\link[=timeAverage]{timeAverage()}}, which by default will calculate the daily mean
concentration.

If wind direction is available it is then possible to plot the wind direction
vector on each day. This is very useful for getting a feel for the
meteorological conditions that affect pollutant concentrations. Note that if
hourly or higher time resolution are supplied, then \code{\link[=calendarPlot]{calendarPlot()}} will
calculate daily averages using \code{\link[=timeAverage]{timeAverage()}}, which ensures that wind
directions are vector-averaged.

If wind speed is also available, then setting the option \code{annotate = "ws"}
will plot the wind vectors whose length is scaled to the wind speed. Thus
information on the daily mean wind speed and direction are available.

It is also possible to plot categorical scales. This is useful where, for
example, an air quality index defines concentrations as bands, e.g., "good",
"poor". In these cases users must supply \code{labels} and corresponding \code{breaks}.

Note that is is possible to pre-calculate concentrations in some way before
passing the data to \code{\link[=calendarPlot]{calendarPlot()}}. For example \code{\link[=rollingMean]{rollingMean()}} could be
used to calculate rolling 8-hour mean concentrations. The data can then be
passed to \code{\link[=calendarPlot]{calendarPlot()}} and \code{statistic = "max"} chosen, which will plot
maximum daily 8-hour mean concentrations.
}
\examples{
# basic plot
calendarPlot(mydata, pollutant = "o3", year = 2003)

# show wind vectors
calendarPlot(mydata, pollutant = "o3", year = 2003, annotate = "wd")
\dontrun{
# show wind vectors scaled by wind speed and different colours
calendarPlot(mydata,
  pollutant = "o3", year = 2003, annotate = "ws",
  cols = "heat"
)

# show only specific months with selectByDate
calendarPlot(selectByDate(mydata, month = c(3, 6, 10), year = 2003),
  pollutant = "o3", year = 2003, annotate = "ws", cols = "heat"
)

# categorical scale example
calendarPlot(mydata,
  pollutant = "no2", breaks = c(0, 50, 100, 150, 1000),
  labels = c("Very low", "Low", "High", "Very High"),
  cols = c("lightblue", "green", "yellow", "red"), statistic = "max"
)

# UK daily air quality index
pm10.breaks <- c(0, 17, 34, 50, 59, 67, 75, 84, 92, 100, 1000)
calendarPlot(mydata, "pm10",
  year = 1999, breaks = pm10.breaks,
  labels = c(1:10), cols = "daqi", statistic = "mean", key.header = "DAQI"
)
}
}
\seealso{
Other time series and trend functions: 
\code{\link{TheilSen}()},
\code{\link{runRegression}()},
\code{\link{smoothTrend}()},
\code{\link{timePlot}()},
\code{\link{timeProp}()},
\code{\link{timeVariation}()},
\code{\link{trendLevel}()}
}
\author{
David Carslaw
}
\concept{time series and trend functions}
