% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\title{Start or continue a chat conversation}
\usage{
chat(message, chatlog_id = ".__CURRENTCHAT__", output = "message_to_console")
}
\arguments{
\item{message}{The message to be added to the chat conversation}

\item{chatlog_id}{The ID of the chat conversation to start or continue. Default is "__CURRENTCHAT__".}

\item{output}{The output format of the response. Default is "message_to_console".
Valid options are "message_to_console", "message", or "response_object".}
}
\value{
Depending on the value of the 'output' argument, this function returns one of the following:
* "message_to_console": a message containing the response text to be printed to the console.
* "message": the response text as a character vector.
* "response_object": the full response object from the ChatGPT API.
}
\description{
This function starts or continues a chat conversation and adds the user's message to the conversation.
If the conversation does not exist, a new one will be initiated.
}
