% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourOpts.R
\name{colourOpts}
\alias{colourOpts}
\alias{colorOpts}
\title{Define \code{cols} options for \code{openair} plots}
\usage{
colourOpts(scheme = "default", alpha = 1, begin = 0, end = 1, direction = 1)

colorOpts(scheme = "default", alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{scheme}{Any one of the pre-defined \code{openair} schemes (e.g.,
\code{"increment"}) or a user-defined palette (e.g., \code{c("red", "orange", "gold")}). See \code{\link[=openColours]{openColours()}} for a full list of available schemes.}

\item{alpha}{The alpha transparency level (between \code{0} and \code{1}) for the
colours. \code{0} is fully transparent, and \code{1} is fully opaque.}

\item{begin, end}{For sequential schemes, the fraction (between \code{0} and \code{1})
of the colour scheme to use. For example, if \code{begin = 0.2} and \code{end = 0.8},
only the middle 60\% of the colour scheme will be used. This can be useful
for avoiding very light or dark colours at the ends of schemes.}

\item{direction}{The order of the colours. \code{1} is the default and gives the
normal order. \code{-1} will reverse the order of the colours.}
}
\value{
A list of options that can be passed to the \code{cols} argument of
plotting functions like \code{\link[=polarPlot]{polarPlot()}}.
}
\description{
This function provides a convenient way to set options for customising
colours in openair plots. It returns a list of options that can be passed to
the \code{cols} argument of most functions. All arguments are the same as
\code{\link[=openColours]{openColours()}}, but passing \code{\link[=colourOpts]{colourOpts()}} allows each function to set an
appropriate \code{n} value and overwrite certain options if required.
\code{colourOpts()} and \code{colorOpts()} are synonyms.
}
\examples{
trendLevel(
  mydata,
  "no2",
  cols = colourOpts("viridis", direction = 1, alpha = 0.5)
)
}
