% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_code.R
\name{explain_code}
\alias{explain_code}
\title{Explain a Piece of R Code}
\usage{
explain_code(code)
}
\arguments{
\item{code}{A character string containing the R code to be processed.}
}
\description{
TODO: FUNCTION_DESCRIPTION
}
