% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_completions.R
\name{get_completions}
\alias{get_completions}
\title{Title}
\usage{
get_completions(
  prompt,
  api_key,
  api_base_url = "https://api.openai.com/v1/chat/completions",
  model = "gpt-3.5-turbo",
  max_tries = 1,
  ...
)
}
\value{
A list object containing response of system.
}
\description{
Title
}
