% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importWAQN.R
\name{importWAQN}
\alias{importWAQN}
\title{Welsh Air Quality Network data import for openair}
\usage{
importWAQN(site = "card", year = 2018, pollutant = "all")
}
\arguments{
\item{site}{Site code of the WAQN site to import e.g. "card" is Cardiff Centre. 
Several sites can be imported with \code{site = c("card", "cae6")}
  --- to import Cardiff Centre and Hafod-yr-ynys Roadside for example.}

\item{year}{Year or years to import. To import a sequence of years from 1990
to 2000 use \code{year = 1990:2000}. To import several specfic years use
\code{year = c(1990, 1995, 2000)} for example.}

\item{pollutant}{Pollutants to import. If omitted will import all pollutants
from a site. To import only NOx and NO2 for example use \code{pollutant =
c("nox", "no2")}.}
}
\value{
Returns a data frame of hourly mean values with date in POSIXct
  class and time zone GMT.
}
\description{
Function for importing hourly mean Welsh Air Quality Network (WAQN)
archive data files for use with the \code{openair} package. Files are
imported from a remote server operated by Ricardo that provides air quality data
files as R data objects.
}
\details{
The \code{importWAQN} function has been written to make it easy to import
data from the Welsh Air Quality Network (WAQN) ---
\url{https://airquality.gov.wales/}. Ricardo have provided
.RData files (R workspaces) of all individual sites and years for the WAQN.
These files are updated on a daily basis. This approach requires a link to
the Internet to work.

For a list of up to date site codes and site information, see
\code{\link{importMeta}} and in particular \code{importMeta(source =
"waqn")}.

The site codes and pollutant names can be upper or lower case. The function
will issue a warning when data less than six months old is downloaded, which
may not be ratified.

The function also returns wind speed (ws) and direction (wd) for more recent
years derived from WRF (regional meteorological model).
}
\examples{

## see what sites are available
\dontrun{meta <- importMeta("waqn")}

## import all pollutants from Cardiff Centre
\dontrun{cardiff <- importWAQN(site = "card", year = 2010:2018)}

## import all pollutants from two sites for 2018
\dontrun{all <- importWAQN(site = site = c("card", "cae6"), year = 2018)}


}
\seealso{
See \code{\link{importAURN}} for data elsewhere in the UK and
  \code{\link{importKCL}} for importing comprehensive data in and around
  London.
}
\author{
David Carslaw and Trevor Davies
}
\keyword{methods}
