% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messages}
\alias{messages}
\title{Extract choices/messages from OpenAI API response}
\usage{
messages(response)
}
\arguments{
\item{response}{a list object representing the HTTP response}
}
\value{
a dataframe representing the choices/messages from the response
}
\description{
This function extracts the choices/messages from the HTTP response of an API call to the
OpenAI chat completions endpoint.
}
\author{
Ulrich Matter umatter@protonmail.com
}
