% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Plot air quality trends in a conventional 'calendar' format}
\usage{
plot_calendar(
  data,
  pollutant,
  statistic = "mean",
  data.thresh = 0,
  percentile = NA,
  year = NULL,
  month = NULL,
  discretise = NULL,
  windflow = FALSE,
  cols = "heat",
  limits = NULL,
  show_year = TRUE,
  w_shift = 0,
  w_abbr_length = 1,
  auto_text = TRUE,
  facet_opts = openair::facet_opts(axes = "all_x"),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The name of the pollutant to summarise.}

\item{statistic}{Statistic passed to \code{\link[=timeAverage]{timeAverage()}}. Note that if \code{statistic \%in\% c("max", "min")} and \code{annotate} is "ws" or "wd", the hour
corresponding to the maximum/minimum concentration of \code{polluant} is used to
provide the associated \code{ws} or \code{wd} and not the maximum/minimum daily \code{ws}
or \code{wd}.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{year}{Year to plot e.g. \code{year = 2003}. If not supplied all data
potentially spanning several years will be plotted.}

\item{month}{If only certain month are required. By default the function will
plot an entire year even if months are missing. To only plot certain months
use the \code{month} option where month is a numeric 1:12 e.g. \code{month = c(1, 12)} to only plot January and December.}

\item{discretise}{Options to discretise the continuous scale into discrete
bins. Providing an integer will split the scale into that number of
quantiles. Alternatively, \code{\link[=discretise_opts]{discretise_opts()}} allows for extensive
customisation (e.g., defining specific bin widths, breaks, labels, and so
on).}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{limits}{The limits of the colour scale, in the form \code{c(min, max)}. An
\code{NA} uses the relevant limit of the input data. Not applied if the data is
discretised.}

\item{show_year}{If only a single year is being plotted, should the calendar
labels include the year label? \code{TRUE} creates labels like "January-2000",
\code{FALSE} labels just as "January". If multiple years of data are detected,
this option is forced to be \code{TRUE}.}

\item{w_shift}{Controls the order of the days of the week. By default the
plot shows Saturday first (\code{w_shift = 0}). To change this so that it starts
on a Monday for example, set \code{w_shift = 2}, and so on.}

\item{w_abbr_length}{Defines the number of letters to abbreviate the weekday
to. The default (\code{1}) abbreviates the days of the week to a single letter
(e.g., in English, S/S/M/T/W/T/F), \code{w.abbr.len = 3} would abbreviate
"Monday" to "Mon", and so on.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Not currently used.}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2} plot, or \code{data.frame} if \code{plot = FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function will plot data by month laid out in a conventional calendar
format. The main purpose is to help rapidly visualise potentially complex
data in a familiar way. Users can also choose to show daily mean wind vectors
if wind speed and direction are available.

Note that is is possible to pre-calculate concentrations in some way before
passing the data to \code{\link[=plot_calendar]{plot_calendar()}}. For example \code{\link[=rollingMean]{rollingMean()}} could be
used to calculate rolling 8-hour mean concentrations. The data can then be
passed to \code{\link[=plot_calendar]{plot_calendar()}} and \code{statistic = "max"} chosen, which will plot
maximum daily 8-hour mean concentrations.
}
\examples{
# basic plot
plot_calendar(mydata, pollutant = "o3", year = 2003)

# show wind vectors
calendarPlot(data, pollutant = "o3", year = 2003, annotate = "wd")
\dontrun{
# show wind vectors scaled by wind speed and different colours
calendarPlot(data,
  pollutant = "o3", year = 2003, annotate = "ws",
  cols = "heat"
)

# show only specific months with selectByDate
calendarPlot(selectByDate(data, month = c(3, 6, 10), year = 2003),
  pollutant = "o3", year = 2003, annotate = "ws", cols = "heat"
)

# categorical scale example
calendarPlot(data,
  pollutant = "no2", breaks = c(0, 50, 100, 150, 1000),
  labels = c("Very low", "Low", "High", "Very High"),
  cols = c("lightblue", "green", "yellow", "red"), statistic = "max"
)

# UK daily air quality index
pm10.breaks <- c(0, 17, 34, 50, 59, 67, 75, 84, 92, 100, 1000)
calendarPlot(data, "pm10",
  year = 1999, breaks = pm10.breaks,
  labels = c(1:10), cols = "daqi", statistic = "mean", key.header = "DAQI"
)
}
}
\seealso{
The legacy \code{\link[=calendarPlot]{calendarPlot()}} function

Other ggplot2 time series and trend functions: 
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_smooth}()}
}
\author{
Jack Davison

David Carslaw
}
\concept{ggplot2 time series and trend functions}
