% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot an air quality timeseries heatmap chart}
\usage{
plot_heatmap(
  data,
  pollutant = "nox",
  x = "month",
  y = "hour",
  type = "year",
  n.levels = c(10, 10, 4),
  cols = "default",
  auto.text = TRUE,
  labels = NA,
  breaks = NA,
  statistic = c("mean", "max", "min", "median", "frequency", "sum", "sd", "percentile"),
  percentile = 95,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{The openair data frame to use to generate the \code{\link[=trendLevel]{trendLevel()}}
plot.}

\item{pollutant}{The name of the data series in \code{mydata} to sample to produce
the \code{\link[=trendLevel]{trendLevel()}} plot.}

\item{x, y, type}{The name of the data series to use as the \code{\link[=trendLevel]{trendLevel()}}
x-axis, y-axis or conditioning variable, passed to \code{\link[=cutData]{cutData()}}. These are
used before applying \code{statistic}. \code{\link[=trendLevel]{trendLevel()}} does not allow duplication
in \code{x}, \code{y} and \code{type} options.}

\item{n.levels}{The number of levels to split \code{x}, \code{y} and \code{type} data into
if numeric. The default, \code{c(10, 10, 4)}, cuts numeric \code{x} and \code{y} data into
ten levels and numeric \code{type} data into four levels. This option is ignored
for date conditioning and factors. If less than three values are supplied,
three values are determined by recursion; if more than three values are
supplied, only the first three are used.}

\item{cols}{The colour set to use to colour the \code{\link[=trendLevel]{trendLevel()}} surface.
\code{cols} is passed to \code{\link[=openColours]{openColours()}} for evaluation.}

\item{auto.text}{Automatic routine text formatting. \code{auto.text = TRUE} passes
common \code{lattice} labelling terms (e.g. \code{xlab} for the x-axis, \code{ylab} for
the y-axis and \code{main} for the title) to the plot via \code{\link[=quickText]{quickText()}} to
provide common text formatting.  The alternative \code{auto.text = FALSE} turns
this option off and passes any supplied labels to the plot without
modification.}

\item{breaks, labels}{If a categorical colour scale is required then \code{breaks}
should be specified. These should be provided as a numeric vector, e.g.,
\code{breaks = c(0, 50, 100, 1000)}. Users should set the maximum value of
\code{breaks} to exceed the maximum data value to ensure it is within the
maximum final range, e.g., 100--1000 in this case. Labels will
automatically be generated, but can be customised by passing a character
vector to \code{labels}, e.g., \code{labels = c("good", "bad", "very bad")}. In this
example, \code{0 - 50} will be \code{"good"} and so on. Note there is one less label
than break.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option. Functions can also be sent directly via
\code{statistic}; see 'Details' for more information.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}}.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
The \code{\link[=plot_heatmap]{plot_heatmap()}} function provides a way of rapidly showing a large
amount of data in a condensed form. In one plot, the variation in the
concentration of one pollutant can to shown as a function of three other
categorical properties. The default version of the plot uses y = hour of day,
x = month of year and type = year to provide information on trends, seasonal
effects and diurnal variations. However, x, y and type and summarising
statistics can all be modified to provide a range of other similar plots.
}
\examples{
# basic use
# default statistic = "mean"
plot_heatmap(mydata, pollutant = "nox")

\dontrun{
# example with categorical scale
plot_heatmap(
  mydata,
  pollutant = "no2",
  statistic = "max",
  breaks = c(0, 50, 100, 500),
  labels = c("low", "medium", "high"),
  cols = c("forestgreen", "yellow", "red")
)
}
}
\seealso{
Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_smooth}()},
\code{\link{plot_trend_theilsen}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{ggplot2 time series and trend functions}
