% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_bars.R
\name{plot_trend_bars}
\alias{plot_trend_bars}
\title{Plot an air quality timeseries chart with (stacked) bars}
\usage{
plot_trend_bars(
  data,
  pollutant = "nox",
  proportion = "default",
  avg.time = "day",
  type = "default",
  cols = "Set1",
  normalise = FALSE,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the fields \code{date}, \code{pollutant} and,
optionally, a splitting variable \code{proportion}.}

\item{pollutant}{Name of the pollutant to plot contained in \code{data}.}

\item{proportion}{The splitting variable that makes up the bars in the bar
chart e.g. \code{proportion = "cluster"} if the output from \code{polarCluster()} is
being analysed. If \code{proportion} is a numeric variable it is split into 4
quantiles (by default) by \code{cutData}. If \code{proportion} is a factor or
character variable then the categories are used directly. Can also be
\code{"default"}, which just plots an uncategorised bar chart.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be \code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in which case
3-month seasonal values are calculated with spring defined as March, April,
May and so on.

Note that \code{avg.time} when used in \code{\link[=plot_trend_bars]{plot_trend_bars()}} should be greater
than the time gap in the original data. For example, \code{avg.time = "day"} for
hourly data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

\code{type} must be of length one.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Extra arguments passed to \code{\link[=cutData]{cutData()}}.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the plotting
of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and \code{\link[=trajCluster]{trajCluster()}} that
consider local and regional (back trajectory) cluster analysis respectively.
However, the function has more general use for understanding time series
data.
}
\examples{
# monthly plot of SO2 showing the contribution by wind sector
plot_trend_bars(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
\code{\link[=timeProp]{timeProp()}}, the original \code{lattice} equivalent

Other ggplot2 time series and trend functions: 
\code{\link{plot_trend_smooth}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{ggplot2 cluster analysis functions}
\concept{ggplot2 time series and trend functions}
