% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_bars.R
\name{plot_trend_bars}
\alias{plot_trend_bars}
\title{Plot air quality trends as stacked bar charts}
\usage{
plot_trend_bars(
  data,
  pollutant,
  group = NULL,
  type = NULL,
  avg.time = "month",
  cols = "tol",
  normalise = FALSE,
  auto_text = TRUE,
  facet_opts = openair::facet_opts(),
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{pollutant}{The name of the pollutant to summarise.}

\item{group}{The splitting variable that makes up the bars in the bar chart
e.g. \code{group = "cluster"} if the output from \code{polarCluster} is being
analysed. If \code{group} is a numeric variable it is split into 4 quantiles (by
default) by \code{\link[=cutData]{cutData()}}. If \code{group} is a factor or character variable then
the categories are used directly.}

\item{type}{The name of the data series to use as the conditioning variable,
passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with maximum length
2, which creates a 2D grid of plots.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be \code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in which case
3-month seasonal values are calculated with spring defined as March, April,
May and so on.

Note that \code{avg.time} when used in \code{timeProp} should be greater than the
time gap in the original data. For example, \code{avg.time = "day"} for hourly
data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Additional arguments passed to \code{\link[=cutData]{cutData()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the
plotting of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and
\code{\link[=trajCluster]{trajCluster()}} that consider local and regional (back trajectory) cluster
analysis respectively. However, the function has more general use for
understanding time series data. In order to plot time series in this way,
some sort of time aggregation is needed, which is controlled by the option
\code{avg.time}.
}
\examples{
# monthly plot of SO2 showing the contribution by wind sector
timeProp(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
the legacy \code{\link[=timeProp]{timeProp()}} function

Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_smooth}()},
\code{\link{plot_trend_theilsen}()}

Other cluster analysis functions: 
\code{\link{polarCluster}()},
\code{\link{timeProp}()},
\code{\link{trajCluster}()}
}
\author{
Jack Davison

David Carslaw
}
\concept{cluster analysis functions}
\concept{ggplot2 time series and trend functions}
