% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_lines.R
\name{plot_trend_lines}
\alias{plot_trend_lines}
\title{Plot an air quality timeseries line chart}
\usage{
plot_trend_lines(
  data,
  pollutant = "nox",
  group = FALSE,
  normalise = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  date.pad = FALSE,
  type = "default",
  cols = "brewer1",
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame of time series. Must include a \code{date} field and at
least one variable to plot.}

\item{pollutant}{Name of variable to plot. Two or more pollutants can be
plotted, in which case a form like \code{pollutant = c("nox", "co")} should be
used.}

\item{group}{If more than one pollutant is chosen, should they all be plotted
on the same graph together? The default is \code{FALSE}, which means they are
plotted in separate panels with their own scales. If \code{TRUE} then they are
plotted on the same plot with the same scale.}

\item{normalise}{Should variables be normalised? The default is is not to
normalise the data. \code{normalise} can take two values, either \code{"mean"} or a
string representing a date in UK format e.g. "1/1/1998" (in the format
dd/mm/YYYY). If \code{normalise = "mean"} then each time series is divided by
its mean value.  If a date is chosen, then values at that date are set to
100 and the rest of the data scaled accordingly. Choosing a date (say at
the beginning of a time series) is very useful for showing how trends
diverge over time. Setting \code{group = TRUE} is often useful too to show all
time series together in one panel.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option --- see below. Not used if \code{avg.time = "default"}.}

\item{percentile}{The percentile level in percent used when \code{statistic = "percentile"} and when aggregating the data with \code{avg.time}. More than one
percentile level is allowed for \code{type = "default"} e.g. \code{percentile = c(50, 95)}. Not used if \code{avg.time = "default"}.}

\item{date.pad}{Should missing data be padded-out? This is useful where a
data frame consists of two or more "chunks" of data with time gaps between
them. By setting \code{date.pad = TRUE} the time gaps between the chunks are
shown properly, rather than with a line connecting each chunk. For
irregular data, set to \code{FALSE}. Note, this should not be set for \code{type}
other than \code{default}.}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Other arguments passed to \code{\link[=cutData]{cutData()}}.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
The \code{\link[=plot_trend_lines]{plot_trend_lines()}} is the basic time series plotting function in
\code{openair}. Its purpose is to make it quick and easy to plot time series for
pollutants and other variables. The other purpose is to plot potentially many
variables together in as compact a way as possible.
}
\details{
The function is flexible enough to plot more than one variable at once. If
more than one variable is chosen plots it can either show all variables on
the same plot (with different line types) \emph{on the same scale}, or (if \code{group = FALSE}) each variable in its own panels with its own scale.

The general preference is not to plot two variables on the same graph with
two different y-scales. It can be misleading to do so and difficult with more
than two variables. If there is in interest in plotting several variables
together that have very different scales, then it can be useful to normalise
the data first, which can be down be setting the \code{normalise} option.
}
\examples{

# basic use, single pollutant
plot_trend_lines(mydata, pollutant = "nox")

\dontrun{
# two pollutants in separate panels
plot_trend_lines(mydata, pollutant = c("nox", "no2"))

# two pollutants in the same panel with the same scale
plot_trend_lines(mydata, pollutant = c("nox", "no2"), group = TRUE)

# alternative by normalising concentrations and plotting on the same scale
plot_trend_lines(
  mydata,
  pollutant = c("nox", "co", "pm10", "so2"),
  group = TRUE,
  avg.time = "year",
  normalise = "1/1/1998"
)

# daily mean O3
plot_trend_lines(mydata, pollutant = "o3", avg.time = "day")

# daily mean O3 ensuring each day has data capture of at least 75\%
plot_trend_lines(mydata, pollutant = "o3", avg.time = "day", data.thresh = 75)

# 2-week average of O3 concentrations
plot_trend_lines(mydata, pollutant = "o3", avg.time = "2 week")
}

}
\seealso{
Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_smooth}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{ggplot2 time series and trend functions}
