% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_smooth.R
\name{plot_trend_smooth}
\alias{plot_trend_smooth}
\title{Plot an air quality timeseries chart with a non-parametric smooth trend}
\usage{
plot_trend_smooth(
  mydata,
  pollutant = "nox",
  avg.time = "month",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  k = NULL,
  deseason = FALSE,
  simulate = FALSE,
  n = 200,
  autocor = FALSE,
  type = "default",
  cols = "brewer1",
  auto.text = TRUE,
  ci = TRUE,
  alpha = 0.2,
  plot = TRUE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame of time series. Must include a \code{date} field and at
least one variable to plot.}

\item{pollutant}{Name of variable to plot. Two or more pollutants can be
plotted, in which case a form like \code{pollutant = c("nox", "co")} should be
used.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{Statistic used for calculating monthly values. Default is
\code{"mean"}, but can also be \code{"percentile"}. See \code{\link[=timeAverage]{timeAverage()}} for more
details.}

\item{percentile}{Percentile value(s) to use if \code{statistic = "percentile"} is
chosen. Can be a vector of numbers e.g. \code{percentile = c(5, 50, 95)} will
plot the 5th, 50th and 95th percentile values together on the same plot.}

\item{k}{This is the smoothing parameter used by the \code{\link[mgcv:gam]{mgcv::gam()}} function
in package \code{mgcv}. By default it is not used and the amount of smoothing is
optimised automatically. However, sometimes it is useful to set the
smoothing amount manually using \code{k}.}

\item{deseason}{Should the data be de-deasonalized first? If \code{TRUE} the
function \code{stl} is used (seasonal trend decomposition using loess). Note
that if \code{TRUE} missing data are first imputed using a Kalman filter and
Kalman smooth.}

\item{simulate}{Should simulations be carried out to determine the
Mann-Kendall tau and p-value. The default is \code{FALSE}. If \code{TRUE}, bootstrap
simulations are undertaken, which also account for autocorrelation.}

\item{n}{Number of bootstrap simulations if \code{simulate = TRUE}.}

\item{autocor}{Should autocorrelation be considered in the trend uncertainty
estimates? The default is \code{FALSE}. Generally, accounting for
autocorrelation increases the uncertainty of the trend estimate sometimes
by a large amount.}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

Type can be up length two e.g. \code{type = c("season", "weekday")} will produce
a 2x2 plot split by season and day of the week. Note, when two types are
provided the first forms the columns and the second the rows.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{ci}{Should confidence intervals be plotted? The default is \code{TRUE}.}

\item{alpha}{The alpha transparency of shaded confidence intervals - if
plotted. A value of 0 is fully transparent and 1 is fully opaque.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{progress}{Show a progress bar when many groups make up \code{type}? Defaults
to \code{TRUE}.}

\item{...}{Extra arguments passed to \code{\link[=cutData]{cutData()}}.}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
The \code{\link[=plot_trend_smooth]{plot_trend_smooth()}} function provides a flexible way of estimating the
trend in the concentration of a pollutant or other variable. Monthly mean
values are calculated from an hourly (or higher resolution) or daily time
series. There is the option to deseasonalise the data if there is evidence of
a seasonal cycle.
}
\details{
\code{\link[=plot_trend_smooth]{plot_trend_smooth()}} uses a Generalized Additive Model (GAM) from the
\code{\link[mgcv:gam]{mgcv::gam()}} package to find the most appropriate level of smoothing. The
function is particularly suited to situations where trends are not monotonic
(see discussion with \code{\link[=TheilSen]{TheilSen()}} for more details on this). The
\code{\link[=plot_trend_smooth]{plot_trend_smooth()}} function is particularly useful as an exploratory
technique e.g. to check how linear or non-linear trends are.

95\% confidence intervals are shown by shading. Bootstrap estimates of the
confidence intervals are also available through the \code{simulate} option.
Residual resampling is used.
}
\examples{
# trend plot for nox
plot_trend_smooth(mydata, pollutant = "nox")

\dontrun{
# trend plot by each of 8 wind sectors
plot_trend_smooth(mydata, pollutant = "o3", type = "wd", ylab = "o3 (ppb)")

# several pollutants, no plotting symbol
plot_trend_smooth(mydata, pollutant = c("no2", "o3", "pm10", "pm25"))

# percentiles
plot_trend_smooth(
  mydata,
  pollutant = "o3",
  statistic = "percentile",
  percentile = 95
)
}
}
\seealso{
\code{\link[=smoothTrend]{smoothTrend()}}, the original \code{lattice} equivalent

Other ggplot2 time series and trend functions: 
\code{\link{plot_calendar}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_trend_bars}()},
\code{\link{plot_trend_lines}()},
\code{\link{plot_trend_theilsen}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{ggplot2 time series and trend functions}
