% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_scatter.R
\name{plot_xy_scatter}
\alias{plot_xy_scatter}
\alias{plot_xy_bins}
\title{Plot the relationship between two (or more) variables using scatter plot
variations}
\usage{
plot_xy_scatter(
  data,
  x,
  y,
  z = NULL,
  group_col = NULL,
  group_shp = NULL,
  type = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  discretise = NULL,
  cols = NULL,
  windflow = FALSE,
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)

plot_xy_bins(
  data,
  x,
  y,
  type = NULL,
  avg.time = "default",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  method = c("bin", "hexbin", "density"),
  binwidth = NULL,
  bins = 30,
  breaks = NULL,
  adjust = 1,
  dist = 0.02,
  discretise = NULL,
  cols = "viridis",
  facet_opts = openair::facet_opts(),
  auto_text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x, y}{The columns of \code{data} to use as the x and y variables.}

\item{z}{The column of \code{data} to colour by. This is not passed though
\code{\link[=cutData]{cutData()}}, so a colour}

\item{group_col, group_shp}{Columns of \code{data} to use to condition the data
within each panel, passed to \code{\link[=cutData]{cutData()}}. \code{group_col} maps to the colour
aesthetic and \code{group_shp} to the marker shape aesthetic.}

\item{type}{The column(s) of \code{data} to use to split the plot into separate
facets, passed to \code{\link[=cutData]{cutData()}}. \code{type} may be \code{NULL} or a vector with
maximum length 2, which creates a 2D grid of plots.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option --- see below. Not used if \code{avg.time = "default"}.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{discretise}{Options to discretise the continuous scale into discrete
bins. Providing an integer will split the scale into that number of
quantiles. Alternatively, \code{\link[=discretise_breaks]{discretise_breaks()}} allows for extensive
customisation (e.g., defining specific bin widths, breaks, labels, and so
on).}

\item{cols}{The colour palette to use. See \code{\link[=openColours]{openColours()}}.}

\item{windflow}{If \code{TRUE}, arrows showing the average magnitude and direction
of wind within each bin will be shown. Note that this requires the columns
\code{"ws"} and \code{"wd"} to be present in \code{data}. Alternatively, \code{\link[=windflow_opts]{windflow_opts()}}
allows for greater customisation of the arrows (e.g., wind speed ranges,
relative size of arrows, arrow appearance).}

\item{facet_opts}{A list of options to help control how 'facets' (different
panels accessed through \code{type}) behave, e.g., whether different panels
should share the same scales. Use \code{\link[=facet_opts]{facet_opts()}} as a convenient way to
provide all necessary options.}

\item{auto_text}{Perform automatic text formatting on common character
strings? \code{TRUE} will automatically, for example, subscript the 'x' in NOx
in axis labels. \code{FALSE} will leave text unchanged, but users may refine it
with, e.g., \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:guides]{ggplot2::guides()}}.}

\item{plot}{Should a plot be produced? \code{FALSE} will instead return the
plotting data, and can be useful when analysing data to extract plot
components and plotting them in other ways.}

\item{...}{Passed to \code{\link[=cutData]{cutData()}}.}

\item{method}{One of:
\itemize{
\item \code{"bin"}, which bins the data using square bins
\item \code{"hexbin"}, which uses regular hexagonal binning, or
\item \code{"density"}, which fits a 2D kernel density estimation.
}}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that takes x after scale transformation as input and
returns a single numeric value. When specifying a function along with a
grouping structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{breaks}{Alternatively, you can supply a numeric vector giving
the bin boundaries. Overrides \code{binwidth}, \code{bins}, \code{center},
and \code{boundary}. Can also be a function that takes group-wise values as input and returns bin boundaries.}

\item{adjust}{Adjustment factor to kernel density bandwidth. Used in
\code{\link[=plot_xy_bins]{plot_xy_bins()}} when \code{method = "density"}.}

\item{dist}{How far from the original data predictions should be made; passed
to \code{\link[mgcv:exclude.too.far]{mgcv::exclude.too.far()}}. Values should be between 0 and 1. Used in
\code{\link[=plot_xy_bins]{plot_xy_bins()}} when \code{method = "density"}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This family of functions allows for two or more pollutants to be directly
compared using a relational diagram, with convenient time averaging. This
family of plotting functions encompasses:
\itemize{
\item \code{\link[=plot_xy_scatter]{plot_xy_scatter()}}, which plots a conventional scatter plot
\item \code{\link[=plot_xy_bins]{plot_xy_bins()}}, which bins large amount of data either in square or hexagonal bins
}
}
\examples{
dat2004 <- selectByDate(mydata, year = 2004)

# basic use, single pollutant
plot_xy_scatter(dat2004, x = "nox", y = "no2")

\dontrun{
# scatterPlot by year
plot_xy_scatter(mydata, x = "nox", y = "no2", type = "year")
}

# examine group and type arguments
plot_xy_scatter(
  dat2004,
  x = "nox",
  y = "no2",
  group_col = "weekday",
  group_shp = "season",
  type = "weekend"
)

# example of the use of continuous where colour is used to show
# different levels of a third (numeric) variable
# plot daily averages and choose a filled plot symbol (pch = 16)
# select only 2004
\dontrun{

plot_xy_scatter(dat2004, x = "nox", y = "no2", z = "co", avg.time = "day")

# also works with the x-axis in date format (alternative to plot_trend_*())
plot_xy_scatter(
  mydata,
  x = "date",
  y = "no2",
  avg.time = "month"
)

# multiple types and grouping variable and continuous colour scale
plot_xy_scatter(
  mydata,
  x = "nox",
  y = "no2",
  z = "o3",
  type = c("season", "weekend")
)

# use binning
plot_xy_bins(mydata, x = "nox", y = "no2", method = "bin")
plot_xy_bins(mydata, x = "nox", y = "no2", method = "hexbin")
plot_xy_bins(mydata, x = "nox", y = "no2", method = "density")
}
}
\seealso{
the legacy \code{\link[=scatterPlot]{scatterPlot()}} function
}
