% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windRose.R
\name{pollutionRose}
\alias{pollutionRose}
\title{Pollution rose variation of the traditional wind rose plot}
\usage{
pollutionRose(
  mydata,
  pollutant = "nox",
  key.title = pollutant,
  key.position = "right",
  breaks = 6,
  paddle = FALSE,
  seg = 0.9,
  normalise = FALSE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing fields \code{ws} and \code{wd}}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox"}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{breaks}{Most commonly, the number of break points for pollutant
concentrations. The default, 6, attempts to breaks the supplied data at
approximately 6 sensible break points. However, \code{breaks} can also be
used to set specific break points. For example, the argument \code{breaks = c(0, 1, 10, 100)} breaks the data into segments <1, 1-10, 10-100, >100.}

\item{paddle}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE} plots rose using 'paddle'
style spokes. If \code{FALSE} plots rose using 'wedge' style spokes.}

\item{seg}{\code{seg} determines with width of the segments. For example, \code{seg = 0.5} will produce segments 0.5 * \code{angle}.}

\item{normalise}{If \code{TRUE} each wind direction segment is normalised to equal
one. This is useful for showing how the concentrations (or other
parameters) contribute to each wind sector when the proportion of time the
wind is from that direction is low. A line showing the probability that the
wind directions is from a particular wind sector is also shown.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{
  Arguments passed on to \code{\link[=windRose]{windRose}}
  \describe{
    \item{\code{ws}}{Name of the column representing wind speed.}
    \item{\code{wd}}{Name of the column representing wind direction.}
    \item{\code{ws2,wd2}}{The user can supply a second set of wind speed and wind
direction values with which the first can be compared. See
\code{\link[=pollutionRose]{pollutionRose()}} for more details.}
    \item{\code{ws.int}}{The Wind speed interval. Default is 2 m/s but for low met masts
with low mean wind speeds a value of 1 or 0.5 m/s may be better.}
    \item{\code{angle}}{Default angle of \dQuote{spokes} is 30. Other potentially useful
angles are 45 and 10. Note that the width of the wind speed interval may
need adjusting using \code{width}.}
    \item{\code{calm.thresh}}{By default, conditions are considered to be calm when the
wind speed is zero. The user can set a different threshold for calms be
setting \code{calm.thresh} to a higher value. For example, \code{calm.thresh = 0.5}
will identify wind speeds \strong{below} 0.5 as calm.}
    \item{\code{bias.corr}}{When \code{angle} does not divide exactly into 360 a bias is
introduced in the frequencies when the wind direction is already supplied
rounded to the nearest 10 degrees, as is often the case. For example, if
\code{angle = 22.5}, N, E, S, W will include 3 wind sectors and all other angles
will be two. A bias correction can made to correct for this problem. A
simple method according to Applequist (2012) is used to adjust the
frequencies.}
    \item{\code{grid.line}}{Grid line interval to use. If \code{NULL}, as in default, this is
assigned based on the available data range. However, it can also be forced
to a specific value, e.g. \code{grid.line = 10}. \code{grid.line} can also be a list
to control the interval, line type and colour. For example \code{grid.line = list(value = 10, lty = 5, col = "purple")}.}
    \item{\code{width}}{For paddle = TRUE, the adjustment factor for width of wind speed
intervals. For example, width = 1.5 will make the paddle width 1.5 times
wider.}
    \item{\code{max.freq}}{Controls the scaling used by setting the maximum value for
the radial limits. This is useful to ensure several plots use the same
radial limits.}
    \item{\code{dig.lab}}{The number of significant figures at which scientific number
formatting is used in break point and key labelling. Default 5.}
    \item{\code{include.lowest}}{Logical. If \code{FALSE} (the default), the first interval
will be left exclusive and right inclusive. If \code{TRUE}, the first interval
will be left and right inclusive. Passed to the \code{include.lowest} argument
of \code{\link[=cut]{cut()}}.}
    \item{\code{statistic}}{The \code{statistic} to be applied to each data bin in the plot.
Options currently include \dQuote{prop.count}, \dQuote{prop.mean} and
\dQuote{abs.count}. The default \dQuote{prop.count} sizes bins according to
the proportion of the frequency of measurements.  Similarly,
\dQuote{prop.mean} sizes bins according to their relative contribution to
the mean. \dQuote{abs.count} provides the absolute count of measurements in
each bin.}
    \item{\code{annotate}}{If \code{TRUE} then the percentage calm and mean values are
printed in each panel together with a description of the statistic below
the plot. If \code{FALSE} then only the statistic will be printed.}
    \item{\code{border}}{Border colour for shaded areas. Default is no border.}
    \item{\code{type}}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}
    \item{\code{cols}}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}
    \item{\code{angle.scale}}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}
    \item{\code{offset}}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}
    \item{\code{strip.position}}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}
    \item{\code{auto.text}}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}
  }}
}
\value{
an \link[=openair-package]{openair} object. Summarised proportions can be
extracted directly using the \verb{$data} operator, e.g.
\code{object$data} for \code{output <- windRose(mydata)}. This returns a
data frame with three set columns: \code{cond}, conditioning based on
\code{type}; \code{wd}, the wind direction; and \code{calm}, the
\code{statistic} for the proportion of data unattributed to any specific
wind direction because it was collected under calm conditions; and then
several (one for each range binned for the plot) columns giving proportions
of measurements associated with each \code{ws} or \code{pollutant} range
plotted as a discrete panel.
}
\description{
The traditional wind rose plot that plots wind speed and wind direction by
different intervals. The pollution rose applies the same plot structure but
substitutes other measurements, most commonly a pollutant time series, for
wind speed.
}
\details{
\code{\link[=pollutionRose]{pollutionRose()}} is a \code{\link[=windRose]{windRose()}} wrapper which brings \code{pollutant}
forward in the argument list, and attempts to sensibly rescale break points
based on the \code{pollutant} data range by by-passing \code{ws.int}.

By default, \code{\link[=pollutionRose]{pollutionRose()}} will plot a pollution rose of \code{nox} using
"wedge" style segments and placing the scale key to the right of the plot.

It is possible to compare two wind speed-direction data sets using
\code{\link[=pollutionRose]{pollutionRose()}}. There are many reasons for doing so e.g. to see how one
site compares with another or for meteorological model evaluation. In this
case, \code{ws} and \code{wd} are considered to the the reference data sets
with which a second set of wind speed and wind directions are to be compared
(\code{ws2} and \code{wd2}). The first set of values is subtracted from the
second and the differences compared. If for example, \code{wd2} was biased
positive compared with \code{wd} then \code{pollutionRose} will show the bias
in polar coordinates. In its default use, wind direction bias is colour-coded
to show negative bias in one colour and positive bias in another.
}
\examples{
# pollutionRose of nox
pollutionRose(mydata, pollutant = "nox")

# source apportionment plot - contribution to mean
\dontrun{
pollutionRose(mydata, pollutant = "pm10", type = "year", statistic = "prop.mean")

# example of comparing 2 met sites
# first we will make some new ws/wd data with a postive bias
mydata$ws2 <- mydata$ws + 2 * rnorm(nrow(mydata)) + 1
mydata$wd2 <- mydata$wd + 30 * rnorm(nrow(mydata)) + 30

# need to correct negative wd
id <- which(mydata$wd2 < 0)
mydata$wd2[id] <- mydata$wd2[id] + 360

# results show postive bias in wd and ws
pollutionRose(mydata, ws = "ws", wd = "wd", ws2 = "ws2", wd2 = "wd2")

## add some wd bias to some nighttime hours
id <- which(as.numeric(format(mydata$date, "\%H")) \%in\% c(23, 1, 2, 3, 4, 5))
mydata$wd2[id] <- mydata$wd[id] + 30 * rnorm(length(id)) + 120
id <- which(mydata$wd2 < 0)
mydata$wd2[id] <- mydata$wd2[id] + 360

pollutionRose(
  mydata,
  ws = "ws",
  wd = "wd",
  ws2 = "ws2",
  wd2 = "wd2",
  breaks = c(-11, -2, -1, -0.5, 0.5, 1, 2, 11),
  cols = c("dodgerblue4", "white", "firebrick"),
  type = "daylight"
)
}
}
\seealso{
Other polar directional analysis functions: 
\code{\link{percentileRose}()},
\code{\link{polarAnnulus}()},
\code{\link{polarCluster}()},
\code{\link{polarDiff}()},
\code{\link{polarFreq}()},
\code{\link{polarPlot}()},
\code{\link{windRose}()}
}
\concept{polar directional analysis functions}
