% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refOpts.R
\name{refOpts}
\alias{refOpts}
\title{Define \code{ref.x} or \code{ref.y} options for \code{openair} plots}
\usage{
refOpts(
  intercept,
  alpha = 1,
  colour = "black",
  linetype = 1,
  linewidth = 1,
  color = NULL
)
}
\arguments{
\item{intercept}{The axis intercept(s) for the reference lines. Should be
numeric, dates, or date-times depending on the axis types. If another data
type is provided, it will attempt to be coerced to the correct type using
\code{\link[=as.numeric]{as.numeric()}}, \code{\link[lubridate:as_date]{lubridate::as_date()}} or \code{\link[lubridate:as_datetime]{lubridate::as_datetime()}},
respectively.}

\item{alpha}{Numeric value between 0 and 1 specifying the transparency of the
lines. Default is 1 (fully opaque).}

\item{colour, color}{Colour of the lines. Default is \code{"black"}. \code{colour} and
\code{color} are interchangeable, but \code{colour} is used preferentially if both
are given.}

\item{linetype}{Line type. Can be an integer (e.g., 1 for solid, 2 for
dashed) or a string (e.g., "solid", "dashed"). Default is 1 (solid).}

\item{linewidth}{Numeric value specifying the width of the lines. Default is
1.}
}
\value{
A list of options that can be passed to the \code{ref.x} or \code{ref.y}
arguments of functions like \code{\link[=timePlot]{timePlot()}}.
}
\description{
This function provides a convenient way to set default options for \code{ref.x} or
\code{ref.y} layers in \code{openair} plots, which show some form of horizontal or
vertical reference line. \code{intercept} can be a vector of any length; all other
arguments will be recycled to be equal to that length.
}
\examples{
# `ref.y` can just be a value to plot
timePlot(mydata, avg.time = "month", ref.y = 250, ref.x = "2002/01/01")

# use the `refOpts()` function to customise reference lines
timePlot(
  mydata,
  avg.time = "month",
  ref.y = refOpts(
    c(250, 300),
    alpha = c(0.5, 1),
    colour = c("grey50", "blue"),
    linetype = c(2, 1),
    linewidth = c(1, 2)
  )
)
}
