% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rolling_gaussian_cpp}
\alias{rolling_gaussian_cpp}
\title{Rolling Gaussian Smoother}
\usage{
rolling_gaussian_cpp(x, sigma, threshold = 0.5)
}
\arguments{
\item{x}{Numeric vector to smooth}

\item{sigma}{Standard deviation of the Gaussian kernel}

\item{threshold}{Min fraction of valid data in window (0-1)}
}
\description{
Rolling Gaussian Smoother
}
\keyword{internal}
