% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-docs.R
\name{shared_openair_params}
\alias{docs-shared-internal}
\title{Shared openair parameters}
\arguments{
\item{ws}{The name of the column in \code{mydata} representing the wind speed.
Defaults to \code{"ws"}.}

\item{wd}{The name of the column in \code{mydata} representing the decimal wind
direction, 0 to 360 where 0/360 are North and 180 is South. Defaults to
\code{"wd"}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{date.breaks}{Number of major x-axis intervals to use. The function will
try and choose a sensible number of dates/times as well as formatting the
date/time appropriately to the range being considered. The user can
override this behaviour by adjusting the value of \code{date.breaks} up or down.}

\item{date.format}{This option controls the date format on the x-axis. A
sensible format is chosen by default, but the user can set \code{date.format} to
override this. For format types see \code{\link[=strptime]{strptime()}}. For example, to format
the date like "Jan-2012" set \code{date.format = "\\\%b-\\\%Y"}.}

\item{breaks, labels}{If a categorical colour scale is required, \code{breaks}
should be specified. This can be either of:
\itemize{
\item A single value, which will divide the scale into \code{breaks} levels using the
same logic as \code{\link[=cutData]{cutData()}}. For example, \code{breaks = 5} will split the scale
into five quantiles.
\item A numeric vector, which will define the specific breakpoints. For
example, \code{c(0, 50, 100)} will bin the data into \verb{0 to 50}, \verb{50 to 100}, and
so on. If \code{breaks} does not cover the full range of the data, the outer
limits will be extended so that the full colour scale is covered while
retaining the desired number of breaks.
}

By default, \code{breaks} will generate nicely formatted labels for each
category. The \code{labels} argument overrides this - for example, a user could
define \verb{breaks = 3, labels = c("low", "medium", "high")}. Care should be
taken to provide the appropriate number of \code{labels} - it should be equal to
\code{breaks} if a single value is given, or equal to \code{length(breaks)-1} if
\code{breaks} is a vector.}

\item{x.relation, y.relation}{This determines how the x- and y-axis scales are
plotted. \code{"same"} ensures all panels use the same scale and \code{"free"} will
use panel-specific scales. The latter is a useful setting when plotting
data with very different values.}

\item{angle.scale}{In radial plots (e.g., \code{\link[=polarPlot]{polarPlot()}}), the radial scale is
drawn directly on the plot itself. While suitable defaults have been
chosen, sometimes the placement of the scale may interfere with an
interesting feature. \code{angle.scale} can take any value between \code{0} and \code{360}
to place the scale at a different angle, or \code{FALSE} to move it to the side
of the plots.}

\item{windflow}{If \code{TRUE}, the vector-averaged wind speed and direction will
be plotted using arrows. Alternatively, can be a list of arguments to
control the appearance of the arrows (colour, linewidth, alpha value,
etc.). See \code{\link[=windflowOpts]{windflowOpts()}} for details.}

\item{offset}{\code{offset} controls the size of the 'hole' in the middle and is
expressed on a scale of \code{0} to \code{100}, where \code{0} is no hole and \code{100} is a
hole that takes up the entire plotting area.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label. All of these
are passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\description{
This is a central place for describing common shared parameters.
This ensures consistency across \code{openair}
}
\keyword{internal}
