% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothTrend.R
\name{smoothTrend}
\alias{smoothTrend}
\title{Calculate nonparametric smooth trends}
\usage{
smoothTrend(
  mydata,
  pollutant = "nox",
  avg.time = "month",
  data.thresh = 0,
  statistic = "mean",
  percentile = NA,
  k = NULL,
  deseason = FALSE,
  simulate = FALSE,
  n = 200,
  autocor = FALSE,
  type = "default",
  cols = "brewer1",
  x.relation = "same",
  y.relation = "same",
  ref.x = NULL,
  ref.y = NULL,
  key.columns = 1,
  key.position = "bottom",
  strip.position = "top",
  name.pol = NULL,
  date.breaks = 7,
  date.format = NULL,
  auto.text = TRUE,
  ci = TRUE,
  alpha = 0.2,
  progress = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame of time series. Must include a \code{date} field and at
least one variable to plot.}

\item{pollutant}{Name of variable to plot. Two or more pollutants can be
plotted, in which case a form like \code{pollutant = c("nox", "co")} should be
used.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

\strong{Period boundary behaviour:} how bin boundaries are determined depends on
the type of period:
\itemize{
\item \emph{Single-unit periods} (\code{"hour"}, \code{"day"}, \code{"week"}, etc.) are floored to
the start of the enclosing unit in the data's timezone (e.g. \code{"day"}
floors to midnight).
\item \emph{Multi-unit fixed-length periods} (\code{"3 day"}, \code{"6 hour"}, \code{"2 week"},
etc.) use epoch-aligned arithmetic: bin boundaries are fixed multiples of
the period length counted from 1970-01-01, so the same calendar dates
always fall in the same bin regardless of where the data starts, and bins
run continuously across month boundaries without resetting at the start of
each month. For example, with \code{avg.time = "3 day"} a bin that begins on
29 January will end on 31 January, and the next bin begins on 1 February
— the month boundary does not start a new bin.
\item \emph{Calendar periods} (\code{"month"}, \code{"quarter"}, \code{"year"}) are floored to the
start of the enclosing calendar unit, so they correctly respect variable
month and year lengths.
}

Note that \code{avg.time} can be \emph{less} than the time interval of the original
series, in which case the series is expanded to the new time interval. This
is useful, for example, for calculating a 15-minute time series from an
hourly one where an hourly value is repeated for each new 15-minute period.
Note that when expanding data in this way it is necessary to ensure that
the time interval of the original series is an exact multiple of \code{avg.time}
e.g. hour to 10 minutes, day to hour. Also, the input time series must have
consistent time gaps between successive intervals so that \code{\link[=timeAverage]{timeAverage()}}
can work out how much 'padding' to apply. To pad-out data in this way
choose \code{fill = TRUE}.}

\item{data.thresh}{The data capture threshold to use (\%). A value of zero
means that all available data will be used in a particular period
regardless if of the number of values available. Conversely, a value of 100
will mean that all data will need to be present for the average to be
calculated, else it is recorded as \code{NA}. See also \code{interval}, \code{start.date}
and \code{end.date} to see whether it is advisable to set these other options.}

\item{statistic}{Statistic used for calculating monthly values. Default is
\code{"mean"}, but can also be \code{"percentile"}. See \code{\link[=timeAverage]{timeAverage()}} for more
details.}

\item{percentile}{Percentile value(s) to use if \code{statistic = "percentile"} is
chosen. Can be a vector of numbers e.g. \code{percentile = c(5, 50, 95)} will
plot the 5th, 50th and 95th percentile values together on the same plot.}

\item{k}{This is the smoothing parameter used by the \code{\link[mgcv:gam]{mgcv::gam()}} function
in package \code{mgcv}. By default it is not used and the amount of smoothing is
optimised automatically. However, sometimes it is useful to set the
smoothing amount manually using \code{k}.}

\item{deseason}{Should the data be de-deasonalized first? If \code{TRUE} the
function \code{stl} is used (seasonal trend decomposition using loess). Note
that if \code{TRUE} missing data are first imputed using a Kalman filter and
Kalman smooth.}

\item{simulate}{Should simulations be carried out to determine the
Mann-Kendall tau and p-value. The default is \code{FALSE}. If \code{TRUE}, bootstrap
simulations are undertaken, which also account for autocorrelation.}

\item{n}{Number of bootstrap simulations if \code{simulate = TRUE}.}

\item{autocor}{Should autocorrelation be considered in the trend uncertainty
estimates? The default is \code{FALSE}. Generally, accounting for
autocorrelation increases the uncertainty of the trend estimate sometimes
by a large amount.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{x.relation, y.relation}{This determines how the x- and y-axis scales are
plotted. \code{"same"} ensures all panels use the same scale and \code{"free"} will
use panel-specific scales. The latter is a useful setting when plotting
data with very different values.}

\item{ref.x}{See \code{ref.y} for details. In this case the correct date format
should be used for a vertical line e.g. \code{ref.x = list(v = as.POSIXct("2000-06-15"), lty = 5)}.}

\item{ref.y}{A list with details of the horizontal lines to be added
representing reference line(s). For example, \code{ref.y = list(h = 50, lty = 5)} will add a dashed horizontal line at 50. Several lines can be plotted
e.g. \code{ref.y = list(h = c(50, 100), lty = c(1, 5), col = c("green", "blue"))}.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{name.pol}{This option can be used to give alternative names for the
variables plotted. Instead of taking the column headings as names, the user
can supply replacements. For example, if a column had the name "nox" and
the user wanted a different description, then setting \code{name.pol = "nox before change"} can be used. If more than one pollutant is plotted then use
\code{c} e.g. \code{name.pol = c("nox here", "o3 there")}.}

\item{date.breaks}{Number of major x-axis intervals to use. The function will
try and choose a sensible number of dates/times as well as formatting the
date/time appropriately to the range being considered. The user can
override this behaviour by adjusting the value of \code{date.breaks} up or down.}

\item{date.format}{This option controls the date format on the x-axis. A
sensible format is chosen by default, but the user can set \code{date.format} to
override this. For format types see \code{\link[=strptime]{strptime()}}. For example, to format
the date like "Jan-2012" set \code{date.format = "\\\%b-\\\%Y"}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{ci}{Should confidence intervals be plotted? The default is \code{TRUE}.}

\item{alpha}{The alpha transparency of shaded confidence intervals - if
plotted. A value of 0 is fully transparent and 1 is fully opaque.}

\item{progress}{Show a progress bar when many groups make up \code{type}? Defaults
to \code{TRUE}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{title} override the x-axis label, y-axis label, and plot title.
\item \code{ylim} and \code{xlim} control axis limits.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{size}, \code{linewidth}, \code{linetype}, \code{alpha}, and \code{shape} control various graphical parameters.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
Use non-parametric methods to calculate time series trends
}
\details{
The \code{\link[=smoothTrend]{smoothTrend()}} function provides a flexible way of estimating the trend
in the concentration of a pollutant or other variable. Monthly mean values
are calculated from an hourly (or higher resolution) or daily time series.
There is the option to deseasonalise the data if there is evidence of a
seasonal cycle.

\code{\link[=smoothTrend]{smoothTrend()}} uses a Generalized Additive Model (GAM) from the
\code{\link[mgcv:gam]{mgcv::gam()}} package to find the most appropriate level of smoothing. The
function is particularly suited to situations where trends are not monotonic
(see discussion with \code{\link[=TheilSen]{TheilSen()}} for more details on this). The
\code{\link[=smoothTrend]{smoothTrend()}} function is particularly useful as an exploratory technique
e.g. to check how linear or non-linear trends are.

95\% confidence intervals are shown by shading. Bootstrap estimates of the
confidence intervals are also available through the \code{simulate} option.
Residual resampling is used.

Trends can be considered in a very wide range of ways, controlled by setting
\code{type} - see examples below.
}
\examples{
# trend plot for nox
smoothTrend(mydata, pollutant = "nox")

# trend plot by each of 8 wind sectors
\dontrun{
smoothTrend(mydata, pollutant = "o3", type = "wd", ylab = "o3 (ppb)")

# several pollutants, no plotting symbol
smoothTrend(mydata, pollutant = c("no2", "o3", "pm10", "pm25"), shape = NA)

# percentiles
smoothTrend(mydata,
  pollutant = "o3", statistic = "percentile",
  percentile = 95
)

# several percentiles with control over lines used
smoothTrend(mydata,
  pollutant = "o3", statistic = "percentile",
  percentile = c(5, 50, 95), linewidth = c(1, 2, 1), linetype = c(5, 1, 5)
)
}
}
\seealso{
Other time series and trend functions: 
\code{\link{TheilSen}()},
\code{\link{calendarPlot}()},
\code{\link{timePlot}()},
\code{\link{timeProp}()},
\code{\link{timeVariation}()}
}
\author{
David Carslaw
}
\concept{time series and trend functions}
