% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheilSen.R
\name{TheilSen}
\alias{TheilSen}
\title{Tests for trends using Theil-Sen estimates}
\usage{
TheilSen(
  mydata,
  pollutant = "nox",
  deseason = FALSE,
  type = "default",
  avg.time = "month",
  statistic = "mean",
  percentile = NA,
  data.thresh = 0,
  alpha = 0.05,
  dec.place = 2,
  lab.frac = 0.99,
  lab.cex = 0.8,
  x.relation = "same",
  y.relation = "same",
  data.col = "cornflowerblue",
  trend = list(lty = c(1, 5), lwd = c(2, 1), col = c("red", "red")),
  text.col = "darkgreen",
  slope.text = NULL,
  cols = NULL,
  auto.text = TRUE,
  autocor = FALSE,
  slope.percent = FALSE,
  date.breaks = 7,
  date.format = NULL,
  strip.position = "top",
  plot = TRUE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the field \code{date} and at least one other
parameter for which a trend test is required; typically (but not
necessarily) a pollutant.}

\item{pollutant}{The parameter for which a trend test is required. Mandatory.}

\item{deseason}{Should the data be de-deasonalized first? If \code{TRUE} the
function \code{stl} is used (seasonal trend decomposition using loess). Note
that if \code{TRUE} missing data are first imputed using a linear regression by month because \code{stl} cannot handle missing data. In this case the plot shows where the missing data have been imputed as a grey filled circle.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{avg.time}{Can be \dQuote{month} (the default), \dQuote{season} or
\dQuote{year}. Determines the time over which data should be averaged. Note
that for \dQuote{year}, six or more years are required. For \dQuote{season}
the data are split up into spring: March, April, May etc. Note that
December is considered as belonging to winter of the following year.}

\item{statistic}{Statistic used for calculating monthly values. Default is
\dQuote{mean}, but can also be \dQuote{percentile}. See \code{\link[=timeAverage]{timeAverage()}} for
more details.}

\item{percentile}{Single percentile value to use if \code{statistic = "percentile"} is chosen.}

\item{data.thresh}{The data capture threshold to use (\%) when aggregating the
data using \code{avg.time}. A value of zero means that all available data will
be used in a particular period regardless if of the number of values
available. Conversely, a value of 100 will mean that all data will need to
be present for the average to be calculated, else it is recorded as \code{NA}.}

\item{alpha}{For the confidence interval calculations of the slope. The
default is 0.05. To show 99\\% confidence intervals for the value of the
trend, choose alpha = 0.01 etc.}

\item{dec.place}{The number of decimal places to display the trend estimate
at. The default is 2.}

\item{lab.frac}{Fraction along the y-axis that the trend information should
be printed at, default 0.99.}

\item{lab.cex}{Size of text for trend information.}

\item{x.relation, y.relation}{This determines how the x- and y-axis scales are
plotted. \code{"same"} ensures all panels use the same scale and \code{"free"} will
use panel-specific scales. The latter is a useful setting when plotting
data with very different values.}

\item{data.col}{Colour name for the data}

\item{trend}{list containing information on the line width, line type and
line colour for the main trend line and confidence intervals respectively.}

\item{text.col}{Colour name for the slope/uncertainty numeric estimates}

\item{slope.text}{The text shown for the slope (default is
\sQuote{units/year}).}

\item{cols}{Predefined colour scheme, currently only enabled for
\code{"greyscale"}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{autocor}{Should autocorrelation be considered in the trend uncertainty
estimates? The default is \code{FALSE}. Generally, accounting for
autocorrelation increases the uncertainty of the trend estimate ---
sometimes by a large amount.}

\item{slope.percent}{Should the slope and the slope uncertainties be
expressed as a percentage change per year? The default is \code{FALSE} and the
slope is expressed as an average units/year change e.g. ppb. Percentage
changes can often be confusing and should be clearly defined. Here the
percentage change is expressed as 100 * (C.end/C.start - 1) / (end.year -
start.year). Where C.start is the concentration at the start date and C.end
is the concentration at the end date.

For \code{avg.time = "year"} (end.year - start.year) will be the total number of
years - 1. For example, given a concentration in year 1 of 100 units and a
percentage reduction of 5\%/yr, after 5 years there will be 75 units but the
actual time span will be 6 years i.e. year 1 is used as a reference year.
Things are slightly different for monthly values e.g. \code{avg.time = "month"},
which will use the total number of months as a basis of the time span and
is therefore able to deal with partial years. There can be slight
differences in the \%/yr trend estimate therefore, depending on whether
monthly or annual values are considered.}

\item{date.breaks}{Number of major x-axis intervals to use. The function will
try and choose a sensible number of dates/times as well as formatting the
date/time appropriately to the range being considered. The user can
override this behaviour by adjusting the value of \code{date.breaks} up or down.}

\item{date.format}{This option controls the date format on the x-axis. A
sensible format is chosen by default, but the user can set \code{date.format} to
override this. For format types see \code{\link[=strptime]{strptime()}}. For example, to format
the date like "Jan-2012" set \code{date.format = "\\\%b-\\\%Y"}.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{silent}{When \code{FALSE} the function will give updates on trend-fitting
progress.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label. All of these
are passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object. The \code{data} component of the
\code{TheilSen} output includes two subsets: \code{main.data}, the monthly data
\code{res2} the trend statistics. For \code{output <- TheilSen(mydata, "nox")}, these
can be extracted as \code{object$data$main.data} and \code{object$data$res2},
respectively. Note: In the case of the intercept, it is assumed the y-axis
crosses the x-axis on 1/1/1970.
}
\description{
Theil-Sen slope estimates and tests for trend. The \code{TheilSen} function is
flexible in the sense that it can be applied to data in many ways e.g. by day
of the week, hour of day and wind direction. This flexibility makes it much
easier to draw inferences from data e.g. why is there a strong downward trend
in concentration from one wind sector and not another, or why trends on one
day of the week or a certain time of day are unexpected.
}
\details{
For data that are strongly seasonal, perhaps from a background site, or a
pollutant such as ozone, it will be important to deseasonalise the data
(using the option \code{deseason = TRUE}.Similarly, for data that increase, then
decrease, or show sharp changes it may be better to use \code{\link[=smoothTrend]{smoothTrend()}}.

A minimum of 6 points are required for trend estimates to be made.

Note! that since version 0.5-11 openair uses Theil-Sen to derive the p values
also for the slope. This is to ensure there is consistency between the
calculated p value and other trend parameters i.e. slope estimates and
uncertainties. The p value and all uncertainties are calculated through
bootstrap simulations.

Note that the symbols shown next to each trend estimate relate to how
statistically significant the trend estimate is: p $<$ 0.001 =
***, p $<$ 0.01 = **, p $<$ 0.05 = * and p $<$ 0.1 = $+$.

Some of the code used in \code{TheilSen} is based on that from Rand Wilcox. This
mostly relates to the Theil-Sen slope estimates and uncertainties. Further
modifications have been made to take account of correlated data based on
Kunsch (1989). The basic function has been adapted to take account of
auto-correlated data using block bootstrap simulations if \code{autocor = TRUE}
(Kunsch, 1989). We follow the suggestion of Kunsch (1989) of setting the
block length to n(1/3) where n is the length of the time series.

The slope estimate and confidence intervals in the slope are plotted and
numerical information presented.
}
\examples{
# trend plot for nox
TheilSen(mydata, pollutant = "nox")

# trend plot for ozone with p=0.01 i.e. uncertainty in slope shown at
# 99 \% confidence interval

\dontrun{
TheilSen(mydata, pollutant = "o3", ylab = "o3 (ppb)", alpha = 0.01)
}

# trend plot by each of 8 wind sectors
\dontrun{
TheilSen(mydata, pollutant = "o3", type = "wd", ylab = "o3 (ppb)")
}

# and for a subset of data (from year 2000 onwards)
\dontrun{
TheilSen(selectByDate(mydata, year = 2000:2005), pollutant = "o3", ylab = "o3 (ppb)")
}
}
\references{
Helsel, D., Hirsch, R., 2002. Statistical methods in water resources. US
Geological Survey. Note that this is a very good resource for statistics as
applied to environmental data.

Hirsch, R. M., Slack, J. R., Smith, R. A., 1982. Techniques of trend analysis
for monthly water-quality data. Water Resources Research 18 (1), 107-121.

Kunsch, H. R., 1989. The jackknife and the bootstrap for general stationary
observations. Annals of Statistics 17 (3), 1217-1241.

Sen, P. K., 1968. Estimates of regression coefficient based on Kendall's tau.
Journal of the American Statistical Association 63(324).

Theil, H., 1950. A rank invariant method of linear and polynomial regression
analysis, i, ii, iii. Proceedings of the Koninklijke Nederlandse Akademie
Wetenschappen, Series A - Mathematical Sciences 53, 386-392, 521-525,
1397-1412.

\dots{} see also several of the Air Quality Expert Group (AQEG) reports for
the use of similar tests applied to UK/European air quality data.
}
\seealso{
Other time series and trend functions: 
\code{\link{calendarPlot}()},
\code{\link{smoothTrend}()},
\code{\link{timePlot}()},
\code{\link{timeProp}()},
\code{\link{timeVariation}()}
}
\author{
David Carslaw with some trend code from Rand Wilcox
}
\concept{time series and trend functions}
