% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAIR-package.R
\docType{package}
\name{TheOpenAIR-package}
\alias{TheOpenAIR}
\alias{TheOpenAIR-package}
\title{TheOpenAIR: Integrate OpenAI's Large Language Models into Your R Workflows}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utilizing OpenAI's API as the backend, TheOpenAIR offers R wrapper functions for the ChatGPT endpoint and several high-level functions that enable the integration of ChatGPT's capabilities in diverse data-related tasks, such as data cleansing and automated analytics script generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://openair-lib.org/}
}

}
\author{
\strong{Maintainer}: Ulrich Matter \email{umatter@protonmail.com} [copyright holder]

Authors:
\itemize{
  \item Jonathan Chassot \email{jonathan.chassot@unisg.ch}
}

}
\keyword{internal}
