% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeProp.R
\name{timeProp}
\alias{timeProp}
\title{Time series plot with categories shown as a stacked bar chart}
\usage{
timeProp(
  mydata,
  pollutant = "nox",
  proportion = "wd",
  avg.time = "day",
  type = "default",
  cols = "Set1",
  normalise = FALSE,
  x.relation = "same",
  y.relation = "same",
  key.columns = 1,
  key.position = "right",
  key.title = proportion,
  strip.position = "top",
  date.breaks = 7,
  date.format = NULL,
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the fields \code{date}, \code{pollutant} and a
splitting variable \code{proportion}}

\item{pollutant}{Name of the pollutant to plot contained in \code{mydata}.}

\item{proportion}{The splitting variable that makes up the bars in the bar
chart, defaulting to \code{"wd"}. Could be \code{"cluster"} if the output from
\code{\link[=polarCluster]{polarCluster()}} or \code{\link[=trajCluster]{trajCluster()}} is being analysed. If \code{proportion} is a
numeric variable it is split into 4 quantiles (by default) by \code{\link[=cutData]{cutData()}}.
If \code{proportion} is a factor or character variable then the categories are
used directly.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be \code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in which case
3-month seasonal values are calculated with spring defined as March, April,
May and so on.

Note that \code{avg.time} when used in \code{timeProp} should be greater than the
time gap in the original data. For example, \code{avg.time = "day"} for hourly
data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{x.relation, y.relation}{This determines how the x- and y-axis scales are
plotted. \code{"same"} ensures all panels use the same scale and \code{"free"} will
use panel-specific scales. The latter is a useful setting when plotting
data with very different values.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{date.breaks}{Number of major x-axis intervals to use. The function will
try and choose a sensible number of dates/times as well as formatting the
date/time appropriately to the range being considered. The user can
override this behaviour by adjusting the value of \code{date.breaks} up or down.}

\item{date.format}{This option controls the date format on the x-axis. A
sensible format is chosen by default, but the user can set \code{date.format} to
override this. For format types see \code{\link[=strptime]{strptime()}}. For example, to format
the date like "Jan-2012" set \code{date.format = "\\\%b-\\\%Y"}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, x-axis label and y-axis label. All of these are
passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2
columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the plotting
of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and \code{\link[=trajCluster]{trajCluster()}} that
consider local and regional (back trajectory) cluster analysis respectively.
However, the function has more general use for understanding time series
data.
}
\details{
In order to plot time series in this way, some sort of time aggregation is
needed, which is controlled by the option \code{avg.time}.

The plot shows the value of \code{pollutant} on the y-axis (averaged according to
\code{avg.time}). The time intervals are made up of bars split according to
\code{proportion}. The bars therefore show how the total value of \code{pollutant} is
made up for any time interval.
}
\examples{
# monthly plot of SO2 showing the contribution by wind sector
timeProp(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
Other time series and trend functions: 
\code{\link{TheilSen}()},
\code{\link{calendarPlot}()},
\code{\link{smoothTrend}()},
\code{\link{timePlot}()},
\code{\link{timeVariation}()}

Other cluster analysis functions: 
\code{\link{polarCluster}()},
\code{\link{trajCluster}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{cluster analysis functions}
\concept{time series and trend functions}
