% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeProp.R
\name{timeProp}
\alias{timeProp}
\title{Time series plot with categories shown as a stacked bar chart}
\usage{
timeProp(
  mydata,
  pollutant = "nox",
  proportion = "cluster",
  avg.time = "day",
  type = "default",
  cols = "Set1",
  normalise = FALSE,
  key = TRUE,
  key.columns = 1,
  key.position = "right",
  key.title = proportion,
  date.breaks = 7,
  date.format = NULL,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame containing the fields \code{date}, \code{pollutant} and a
splitting variable \code{proportion}}

\item{pollutant}{Name of the pollutant to plot contained in \code{mydata}.}

\item{proportion}{The splitting variable that makes up the bars in the bar
chart e.g. \code{proportion = "cluster"} if the output from \code{polarCluster} is
being analysed. If \code{proportion} is a numeric variable it is split into 4
quantiles (by default) by \code{cutData}. If \code{proportion} is a factor or
character variable then the categories are used directly.}

\item{avg.time}{This defines the time period to average to. Can be \code{"sec"},
\code{"min"}, \code{"hour"}, \code{"day"}, \code{"DSTday"}, \code{"week"}, \code{"month"}, \code{"quarter"} or
\code{"year"}. For much increased flexibility a number can precede these options
followed by a space. For example, an average of 2 months would be
\code{avg.time = "2 month"}. In addition, \code{avg.time} can equal \code{"season"}, in
which case 3-month seasonal values are calculated with spring defined as
March, April, May and so on.

Note that \code{avg.time} when used in \code{timeProp} should be greater than the
time gap in the original data. For example, \code{avg.time = "day"} for hourly
data is OK, but \code{avg.time = "hour"} for daily data is not.}

\item{type}{\code{type} determines how the data are split i.e. conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{\link[=cutData]{cutData()}},
e.g., \code{"season"}, \code{"year"}, \code{"weekday"} and so on. For example, \code{type = "season"} will produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

\code{type} must be of length one.}

\item{cols}{Colours to be used for plotting; see \code{\link[=openColours]{openColours()}} for details.}

\item{normalise}{If \code{normalise = TRUE} then each time interval is scaled to
100. This is helpful to show the relative (percentage) contribution of the
proportions.}

\item{key}{Should a key be drawn? The default is \code{TRUE}.}

\item{key.columns}{Number of columns to be used in the key. With many
pollutants a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{columns} to be less than the
number of pollutants.}

\item{key.position}{Location where the scale key is to plotted. Can include
\code{"top"}, \code{"bottom"}, \code{"right"} and \code{"left"}.}

\item{key.title}{The title of the key.}

\item{date.breaks}{Number of major x-axis intervals to use. The function will
try and choose a sensible number of dates/times as well as formatting the
date/time appropriately to the range being considered. This does not always
work as desired automatically. The user can therefore increase or decrease
the number of intervals by adjusting the value of \code{date.breaks} up or down.}

\item{date.format}{This option controls the date format on the x-axis. While
\code{\link[=timePlot]{timePlot()}} generally sets the date format sensibly there can be some
situations where the user wishes to have more control. For format types see
\code{\link[=strptime]{strptime()}}. For example, to format the date like "Jan-2012" set
\code{date.format = "\\\%b-\\\%Y"}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the '2' in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{border} sets the border colour of each bar.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function shows time series plots as stacked bar charts. The different
categories in the bar chart are made up from a character or factor variable
in a data frame. The function is primarily developed to support the plotting
of cluster analysis output from \code{\link[=polarCluster]{polarCluster()}} and \code{\link[=trajCluster]{trajCluster()}} that
consider local and regional (back trajectory) cluster analysis respectively.
However, the function has more general use for understanding time series
data.
}
\details{
In order to plot time series in this way, some sort of time aggregation is
needed, which is controlled by the option \code{avg.time}.

The plot shows the value of \code{pollutant} on the y-axis (averaged according to
\code{avg.time}). The time intervals are made up of bars split according to
\code{proportion}. The bars therefore show how the total value of \code{pollutant} is
made up for any time interval.
}
\examples{
# monthly plot of SO2 showing the contribution by wind sector
timeProp(mydata, pollutant = "so2", avg.time = "month", proportion = "wd")
}
\seealso{
Other time series and trend functions: 
\code{\link{TheilSen}()},
\code{\link{calendarPlot}()},
\code{\link{runRegression}()},
\code{\link{smoothTrend}()},
\code{\link{timePlot}()},
\code{\link{timeVariation}()},
\code{\link{trendLevel}()}

Other cluster analysis functions: 
\code{\link{polarCluster}()},
\code{\link{trajCluster}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{cluster analysis functions}
\concept{time series and trend functions}
