% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajLevel.R
\name{trajLevel}
\alias{trajLevel}
\title{Trajectory level plots with conditioning}
\usage{
trajLevel(
  mydata,
  lon = "lon",
  lat = "lat",
  pollutant = "height",
  type = "default",
  smooth = FALSE,
  statistic = "frequency",
  percentile = 90,
  lon.inc = 1,
  lat.inc = lon.inc,
  min.bin = 1,
  .combine = NULL,
  sigma = 1.5,
  cols = "default",
  crs = 4326,
  map = TRUE,
  map.fill = TRUE,
  map.cols = "grey40",
  map.border = "black",
  map.alpha = 0.3,
  map.lwd = 1,
  map.lty = 1,
  grid.col = "deepskyblue",
  grid.nx = 9,
  grid.ny = grid.nx,
  origin = TRUE,
  key.title = NULL,
  key.position = "right",
  key.columns = NULL,
  strip.position = "top",
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{Data frame, the result of importing a trajectory file using
\code{\link[=importTraj]{importTraj()}}.}

\item{lon, lat}{Columns containing the decimal longitude and latitude.}

\item{pollutant}{Pollutant (or any numeric column) to be plotted, if any.
Alternatively, use \code{group}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{smooth}{Should the trajectory surface be smoothed?}

\item{statistic}{One of:
\itemize{
\item \code{"frequency"} (the default) shows trajectory frequencies.
\item \code{"hexbin"}, which is similar to \code{"frequency"} but shows a hexagonal
grid of counts.
\item \code{"difference"} - in this case trajectories where the associated
concentration is greater than \code{percentile} are compared with the the full
set of trajectories to understand the differences in frequencies of the
origin of air masses. The comparison is made by comparing the percentage
change in gridded frequencies. For example, such a plot could show that the
top 10\\% of concentrations of PM10 tend to originate from air-mass origins
to the east.
\item \code{"pscf"} for a Potential Source Contribution Function map. This statistic
method interacts with \code{percentile}.
\item \code{"cwt"} for concentration weighted trajectories.
\item \code{"sqtba"} to undertake Simplified Quantitative Transport Bias
Analysis. This statistic method interacts with \code{.combine} and \code{sigma}.
}}

\item{percentile}{The percentile concentration of \code{pollutant} against which
the all trajectories are compared.}

\item{lon.inc, lat.inc}{The longitude and latitude intervals to be used for
binning data. If \code{statistic = "hexbin"}, the minimum value out of of
\code{lon.inc} and \code{lat.inc} is passed to the \code{binwidth} argument of
\code{\link[ggplot2:geom_hex]{ggplot2::geom_hex()}}.}

\item{min.bin}{The minimum number of unique points in a grid cell. Counts
below \code{min.bin} are set as missing.}

\item{.combine}{When statistic is "SQTBA" it is possible to combine lots of
receptor locations to derive a single map. \code{.combine} identifies the column
that differentiates different sites (commonly a column named \code{"site"}).
Note that individual site maps are normalised first by dividing by their
mean value.}

\item{sigma}{For the SQTBA approach \code{sigma} determines the amount of back
trajectory spread based on the Gaussian plume equation. Values in the
literature suggest 5.4 km after one hour. However, testing suggests lower
values reveal source regions more effectively while not introducing too
much noise.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{crs}{The coordinate reference system to use for plotting. Defaults to
\code{4326}, which is the WGS84 geographic coordinate system, the standard,
unprojected latitude/longitude system used in GPS, Google Earth, and GIS
mapping. Other \code{crs} values are available - for example, \code{27700} will use
the the OSGB36/British National Grid.}

\item{map}{Should a base map be drawn? If \code{TRUE} the world base map provided
by \code{\link[ggplot2:map_data]{ggplot2::map_data()}} will be used.}

\item{map.fill}{Should the base map be a filled polygon? Default is to fill
countries.}

\item{map.cols}{If \code{map.fill = TRUE} \code{map.cols} controls the fill colour.
Examples include \code{map.fill = "grey40"} and \code{map.fill = openColours("default", 10)}. The latter colours the countries and can help
differentiate them.}

\item{map.border}{The colour to use for the map outlines/borders. Defaults to
\code{"black"}.}

\item{map.alpha}{The transparency level of the filled map which takes values
from 0 (full transparency) to 1 (full opacity). Setting it below 1 can help
view trajectories, trajectory surfaces etc. \emph{and} a filled base map.}

\item{map.lwd}{The map line width, a positive number, defaulting to \code{1}.}

\item{map.lty}{The map line type. Line types can either be specified as an
integer (\code{0} = blank, \code{1} = solid (default), \code{2} = dashed, \code{3} = dotted,
\code{4} = dotdash, \code{5} = longdash, \code{6} = twodash) or as one of the character
strings "blank", "solid", "dashed", "dotted", "dotdash", "longdash", or
"twodash", where "blank" uses 'invisible lines' (i.e., does not draw them).}

\item{grid.col}{The colour of the map grid to be used. To remove the grid set
\code{grid.col = "transparent"}.}

\item{grid.nx, grid.ny}{The approximate number of ticks to draw on the map
grid. \code{grid.nx} defaults to \code{9}, and \code{grid.ny} defaults to whatever value
is passed to \code{grid.nx}. Setting both values to \code{0} will remove the grid
entirely. The number of ticks is approximate as this value is passed to
\code{\link[scales:breaks_pretty]{scales::breaks_pretty()}} to determine nice-looking, round breakpoints.}

\item{origin}{If true a filled circle dot is shown to mark the receptor
point.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{border} sets the border colour of each tile.
}}
}
\value{
an \link[=openair-package]{openair} object
}
\description{
This function plots gridded back trajectories. This function requires that
data are imported using the \code{\link[=importTraj]{importTraj()}} function.
}
\details{
An alternative way of showing the trajectories compared with plotting
trajectory lines is to bin the points into latitude/longitude intervals. For
these purposes \code{\link[=trajLevel]{trajLevel()}} should be used. There are several trajectory
statistics that can be plotted as gridded surfaces. First, \code{statistic} can be
set to "frequency" to show the number of back trajectory points in a grid
square. Grid squares are by default at 1 degree intervals, controlled by
\code{lat.inc} and \code{lon.inc}. Such plots are useful for showing the frequency of
air mass locations. Note that it is also possible to set \code{statistic = "hexbin"} for plotting frequencies (not concentrations), which will produce a
plot by hexagonal binning.

If \code{statistic = "difference"} the trajectories associated with a
concentration greater than \code{percentile} are compared with the the full set of
trajectories to understand the differences in frequencies of the origin of
air masses of the highest concentration trajectories compared with the
trajectories on average. The comparison is made by comparing the percentage
change in gridded frequencies. For example, such a plot could show that the
top 10\\% of concentrations of PM10 tend to originate from air-mass origins to
the east.

If \code{statistic = "pscf"} then the Potential Source Contribution Function is
plotted. The PSCF calculates the probability that a source is located at
latitude \eqn{i} and longitude \eqn{j} (Pekney et al., 2006).The basis of
PSCF is that if a source is located at (i,j), an air parcel back trajectory
passing through that location indicates that material from the source can be
collected and transported along the trajectory to the receptor site. PSCF
solves \deqn{PSCF = m_{ij}/n_{ij}} where \eqn{n_{ij}} is the number of times
that the trajectories passed through the cell (i,j) and \eqn{m_{ij}} is the
number of times that a source concentration was high when the trajectories
passed through the cell (i,j). The criterion for determining \eqn{m_{ij}} is
controlled by \code{percentile}, which by default is 90. Note also that cells with
few data have a weighting factor applied to reduce their effect.

A limitation of the PSCF method is that grid cells can have the same PSCF
value when sample concentrations are either only slightly higher or much
higher than the criterion. As a result, it can be difficult to distinguish
moderate sources from strong ones. Seibert et al. (1994) computed
concentration fields to identify source areas of pollutants. The
Concentration Weighted Trajectory (CWT) approach considers the concentration
of a species together with its residence time in a grid cell. The CWT
approach has been shown to yield similar results to the PSCF approach. The
openair manual has more details and examples of these approaches.

A further useful refinement is to smooth the resulting surface, which is
possible by setting \code{smooth = TRUE}.
}
\examples{

# show a simple case with no pollutant i.e. just the trajectories
# let's check to see where the trajectories were coming from when
# Heathrow Airport was closed due to the Icelandic volcanic eruption
# 15--21 April 2010.
# import trajectories for London and plot
\dontrun{
lond <- importTraj("london", 2010)
}
# more examples to follow linking with concentration measurements...

# import some measurements from KC1 - London
\dontrun{
kc1 <- importAURN("kc1", year = 2010)
# now merge with trajectory data by 'date'
lond <- merge(lond, kc1, by = "date")

# trajectory plot, no smoothing - and limit lat/lon area of interest
# use PSCF
trajLevel(subset(lond, lat > 40 & lat < 70 & lon > -20 & lon < 20),
  pollutant = "pm10", statistic = "pscf"
)

# can smooth surface, suing CWT approach:
trajLevel(subset(lond, lat > 40 & lat < 70 & lon > -20 & lon < 20),
  pollutant = "pm2.5", statistic = "cwt", smooth = TRUE
)

# plot by season:
trajLevel(subset(lond, lat > 40 & lat < 70 & lon > -20 & lon < 20),
  pollutant = "pm2.5",
  statistic = "pscf", type = "season"
)
}
}
\references{
Pekney, N. J., Davidson, C. I., Zhou, L., & Hopke, P. K. (2006). Application
of PSCF and CPF to PMF-Modeled Sources of PM 2.5 in Pittsburgh. Aerosol
Science and Technology, 40(10), 952-961.

Seibert, P., Kromp-Kolb, H., Baltensperger, U., Jost, D., 1994. Trajectory
analysis of high-alpine air pollution data. NATO Challenges of Modern Society
18, 595-595.

Xie, Y., & Berkowitz, C. M. (2007). The use of conditional probability
functions and potential source contribution functions to identify source
regions and advection pathways of hydrocarbon emissions in Houston, Texas.
Atmospheric Environment, 41(28), 5831-5847.
}
\seealso{
Other trajectory analysis functions: 
\code{\link{importTraj}()},
\code{\link{trajCluster}()},
\code{\link{trajPlot}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{trajectory analysis functions}
