% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajPlot.R
\name{trajPlot}
\alias{trajPlot}
\title{Trajectory line plots with conditioning}
\usage{
trajPlot(
  mydata,
  lon = "lon",
  lat = "lat",
  pollutant = NULL,
  type = "default",
  group = NULL,
  cols = "default",
  crs = 4326,
  map = TRUE,
  map.res = "medium",
  map.fill = TRUE,
  map.cols = "grey40",
  map.border = "black",
  map.alpha = 0.4,
  map.lwd = 1,
  map.lty = 1,
  grid.col = "deepskyblue",
  grid.nx = 9,
  grid.ny = grid.nx,
  npoints = 12,
  origin = TRUE,
  key.title = group,
  key.position = "right",
  key.columns = 1,
  strip.position = "top",
  auto.text = TRUE,
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{Data frame, the result of importing a trajectory file using
\code{\link[=importTraj]{importTraj()}}.}

\item{lon, lat}{Columns containing the decimal longitude and latitude.}

\item{pollutant}{Pollutant (or any numeric column) to be plotted, if any.
Alternatively, use \code{group}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the columns and the second for the rows.}

\item{group}{A condition to colour the plot by, passed to \code{\link[=cutData]{cutData()}}. An
alternative to \code{pollutant}, and used preferentially to \code{pollutant} if both
are set.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{crs}{The coordinate reference system to use for plotting. Defaults to
\code{4326}, which is the WGS84 geographic coordinate system, the standard,
unprojected latitude/longitude system used in GPS, Google Earth, and GIS
mapping. Other \code{crs} values are available - for example, \code{27700} will use
the the OSGB36/British National Grid.}

\item{map}{Should a base map be drawn? If \code{TRUE} the world base map provided
by \code{\link[ggplot2:map_data]{ggplot2::map_data()}} will be used.}

\item{map.res}{The scale of the map to use. One of \code{110}, \code{50}, \code{10} or
\code{small}, \code{medium}, \code{large}. Passed to \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}}.}

\item{map.fill}{Should the base map be a filled polygon? Default is to fill
countries.}

\item{map.cols}{If \code{map.fill = TRUE} \code{map.cols} controls the fill colour.
Examples include \code{map.fill = "grey40"} and \code{map.fill = openColours("default", 10)}. The latter colours the countries and can help
differentiate them.}

\item{map.border}{The colour to use for the map outlines/borders. Defaults to
\code{"black"}.}

\item{map.alpha}{The transparency level of the filled map which takes values
from 0 (full transparency) to 1 (full opacity). Setting it below 1 can help
view trajectories, trajectory surfaces etc. \emph{and} a filled base map.}

\item{map.lwd}{The map line width, a positive number, defaulting to \code{1}.}

\item{map.lty}{The map line type. Line types can either be specified as an
integer (\code{0} = blank, \code{1} = solid (default), \code{2} = dashed, \code{3} = dotted,
\code{4} = dotdash, \code{5} = longdash, \code{6} = twodash) or as one of the character
strings "blank", "solid", "dashed", "dotted", "dotdash", "longdash", or
"twodash", where "blank" uses 'invisible lines' (i.e., does not draw them).}

\item{grid.col}{The colour of the map grid to be used. To remove the grid set
\code{grid.col = "transparent"}.}

\item{grid.nx, grid.ny}{The approximate number of ticks to draw on the map
grid. \code{grid.nx} defaults to \code{9}, and \code{grid.ny} defaults to whatever value
is passed to \code{grid.nx}. Setting both values to \code{0} will remove the grid
entirely. The number of ticks is approximate as this value is passed to
\code{\link[scales:breaks_pretty]{scales::breaks_pretty()}} to determine nice-looking, round breakpoints.}

\item{npoints}{A dot is placed every \code{npoints} along each full trajectory.
For hourly back trajectories points are plotted every \code{npoint} hours. This
helps to understand where the air masses were at particular times and get a
feel for the speed of the air (points closer together correspond to slower
moving air masses). If \code{npoints = NA} then no points are added.}

\item{origin}{If true a filled circle dot is shown to mark the receptor
point.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, x-axis label and y-axis label. All of these are
passed through to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\description{
This function plots back trajectories. This function requires that data are
imported using the \code{\link[=importTraj]{importTraj()}} function, or matches that structure.
}
\details{
Several types of trajectory plot are available:
\itemize{
\item \code{\link[=trajPlot]{trajPlot()}} by default will plot each lat/lon location showing the origin
of each trajectory, if no \code{pollutant} is supplied.
\item If a pollutant is given, by merging the trajectory data with concentration
data, the trajectories are colour-coded by the concentration of \code{pollutant}.
With a long time series there can be lots of overplotting making it difficult
to gauge the overall concentration pattern. In these cases setting \code{alpha} to
a low value e.g. 0.1 can help.
}

The user can also show points instead of lines by \code{plot.type = "p"}.

Note that \code{\link[=trajPlot]{trajPlot()}} will plot only the full length trajectories. This
should be remembered when selecting only part of a year to plot.
}
\examples{
\dontrun{
# show a simple case with no pollutant i.e. just the trajectories
# let's check to see where the trajectories were coming from when
# Heathrow Airport was closed due to the Icelandic volcanic eruption
# 15--21 April 2010.
# import trajectories for London and plot

lond <- importTraj("london", 2010)

# well, HYSPLIT seems to think there certainly were conditions where trajectories
# orginated from Iceland...
trajPlot(selectByDate(lond, start = "15/4/2010", end = "21/4/2010"))

# plot by day, need a column that makes a date
lond$day <- as.Date(lond$date)
trajPlot(
  selectByDate(lond, start = "15/4/2010", end = "21/4/2010"),
  type = "day"
)

# or show each day grouped by colour, with some other options set
trajPlot(
  selectByDate(lond, start = "15/4/2010", end = "21/4/2010"),
  group = "day",
  cols = "turbo",
  key.position = "right",
  key.columns = 1,
  linewidth = 2,
  size = 4
)
}
}
\seealso{
Other trajectory analysis functions: 
\code{\link{importTraj}()},
\code{\link{trajCluster}()},
\code{\link{trajLevel}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{trajectory analysis functions}
