% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajPlot.R
\name{trajPlot}
\alias{trajPlot}
\title{Trajectory line plots with conditioning}
\usage{
trajPlot(
  mydata,
  lon = "lon",
  lat = "lat",
  pollutant = NULL,
  type = "default",
  map = TRUE,
  group = NULL,
  cols = "default",
  crs = 4326,
  map.fill = TRUE,
  map.cols = "grey40",
  map.border = "black",
  map.alpha = 0.4,
  map.lwd = 1,
  map.lty = 1,
  grid.col = "deepskyblue",
  grid.nx = 9,
  grid.ny = grid.nx,
  npoints = 12,
  origin = TRUE,
  key = TRUE,
  key.title = group,
  key.position = "right",
  key.columns = 1,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{Data frame, the result of importing a trajectory file using
\code{\link[=importTraj]{importTraj()}}.}

\item{lon, lat}{Columns containing the decimal longitude and latitude.}

\item{pollutant}{Pollutant (or any numeric column) to be plotted, if any.
Alternatively, use \code{group}.}

\item{type}{\code{type} determines how the data are split, i.e., conditioned, and
then plotted. The default is will produce a single plot using the entire
data. Type can be one of the built-in types as detailed in \code{cutData} e.g.
"season", "year", "weekday" and so on. For example, \code{type = "season"} will
produce four plots --- one for each season.

It is also possible to choose \code{type} as another variable in the data frame.
If that variable is numeric, then the data will be split into four
quantiles (if possible) and labelled accordingly. If type is an existing
character or factor variable, then those categories/levels will be used
directly. This offers great flexibility for understanding the variation of
different variables and how they depend on one another.

\code{type} can be up length two e.g. \code{type = c("season", "weekday")} will
produce a 2x2 plot split by season and day of the week. Note, when two
types are provided the first forms the columns and the second the rows.}

\item{map}{Should a base map be drawn? If \code{TRUE} the world base map provided
by \code{\link[ggplot2:map_data]{ggplot2::map_data()}} will be used.}

\item{group}{A condition to colour the plot by, passed to \code{\link[=cutData]{cutData()}}. An
alternative to \code{pollutant}, and used preferentially to \code{pollutant} if both
are set.}

\item{cols}{Colours for plotting. Passed to \code{\link[=openColours]{openColours()}}.}

\item{crs}{The coordinate reference system to use for plotting. Defaults to
\code{4326}, which is the WGS84 geographic coordinate system, the standard,
unprojected latitude/longitude system used in GPS, Google Earth, and GIS
mapping. Other \code{crs} values are available - for example, \code{27700} will use
the the OSGB36/British National Grid.}

\item{map.fill}{Should the base map be a filled polygon? Default is to fill
countries.}

\item{map.cols}{If \code{map.fill = TRUE} \code{map.cols} controls the fill colour.
Examples include \code{map.fill = "grey40"} and \code{map.fill = openColours("default", 10)}. The latter colours the countries and can help
differentiate them.}

\item{map.border}{The colour to use for the map outlines/borders. Defaults to
\code{"black"}.}

\item{map.alpha}{The transparency level of the filled map which takes values
from 0 (full transparency) to 1 (full opacity). Setting it below 1 can help
view trajectories, trajectory surfaces etc. \emph{and} a filled base map.}

\item{map.lwd}{The map line width, a positive number, defaulting to \code{1}.}

\item{map.lty}{The map line type. Line types can either be specified as an
integer (\code{0} = blank, \code{1} = solid (default), \code{2} = dashed, \code{3} = dotted,
\code{4} = dotdash, \code{5} = longdash, \code{6} = twodash) or as one of the character
strings "blank", "solid", "dashed", "dotted", "dotdash", "longdash", or
"twodash", where "blank" uses 'invisible lines' (i.e., does not draw them).}

\item{grid.col}{The colour of the map grid to be used. To remove the grid set
\code{grid.col = "transparent"}.}

\item{grid.nx, grid.ny}{The approximate number of ticks to draw on the map
grid. \code{grid.nx} defaults to \code{9}, and \code{grid.ny} defaults to whatever value
is passed to \code{grid.nx}. Setting both values to \code{0} will remove the grid
entirely. The number of ticks is approximate as this value is passed to
\code{\link[scales:breaks_pretty]{scales::breaks_pretty()}} to determine nice-looking, round breakpoints.}

\item{npoints}{A dot is placed every \code{npoints} along each full trajectory.
For hourly back trajectories points are plotted every \code{npoint} hours. This
helps to understand where the air masses were at particular times and get a
feel for the speed of the air (points closer together correspond to slower
moving air masses). If \code{npoints = NA} then no points are added.}

\item{origin}{If true a filled circle dot is shown to mark the receptor
point.}

\item{key}{Should a key be drawn? Defaults to \code{TRUE}.}

\item{key.title}{The title of the key.}

\item{key.position}{Location where the scale key should be plotted. Allowed
arguments currently include \code{"top"}, \code{"right"}, \code{"bottom"}, and \code{"left"}.}

\item{key.columns}{Number of columns to be used in the key.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE}
titles and axis labels will automatically try and format pollutant names
and units properly e.g.  by subscripting the \sQuote{2} in NO2.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{border} sets the border colour of each bar.
}}
}
\description{
This function plots back trajectories. This function requires that data are
imported using the \code{\link[=importTraj]{importTraj()}} function, or matches that structure.
}
\details{
Several types of trajectory plot are available:
\itemize{
\item \code{\link[=trajPlot]{trajPlot()}} by default will plot each lat/lon location showing the origin
of each trajectory, if no \code{pollutant} is supplied.
\item If a pollutant is given, by merging the trajectory data with concentration
data, the trajectories are colour-coded by the concentration of \code{pollutant}.
With a long time series there can be lots of overplotting making it difficult
to gauge the overall concentration pattern. In these cases setting \code{alpha} to
a low value e.g. 0.1 can help.
}

The user can also show points instead of lines by \code{plot.type = "p"}.

Note that \code{\link[=trajPlot]{trajPlot()}} will plot only the full length trajectories. This
should be remembered when selecting only part of a year to plot.
}
\examples{
\dontrun{
# show a simple case with no pollutant i.e. just the trajectories
# let's check to see where the trajectories were coming from when
# Heathrow Airport was closed due to the Icelandic volcanic eruption
# 15--21 April 2010.
# import trajectories for London and plot

lond <- importTraj("london", 2010)

# well, HYSPLIT seems to think there certainly were conditions where trajectories
# orginated from Iceland...
trajPlot(selectByDate(lond, start = "15/4/2010", end = "21/4/2010"))

# plot by day, need a column that makes a date
lond$day <- as.Date(lond$date)
trajPlot(
  selectByDate(lond, start = "15/4/2010", end = "21/4/2010"),
  type = "day"
)

# or show each day grouped by colour, with some other options set
trajPlot(
  selectByDate(lond, start = "15/4/2010", end = "21/4/2010"),
  group = "day",
  cols = "turbo",
  key.position = "right",
  key.columns = 1,
  lwd = 2,
  cex = 4
)
}
}
\seealso{
Other trajectory analysis functions: 
\code{\link{importTraj}()},
\code{\link{trajCluster}()},
\code{\link{trajLevel}()}
}
\author{
David Carslaw

Jack Davison
}
\concept{trajectory analysis functions}
