% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendLevel.R
\name{trendLevel}
\alias{trendLevel}
\title{Plot heat maps of atmospheric composition data}
\usage{
trendLevel(
  mydata,
  pollutant = "nox",
  x = "month",
  y = "hour",
  type = "default",
  rotate.axis = c(90, 0),
  n.levels = c(10, 10, 4),
  windflow = NULL,
  limits = NULL,
  min.bin = 1,
  cols = "default",
  auto.text = TRUE,
  key.title = paste("use.stat.name", pollutant, sep = " "),
  key.position = "right",
  strip.position = "top",
  labels = NULL,
  breaks = NULL,
  statistic = c("mean", "max", "min", "median", "frequency", "sum", "sd", "percentile"),
  percentile = 95,
  stat.args = NULL,
  stat.safe.mode = TRUE,
  drop.unused.types = TRUE,
  col.na = "white",
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{The openair data frame to use to generate the \code{\link[=trendLevel]{trendLevel()}}
plot.}

\item{pollutant}{The name of the data series in \code{mydata} to sample to produce
the \code{\link[=trendLevel]{trendLevel()}} plot.}

\item{x, y, type}{The name of the data series to use as the \code{\link[=trendLevel]{trendLevel()}}
x-axis, y-axis or conditioning variable, passed to \code{\link[=cutData]{cutData()}}. These are
used before applying \code{statistic}. \code{\link[=trendLevel]{trendLevel()}} does not allow duplication
in \code{x}, \code{y} and \code{type} options.}

\item{rotate.axis}{The rotation to be applied to \code{trendLevel} \code{x} and \code{y}
axes. The default, \code{c(90, 0)}, rotates the x axis by 90 degrees but does
not rotate the y axis. If only one value is supplied, this is applied to
both axes; if more than two values are supplied, only the first two are
used.}

\item{n.levels}{The number of levels to split \code{x}, \code{y} and \code{type} data into
if numeric. The default, \code{c(10, 10, 4)}, cuts numeric \code{x} and \code{y} data into
ten levels and numeric \code{type} data into four levels. This option is ignored
for date conditioning and factors. If less than three values are supplied,
three values are determined by recursion; if more than three values are
supplied, only the first three are used.}

\item{windflow}{If \code{TRUE}, the vector-averaged wind speed and direction will
be plotted using arrows. Alternatively, can be a list of arguments to
control the appearance of the arrows (colour, linewidth, alpha value,
etc.). See \code{\link[=windflowOpts]{windflowOpts()}} for details.}

\item{limits}{The colour scale range to use when generating the
\code{\link[=trendLevel]{trendLevel()}} plot.}

\item{min.bin}{The minimum number of records required in a bin to show a
value. Bins with fewer than \code{min.bin} records are set to \code{NA}. The default
is 1, i.e., all bins with no records are set to \code{NA}. Setting \code{min.bin} to
a value greater than 1 can be useful to exclude bins with very few records
that might produce unreliable statistic values.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{breaks, labels}{If a categorical colour scale is required then \code{breaks}
should be specified. These should be provided as a numeric vector, e.g.,
\code{breaks = c(0, 50, 100, 1000)}. Users should set the maximum value of
\code{breaks} to exceed the maximum data value to ensure it is within the
maximum final range, e.g., 100--1000 in this case. Labels will
automatically be generated, but can be customised by passing a character
vector to \code{labels}, e.g., \code{labels = c("good", "bad", "very bad")}. In this
example, \code{0 - 50} will be \code{"good"} and so on. Note there is one less label
than break.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option. Functions can also be sent directly via
\code{statistic}; see 'Details' for more information.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{stat.args}{Additional options to be used with \code{statistic} if this is a
function. The extra options should be supplied as a list of named
parameters; see 'Details' for more information.}

\item{stat.safe.mode}{An addition protection applied when using functions
directly with \code{statistic} that most users can ignore. This option returns
\code{NA} instead of running \code{statistic} on binned sub samples that are empty.
Many common functions terminate with an error message when applied to an
empty dataset. So, this option provides a mechanism to work with such
functions. For a very few cases, e.g., for a function that counted missing
entries, it might need to be set to \code{FALSE}; see 'Details' for more
information.}

\item{drop.unused.types}{Hide unused/empty \code{type} conditioning cases. Some
conditioning options may generate empty cases for some data sets, e.g. a
hour of the day when no measurements were taken. Empty \code{x} and \code{y} cases
generate 'holes' in individual plots. However, empty \code{type} cases would
produce blank panels if plotted. Therefore, the default, \code{TRUE}, excludes
these empty panels from the plot. The alternative \code{FALSE} plots all \code{type}
panels.}

\item{col.na}{Colour to be used to show missing data.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{title} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{border} sets the border colour of each tile.
}}
}
\value{
an \link[=openair-package]{openair} object.
}
\description{
The \code{\link[=trendLevel]{trendLevel()}} function provides a way of rapidly showing a large amount
of data in a condensed form. In one plot, the variation in the concentration
of one pollutant can to shown as a function of between two and four
categorical properties. The default arguments plot hour of day on the x-axis
and month of year on the y-axis. However, \code{x}, \code{y} and \code{type} and summarising
statistics can all be modified to provide a range of other similar plots, all
being passed to \code{\link[=cutData]{cutData()}} for discretisation. The average wind speed and
direction in each bin can also be plotted using the \code{windflow} argument.
}
\details{
\code{\link[=trendLevel]{trendLevel()}} allows the use of third party summarising functions via the
\code{statistic} option. Any additional function arguments not included within a
function called using \code{statistic} should be supplied as a list of named
parameters and sent using \code{stat.args}. For example, the encoded option
\code{statistic = "mean"} is equivalent to \verb{statistic = mean, stat.args = list(na.rm = TRUE)} or the R command \code{mean(x, na.rm = TRUE)}. Many R
functions and user's own code could be applied in a similar fashion, subject
to the following restrictions: the first argument sent to the function must
be the data series to be analysed; the name 'x' cannot be used for any of the
extra options supplied in \code{stat.args}; and the function should return the
required answer as a numeric or \code{NA}. Note: If the supplied function returns
more than one answer, currently only the first of these is retained and used
by \code{\link[=trendLevel]{trendLevel()}}. All other returned information will be ignored without
warning. If the function terminates with an error when it is sent an empty
data series, the option \code{stat.safe.mode} should not be set to \code{FALSE} or
\code{\link[=trendLevel]{trendLevel()}} may fail. Note: The \code{stat.safe.mode = TRUE} option returns an
NA without warning for empty data series.
}
\examples{
# basic use
# default statistic = "mean"
trendLevel(mydata, pollutant = "nox")

# applying same as 'own' statistic
my.mean <- function(x) mean(x, na.rm = TRUE)
trendLevel(mydata, pollutant = "nox", statistic = my.mean)

# alternative for 'third party' statistic
# trendLevel(mydata, pollutant = "nox", statistic = mean,
#           stat.args = list(na.rm = TRUE))

\dontrun{
# example with categorical scale
trendLevel(mydata,
  pollutant = "no2",
  border = "white", statistic = "max",
  breaks = c(0, 50, 100, 500),
  labels = c("low", "medium", "high"),
  cols = c("forestgreen", "yellow", "red")
)
}
}
\author{
Karl Ropkins

David Carslaw

Jack Davison
}
