% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendLevel.R
\name{trendLevel}
\alias{trendLevel}
\title{Plot heat maps of atmospheric composition data}
\usage{
trendLevel(
  mydata,
  pollutant = "nox",
  x = "month",
  y = "hour",
  type = "default",
  rotate.axis = c(90, 0),
  n.levels = c(10, 10, 4),
  windflow = NULL,
  limits = NULL,
  min.bin = 1,
  cols = "default",
  auto.text = TRUE,
  key.title = paste("use.stat.name", pollutant, sep = " "),
  key.position = "right",
  breaks = NULL,
  statistic = c("mean", "max", "min", "median", "frequency", "sum", "sd", "percentile"),
  percentile = 95,
  stat.args = NULL,
  stat.safe.mode = TRUE,
  drop.unused.types = TRUE,
  col.na = "white",
  plot = TRUE,
  key = NULL,
  ...
)
}
\arguments{
\item{mydata}{The openair data frame to use to generate the \code{\link[=trendLevel]{trendLevel()}}
plot.}

\item{pollutant}{The name of the data series in \code{mydata} to sample to produce
the \code{\link[=trendLevel]{trendLevel()}} plot.}

\item{x, y, type}{The name of the data series to use as the \code{\link[=trendLevel]{trendLevel()}}
x-axis, y-axis or conditioning variable, passed to \code{\link[=cutData]{cutData()}}. These are
used before applying \code{statistic}. \code{\link[=trendLevel]{trendLevel()}} does not allow duplication
in \code{x}, \code{y} and \code{type} options.}

\item{rotate.axis}{The rotation to be applied to \code{trendLevel} \code{x} and \code{y}
axes. The default, \code{c(90, 0)}, rotates the x axis by 90 degrees but does
not rotate the y axis. If only one value is supplied, this is applied to
both axes; if more than two values are supplied, only the first two are
used.}

\item{n.levels}{The number of levels to split \code{x}, \code{y} and \code{type} data into
if numeric. The default, \code{c(10, 10, 4)}, cuts numeric \code{x} and \code{y} data into
ten levels and numeric \code{type} data into four levels. This option is ignored
for date conditioning and factors. If less than three values are supplied,
three values are determined by recursion; if more than three values are
supplied, only the first three are used.}

\item{windflow}{If \code{TRUE}, the vector-averaged wind speed and direction will
be plotted using arrows. Alternatively, can be a list of arguments to
control the appearance of the arrows (colour, linewidth, alpha value,
etc.). See \code{\link[=windflowOpts]{windflowOpts()}} for details.}

\item{limits}{The colour scale range to use when generating the
\code{\link[=trendLevel]{trendLevel()}} plot.}

\item{min.bin}{The minimum number of records required in a bin to show a
value. Bins with fewer than \code{min.bin} records are set to \code{NA}. The default
is 1, i.e., all bins with no records are set to \code{NA}. Setting \code{min.bin} to
a value greater than 1 can be useful to exclude bins with very few records
that might produce unreliable statistic values.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). Alternatively, can be a list of arguments to control the
colour palette more closely (e.g., \code{palette}, \code{direction}, \code{alpha}, etc.).
See \code{\link[=openColours]{openColours()}} and \code{\link[=colourOpts]{colourOpts()}} for more details.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{key.title}{Used to set the title of the legend. The legend title is
passed to \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{breaks}{\code{breaks} bins a continuous axis into discrete bins. It can
either take a single number (e.g., \code{breaks = 5}) to split the scale into
quantiles, a vector of numbers (e.g., \verb{breaks = c(0, 50, 100, 200, 500}) to
define specific break-points, or a named list. See \code{\link[=breakOpts]{breakOpts()}} for more
details.}

\item{statistic}{The statistic to apply when aggregating the data; default is
the mean. Can be one of \code{"mean"}, \code{"max"}, \code{"min"}, \code{"median"},
\code{"frequency"}, \code{"sum"}, \code{"sd"}, \code{"percentile"}. Note that \code{"sd"} is the
standard deviation, \code{"frequency"} is the number (frequency) of valid
records in the period and \code{"data.cap"} is the percentage data capture.
\code{"percentile"} is the percentile level (\%) between 0-100, which can be set
using the \code{"percentile"} option. Functions can also be sent directly via
\code{statistic}; see 'Details' for more information.}

\item{percentile}{The percentile level used when \code{statistic = "percentile"}.
The default is 95\%.}

\item{stat.args}{Additional options to be used with \code{statistic} if this is a
function. The extra options should be supplied as a list of named
parameters; see 'Details' for more information.}

\item{stat.safe.mode}{An addition protection applied when using functions
directly with \code{statistic} that most users can ignore. This option returns
\code{NA} instead of running \code{statistic} on binned sub samples that are empty.
Many common functions terminate with an error message when applied to an
empty dataset. So, this option provides a mechanism to work with such
functions. For a very few cases, e.g., for a function that counted missing
entries, it might need to be set to \code{FALSE}; see 'Details' for more
information.}

\item{drop.unused.types}{Hide unused/empty \code{type} conditioning cases. Some
conditioning options may generate empty cases for some data sets, e.g. a
hour of the day when no measurements were taken. Empty \code{x} and \code{y} cases
generate 'holes' in individual plots. However, empty \code{type} cases would
produce blank panels if plotted. Therefore, the default, \code{TRUE}, excludes
these empty panels from the plot. The alternative \code{FALSE} plots all \code{type}
panels.}

\item{col.na}{Colour to be used to show missing data.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{key}{Deprecated; please use \code{key.position}. If \code{FALSE}, sets
\code{key.position} to \code{"none"}.}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available, varying somewhat in different
plotting functions:
\itemize{
\item \code{title}, \code{subtitle}, \code{caption}, \code{tag}, \code{xlab} and \code{ylab} control the plot
title, subtitle, caption, tag, x-axis label and y-axis label, passed to
\code{\link[ggplot2:labs]{ggplot2::labs()}} via \code{\link[=quickText]{quickText()}} if \code{auto.text = TRUE}.
\item \code{xlim}, \code{ylim} and \code{limits} control the limits of the x-axis, y-axis and
colorbar scales.
\item \code{ncol} and \code{nrow} set the number of columns and rows in a faceted plot.
\item \code{scales} can be \code{"fixed"}, \code{"free_x"}, \code{"free_y"} or \code{"free"} to control
whether axes are shared across facets when using \code{type}. Also supported are
the legacy \code{x.relation} and \code{y.relation}, which can be either \code{"same"} or
\code{"free"} and get remapped to \code{scales} automatically.
\item Similarly, \code{space}, \code{axes}, \code{axis.labels}, \code{switch} and \code{strip.position}
can be used to customise the appearance of faceted plots. See
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} and \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} for the arguments these
take.
\item \code{fontsize} overrides the overall font size of the plot by setting the
\code{text} argument of \code{\link[ggplot2:theme]{ggplot2::theme()}}. It may also be applied
proportionately to any \code{openair} annotations (e.g., N/E/S/W labels on polar
coordinate plots).
\item Various graphical parameters are also supported: \code{linewidth},
\code{linetype},\code{ shape}, \code{size}, \code{border}, and \code{alpha}. Not all parameters
apply to all plots. These can take a single value, or a vector of multiple
values - e.g., \code{shape = c(1, 2)} - which will be recycled to the length of
values needed.
\item \code{lineend}, \code{linejoin} and \code{linemitre} tweak the appearance of line
plots; see \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} for more information.
\item In polar coordinate plots, \code{annotate = FALSE} will remove the N/E/S/W
labels and any other annotations.
}}
}
\value{
an \link[=openair-package]{openair} object.
}
\description{
The \code{\link[=trendLevel]{trendLevel()}} function provides a way of rapidly showing a large amount
of data in a condensed form. In one plot, the variation in the concentration
of one pollutant can to shown as a function of between two and four
categorical properties. The default arguments plot hour of day on the x-axis
and month of year on the y-axis. However, \code{x}, \code{y} and \code{type} and summarising
statistics can all be modified to provide a range of other similar plots, all
being passed to \code{\link[=cutData]{cutData()}} for discretisation. The average wind speed and
direction in each bin can also be plotted using the \code{windflow} argument.
}
\details{
\code{\link[=trendLevel]{trendLevel()}} allows the use of third party summarising functions via the
\code{statistic} option. Any additional function arguments not included within a
function called using \code{statistic} should be supplied as a list of named
parameters and sent using \code{stat.args}. For example, the encoded option
\code{statistic = "mean"} is equivalent to \verb{statistic = mean, stat.args = list(na.rm = TRUE)} or the R command \code{mean(x, na.rm = TRUE)}. Many R
functions and user's own code could be applied in a similar fashion, subject
to the following restrictions: the first argument sent to the function must
be the data series to be analysed; the name 'x' cannot be used for any of the
extra options supplied in \code{stat.args}; and the function should return the
required answer as a numeric or \code{NA}. Note: If the supplied function returns
more than one answer, currently only the first of these is retained and used
by \code{\link[=trendLevel]{trendLevel()}}. All other returned information will be ignored without
warning. If the function terminates with an error when it is sent an empty
data series, the option \code{stat.safe.mode} should not be set to \code{FALSE} or
\code{\link[=trendLevel]{trendLevel()}} may fail. Note: The \code{stat.safe.mode = TRUE} option returns an
NA without warning for empty data series.
}
\examples{
# basic use
# default statistic = "mean"
trendLevel(mydata, pollutant = "nox")

# applying same as 'own' statistic
my.mean <- function(x) mean(x, na.rm = TRUE)
trendLevel(mydata, pollutant = "nox", statistic = my.mean)

# alternative for 'third party' statistic
# trendLevel(mydata, pollutant = "nox", statistic = mean,
#           stat.args = list(na.rm = TRUE))

\dontrun{
# example with categorical scale
trendLevel(mydata,
  pollutant = "no2",
  border = "white", statistic = "max",
  breaks = c(0, 50, 100, 500),
  labels = c("low", "medium", "high"),
  cols = c("forestgreen", "yellow", "red")
)
}
}
\author{
Karl Ropkins

David Carslaw

Jack Davison
}
