% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variationPlot.R
\name{variationPlot}
\alias{variationPlot}
\title{Variation Plot}
\usage{
variationPlot(
  mydata,
  pollutant = "nox",
  x = "hour",
  statistic = "mean",
  type = "default",
  group = "default",
  normalise = FALSE,
  difference = FALSE,
  conf.int = NULL,
  B = 100,
  local.tz = NULL,
  ci = TRUE,
  cols = "hue",
  alpha = 0.4,
  strip.position = "top",
  key.position = "top",
  key.columns = NULL,
  name.pol = NULL,
  auto.text = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{mydata}{A data frame of time series. Must include a \code{date} field and at
least one variable to plot.}

\item{pollutant}{Name of variable to plot. Two or more pollutants can be
plotted, in which case a form like \code{pollutant = c("nox", "co")} should be
used.}

\item{x}{A character value to be passed to \code{\link[=cutData]{cutData()}}; used to define the
category by which \code{pollutant} will be varied and plotted.}

\item{statistic}{Can be \code{"mean"} (default) or \code{"median"}. If the statistic is
\code{"mean"} then the mean line and the 95\% confidence interval in the mean are
plotted by default. If the statistic is \code{"median"} then the median line is
plotted together with the 5/95 and 25/75th quantiles are plotted. Users can
control the confidence intervals with \code{conf.int}.}

\item{type}{Character string(s) defining how data should be split/conditioned
before plotting. \code{"default"} produces a single panel using the entire
dataset. Any other options will split the plot into different panels - a
roughly square grid of panels if one \code{type} is given, or a 2D matrix of
panels if two \code{types} are given. \code{type} is always passed to \code{\link[=cutData]{cutData()}},
and can therefore be any of:
\itemize{
\item A built-in type defined in \code{\link[=cutData]{cutData()}} (e.g., \code{"season"}, \code{"year"},
\code{"weekday"}, etc.). For example, \code{type = "season"} will split the plot into
four panels, one for each season.
\item The name of a numeric column in \code{mydata}, which will be split into
\code{n.levels} quantiles (defaulting to 4).
\item The name of a character or factor column in \code{mydata}, which will be used
as-is. Commonly this could be a variable like \code{"site"} to ensure data from
different monitoring sites are handled and presented separately. It could
equally be any arbitrary column created by the user (e.g., whether a nearby
possible pollutant source is active or not).
}

Most \code{openair} plotting functions can take two \code{type} arguments. If two are
given, the first is used for the rows and the second for the columns.}

\item{group}{This sets the grouping variable to be used. For example, if a
data frame had a column \code{site} setting \code{group = "site"} will plot all sites
together in each panel. Passed to \code{\link[=cutData]{cutData()}}.}

\item{normalise}{Should variables be normalised? The default is \code{FALSE}. If
\code{TRUE} then the variable(s) are divided by their mean values. This helps to
compare the shape of the diurnal trends for variables on very different
scales.}

\item{difference}{If two pollutants are chosen then setting \code{difference = TRUE} will also plot the difference in means between the two variables as
\code{pollutant[2] - pollutant[1]}. Bootstrap 95\\% confidence intervals of the
difference in means are also calculated. A horizontal dashed line is shown
at y = 0. The difference can also be calculated if there is a column that
identifies two groups, e.g., having used \code{\link[=splitByDate]{splitByDate()}}. In this case it
is possible to call the function with the option \code{group = "split.by"} and
\code{difference = TRUE}.}

\item{conf.int}{The confidence intervals to be plotted. If \code{statistic = "mean"} then the confidence intervals in the mean are plotted. If
\code{statistic = "median"} then the \code{conf.int} and \code{1 - conf.int} \emph{quantiles}
are plotted. Any number of \code{conf.int}s can be provided.}

\item{B}{Number of bootstrap replicates to use. Can be useful to reduce this
value when there are a large number of observations available to increase
the speed of the calculations without affecting the 95\% confidence interval
calculations by much.}

\item{local.tz}{Used for identifying whether a date has daylight savings time
(DST) applied or not. Examples include \code{local.tz = "Europe/London"},
\code{local.tz = "America/New_York"}, i.e., time zones that assume DST.
\url{https://en.wikipedia.org/wiki/List_of_zoneinfo_time_zones} shows time
zones that should be valid for most systems. It is important that the
original data are in GMT (UTC) or a fixed offset from GMT.}

\item{ci}{Should confidence intervals be shown? The default is \code{TRUE}.
Setting this to \code{FALSE} can be useful if multiple pollutants are chosen
where over-lapping confidence intervals can over complicate plots.}

\item{cols}{Colours to use for plotting. Can be a pre-set palette (e.g.,
\code{"turbo"}, \code{"viridis"}, \code{"tol"}, \code{"Dark2"}, etc.) or a user-defined vector
of R colours (e.g., \code{c("yellow", "green", "blue", "black")} - see
\code{\link[=colours]{colours()}} for a full list) or hex-codes (e.g., \code{c("#30123B", "#9CF649", "#7A0403")}). See \code{\link[=openColours]{openColours()}} for more details.}

\item{alpha}{The alpha transparency used for plotting confidence intervals.
\code{0} is fully transparent and 1 is opaque. The default is \code{0.4}.}

\item{strip.position}{Location where the facet 'strips' are located when
using \code{type}. When one \code{type} is provided, can be one of \code{"left"},
\code{"right"}, \code{"bottom"} or \code{"top"}. When two \code{type}s are provided, this
argument defines whether the strips are "switched" and can take either
\code{"x"}, \code{"y"}, or \code{"both"}. For example, \code{"x"} will switch the 'top' strip
locations to the bottom of the plot.}

\item{key.position}{Location where the legend is to be placed. Allowed
arguments include \code{"top"}, \code{"right"}, \code{"bottom"}, \code{"left"} and \code{"none"},
the last of which removes the legend entirely.}

\item{key.columns}{Number of columns to be used in a categorical legend. With
many categories a single column can make to key too wide. The user can thus
choose to use several columns by setting \code{key.columns} to be less than the
number of categories.}

\item{name.pol}{This option can be used to give alternative names for the
variables plotted. Instead of taking the column headings as names, the user
can supply replacements. For example, if a column had the name "nox" and
the user wanted a different description, then setting \code{name.pol = "nox before change"} can be used. If more than one pollutant is plotted then use
\code{c} e.g. \code{name.pol = c("nox here", "o3 there")}.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE} titles and
axis labels will automatically try and format pollutant names and units
properly, e.g., by subscripting the "2" in "NO2". Passed to \code{\link[=quickText]{quickText()}}.}

\item{plot}{When \code{openair} plots are created they are automatically printed
to the active graphics device. \code{plot = FALSE} deactivates this behaviour.
This may be useful when the plot \emph{data} is of more interest, or the plot is
required to appear later (e.g., later in a Quarto document, or to be saved
to a file).}

\item{...}{Addition options are passed on to \code{\link[=cutData]{cutData()}} for \code{type} handling.
Some additional arguments are also available:
\itemize{
\item \code{xlab}, \code{ylab} and \code{main} override the x-axis label, y-axis label, and plot title.
\item \code{layout} sets the layout of facets - e.g., \code{layout(2, 5)} will have 2 columns and 5 rows.
\item \code{lwd} and \code{lty} control various graphical parameters.
\item \code{fontsize} overrides the overall font size of the plot.
\item \code{ylim} controls axis limits.
}}
}
\value{
an \link[=openair-package]{openair} object.
}
\description{
The \code{variationPlot()} function is designed to explore how the distribution of
a pollutant (or other variable) changes by another variable (\code{x}). For
example, it can be used to explore how the distribution of \code{nox} varies by
\code{season} or by \code{weekday}. This plot can be extensively conditioned using the
\code{type} and \code{group} arguments, both of which are passed to \code{\link[=cutData]{cutData()}}. An
appropriate plot type will be chosen based on the type of \code{x} - e.g., ordered
variables will be joined by a line.
}
\details{
When \code{statistic = "mean"}, the plot shows the 95\% confidence intervals in the
mean. The 95\% confidence intervals are calculated through bootstrap
simulations, which will provide more robust estimates of the confidence
intervals (particularly when there are relatively few data).

Users can supply their own \code{ylim}, e.g. \code{ylim = c(0, 200)}.

The \code{difference} option calculates the difference in means between two
pollutants, along with bootstrap estimates of the 95\\% confidence intervals
in the difference. This works in two ways: either two pollutants are supplied
in separate columns (e.g. \code{pollutant = c("no2", "o3")}), or there are two
unique values of \code{group}. The difference is calculated as the second
pollutant minus the first and is labelled accordingly. This feature is
particularly useful for model evaluation and identifying where models diverge
from observations across time scales.

Depending on the choice of statistic, a subheading is added. Users can
control the text in the subheading through the use of \code{sub} e.g. \code{sub = ""}
will remove any subheading.
}
\examples{
# example using the 'mydata' dataset
variationPlot(
  mydata,
  pollutant = c("nox", "o3"),
  x = "hour",
  type = "season",
  normalise = TRUE
)
}
\seealso{
\code{\link[=timeVariation]{timeVariation()}}, which conveniently assembles many time-related
variation plots into a single plot
}
\author{
Jack Davison

David Carslaw
}
