% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_test.R
\name{write_test}
\alias{write_test}
\title{Write test for an R function}
\usage{
write_test(file, ...)
}
\arguments{
\item{file}{The file path of the R function.}

\item{...}{Additional arguments to pass to the chat_completion() function.}
}
\value{
If the input is a character string, the function returns the
generated output without creating a test file. Otherwise, it creates a test
file and returns the file name.
}
\description{
This function reads an R function from a file and generates a test file with
documentation.
}
\examples{
\dontrun{
# Write test for an R function
write_test("path/to/file.R")
}
}
