% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yorkFit.R
\name{YorkFit}
\alias{YorkFit}
\title{York regression function}
\usage{
YorkFit(
  input_data,
  X = "X",
  Y = "Y",
  Xstd = "Xstd",
  Ystd = "Ystd",
  weight = NA,
  Ri = 0,
  eps = 1e-07
)
}
\description{
York regression function
}
